% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base2gcl.R
\name{base2gcl}
\alias{base2gcl}
\title{Create \code{.gcl} Objects from a \emph{rubias} Baseline Object}
\usage{
base2gcl(base)
}
\arguments{
\item{base}{a rubias baseline (aka reference) object.}
}
\value{
\code{.gcl} objects are assigned to your global environment and a vector of silly codes (aka \code{sillyvec}) is returned (see details)
}
\description{
This function creates a \code{.gcl} object for each collection in a \emph{rubias} baseline object.
}
\details{
This function is a quick way of creating \code{.gcl} objects from a \emph{rubias} baseline object. The objects will contain the same variables as \code{.gcl} objects produced by \link{loki2r} and \link{loki2r_gaps} for compatibility with the \link{GCLr} package; however, most of the attribute variables will contain NAs.
}
\examples{

GCLr::base2gcl(base = GCLr::baseline)

}
