% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaselineEvalSampleSizes.GCL.r
\name{BaselineEvalSampleSizes.GCL}
\alias{BaselineEvalSampleSizes.GCL}
\title{This function creates a tibble of sample sizes for creating baseline and mixture files for baseline evaluation tests.
The output from this function can be used by CreateRubiasBaselineEval.GCL() to write out mixture and baseline files for baseline evaluation tests in rubias.}
\usage{
BaselineEvalSampleSizes.GCL(
  sillyvec,
  group_names,
  groupvec,
  mixsize,
  scenarios = round(seq(0.01, 1, 0.01), 2),
  maxprop = 0.5,
  seed = 56
)
}
\arguments{
\item{sillyvec}{A vector of silly codes without the ".gcl" extension.}

\item{group_names}{A character vector of group names with the length of max(groupvec).}

\item{groupvec}{A numeric vector indicating the group affiliation of each population in sillyvec.}

\item{mixsize}{The sample size of each test mixture.}

\item{scenarios}{A numeric vector of proportions to test for each group.}

\item{maxprop}{A numeric vector of length 1. This sets the maximum proportion of baseline individuals that can be
selected from each reporting group. For example, if maxprop = 0.5, the output tibble will only contain scenarios
where sample sizes for the test_group do not exceed 50\% of the fish in the baseline for that group.}

\item{seed}{An integer to set the seed for rmultinom(), so sample sizes are reproducible.}
}
\value{
The output of this function is a tibble with 4 variables: test_group, scenario, repunit, and samps.
For each test_group and scenario, the number of rows will be length(group_names).
}
\description{
This function creates a tibble of sample sizes for creating baseline and mixture files for baseline evaluation tests.
The output from this function can be used by CreateRubiasBaselineEval.GCL() to write out mixture and baseline files for baseline evaluation tests in rubias.
}
\examples{
BaselineEvalSampleSizes.GCL(sillyvec = c("KQUART06", "KQUART08", "KQUART10"), group_names = c("Upper_Susitna", "Chulitna", "Talkeetna", "Eastern_Susitna", "Deshka", "Yentna"), groupvec = c(1, 2, 3, 4, 5, 6), scenarios = round(seq(.01, 1, .01), 2), mixsize = 200, maxprop = 0.5, seed = 123)
}
