% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_indiv_assign_summary.r
\name{bayes_indiv_assign_summary}
\alias{bayes_indiv_assign_summary}
\title{Summarize BAYES Individual Assignment .CLS Files}
\usage{
bayes_indiv_assign_summary(
  group_names,
  groupvec,
  mixnames,
  maindir,
  nchains,
  nreps,
  burn = 0.5,
  thin = 100
)
}
\arguments{
\item{group_names}{Character vector of group names. Length(group_names) should be equal to max(groupvec).}

\item{groupvec}{Numeric vector indicating reporting group affiliation for each baseline population. Numbers correspond to group_names.}

\item{mixnames}{Character vector indicating mixture names (i.e., name of the folder where BAYES output files exist).}

\item{maindir}{Character vector indicating the directory where the mixture folders are located. \strong{Note: The results for each mixture must be in their own folder with the same name as the mixture.}}

\item{nchains}{Number of chains to summarize (there should be output files for each chain).}

\item{nreps}{Numeric vector indicating the number of MCMC reps. Can be a vector of length(mixnames) if there are a different number of reps per mixture.}

\item{burn}{Proportion of iterations to drop from the beginning of each chain. Default is 0.5, which drops the first 50\% of iterations.}

\item{thin}{Number for the 3rd argument for thin in CreateControlFile.GCL. Recommend thin = 100 for IA.}
}
\value{
A tibble with the following variables:
\describe{
\item{mixname}{The name of the mixture}
\item{id}{The individual id}
\item{group_names}{A variable for each group in group_names containing the proportion of thinned MCMC iterations in which an individual was assigned to a specific reporting group}
}
}
\description{
This function summarizes BAYES individual assignment .CLS files.
}
\note{
\describe{
\item{Individual ids}{Come from FK_FISH_ID if the mixture .gcl object exists, otherwise they are sequential. Mixture objects can be either new- (tibble) or old-style (list) .gcl objects.}
\item{Dropped individuals}{BAYES will drop mixture individuals from the analysis if they contain an allele that does not exist in the baseline. Check the BAYES summary files for each mixture and remove dropped individuals from the mixture .gcl objects before running this function. Otherwise, the function will produce an error message stating that the number of individuals in the mixture object exceeds the number of individuals in the BAYES output.}
}
}
\examples{
  \dontrun{
  group_names <- c("Crescent", "West", "JCL", "SusYen", "Fish", "KTNE", "Kenai", "Kasilof")
  groupvec <- c(1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 4, 4, 4, 4, 4, 4, 3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 3,
               4, 3, 6, 5, 5, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 8, 8, 8, 8)
  mixnames <- c("SUCICD16", "SUCICD16corr", "SUCIND16")
  maindir <- "V:/Analysis/2_Central/Sockeye/Cook Inlet/2012 Baseline/Mixture/2016 UCIfisheryMixtures/BAYES/Output"
  source("V:/Analysis/2_Central/Sockeye/Cook Inlet/2012 Baseline/Mixture/2016 UCIfisheryMixtures/Output/2016mixturegcls.R") # Get .gcl objects
  nchains <- 5
  nreps <- 40000
  burn <- 0.5
  thin <- 100
  bayes_indiv_assign_summary(group_names, groupvec, mixnames, maindir, nchains, nreps, burn, thin)
  }

}
