% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_output_copy_paste.r
\name{bayes_output_copy_paste}
\alias{bayes_output_copy_paste}
\title{Copy and Paste BAYES Output Files}
\usage{
bayes_output_copy_paste(origindir, targetdir, mixvec)
}
\arguments{
\item{origindir}{Original directory of files.
A character vector indicating where BAYES output files are located on the server.}

\item{targetdir}{Target directory structure.
A character vector indicating where to paste BAYES files (i.e., V drive).}

\item{mixvec}{Mixtures to be moved.
A character vector indicating which mixtures to move.}
}
\value{
This function has no explicit return value. It cuts and pastes BAYES output files
from the server to the target directory, and deletes the BAYES input files from the server.
}
\description{
This function copies the BAYES output files (.BO1, .BOT, .RGN, .SUM, .CLS) from the server
and pastes them onto the target directory (i.e., V drive). It also removes all BAYES input
files from the server. The function recursively searches the original directory for BAYES files
that match the mixvec.
}
\examples{
\dontrun{
bayes_output_copy_paste(origindir = "C:/Users/krshedd/BAYES",
                        targetdir = "V:/Analysis/4_Westward/Sockeye/KMA Commercial Harvest 2014-2016/Mixtures/BAYES/Late August 89loci",
                        mixvec = dget(file = "V:/Analysis/4_Westward/Sockeye/KMA Commercial Harvest 2014-2016/Mixtures/Objects/LateAugustMixtures2014Strata.txt"))

}

}
