% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_freq_pop.r
\name{calc_freq_pop}
\alias{calc_freq_pop}
\title{Calculate Allele Frequency for Multiple Loci and Silly Codes}
\usage{
calc_freq_pop(
  sillyvec,
  loci = LocusControl$locusnames,
  ncores = 4,
  LocusCtl = LocusControl
)
}
\arguments{
\item{sillyvec}{A character vector of silly codes without the ".gcl" extension.}

\item{loci}{A character vector of locus names; default is all \code{LocusControl$locusnames}.}

\item{ncores}{A numeric value for the number of cores to use in a \pkg{foreach} \verb{\%dopar\%} loop (default = 4).
If the number of cores exceeds the number on your device, \code{ncores} defaults to \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}} (default = LocusControl).}
}
\value{
A tibble with the following 6 columns:
\itemize{
\item \code{silly}: silly code
\item \code{locus}: locus name
\item \code{allele_no}: the allele number from \code{LocusControl} (1 or 2 for SNPs)
\item \code{allele}: the allele call from \code{LocusControl} (e.g. -, A, C, G, T)
\item \code{freq}: the allele frequency (count)
\item \code{proportion}: the allele frequency (proportion); these sum to 1 within a locus
}
}
\description{
This function calculates the allele frequency for each locus for each collection (silly code).
}
\examples{

sillyvec <- GCLr::ex_baseline$collection \%>\% unique()

loci <- names(GCLr::ex_baseline)[-c(1:4)][c(TRUE, FALSE)]

calc_freq_pop(sillyvec = sillyvec, loci = loci, LocusCtl = GCLr::ex_LocusControl)

}
