% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ind_het.R
\name{calc_ind_het}
\alias{calc_ind_het}
\title{Calculate individual heterozygosity}
\usage{
calc_ind_het(
  sillyvec,
  loci = LocusControl$locusnames,
  ncores = 4,
  LocusCtl = LocusControl
)
}
\arguments{
\item{sillyvec}{A character vector of silly codes without the ".gcl" extension.}

\item{loci}{A character vector of locus names; default is all \code{LocusControl$locusnames}.}

\item{ncores}{A numeric value for the number of cores to use in a \pkg{foreach} \verb{\%dopar\%} loop (default = 4).
If the number of cores exceeds the number on your device, \code{ncores} defaults to \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}} (default = LocusControl).}
}
\value{
A tibble with 1 row per individual and the following 6 columns:
\itemize{
\item \code{SillySource}: SillySource (i.e., silly_fishID)
\item \code{silly}: silly code
\item \code{fish_id}: GCL fish ID
\item \code{nloci}: the number of non-\code{NA} diploid loci genotyped per individual
\item \code{ploci}: the proportion of non-\code{NA} diploid loci genotyped per individual
\item \code{het}: individual heterozygosity (i.e., the proportion of non-\code{NA} diploid loci with heterozygous genotypes)
}
}
\description{
This function calculates individual heterozygosity across all non-\code{NA} diploid loci. This can be used as a proxy for detecting contaminated samples.
}
\examples{
\dontrun{
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)
loci <- names(GCLr::ex_baseline)[-c(1:4)][c(TRUE, FALSE)]

GCLr::calc_ind_het(sillyvec = sillyvec, loci = loci)
}

}
