% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex.r
\name{col2hex}
\alias{col2hex}
\title{Convert R color to hexadecimal}
\usage{
col2hex(color)
}
\arguments{
\item{color}{an R color name}
}
\value{
a tibble containing the following variables:
\itemize{
\item \code{col}: R color name
\item \code{hex}: hexadecimal color
\item \code{r}: red saturation
\item \code{g}: green saturation
\item \code{b}: blue saturation
}
}
\description{
This function takes an R color name and convert it to hexidecimal format.
}
\examples{
col2hex(color = "red")

}
