% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_map.R
\name{collections_map}
\alias{collections_map}
\title{Plot Collections on a Map}
\usage{
collections_map(input, file = NULL, select.basemap = FALSE, png = FALSE)
}
\arguments{
\item{input}{a \code{tibble} containing the following variables:
\itemize{
\item{number}{unique collection number; if NULL, this will default to 1:length(location); this will appear when you hover your mouse pointer over a waypoint}
\item{location}{location name for each collection; this will pop up when you click on a waypoint}
\item{group_name}{reporting group name for each collection}
\item{group_color}{the group color for each collection (see details)}
\item{latitude}{latitude in decimal degrees for each collection (see details)}
\item{longitude}{longitude in decimal degrees for each collection (see details)}
}}

\item{file}{the file path to output the html map with .html extension (default = NULL)}

\item{select.basemap}{if TRUE, select from a list of base maps in the console. If FALSE (default), basemap will be set to leaflet::providers$Esri.OceanBasemap}

\item{png}{save map as png (default: FALSE)}
}
\description{
This function plots collection locations on an interactive map.
}
\details{
If file is set to NULL, the map will not be save to an html file. Either R color names or hexidecimal colors can be used by this function. This function only works with coordinates in decimal degrees format with a negative sign to indicate western longitudes (e.g. 61.1596, -149.8876).
}
\examples{
 input <- tibble::tibble(number = 1:4,
                 location = c("Juneau Creek", "Russian River", "Killey River", "Funny River"),
                 group_name = c("Upper Kenai", "Upper Kenai", "Lower Kenai", "Lower Kenai"),
                 group_color = c("red", "red", "blue", "blue"),
                 latitude = c(60.49188, 60.4477, 60.478727, 60.42944),
                 longitude = c(-149.88454, -149.9865, -150.646294, -150.78141)
 )
 
 collections_map(input)
 
}
