% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_conflicts.r
\name{combine_conflicts}
\alias{combine_conflicts}
\alias{CombineConflictsWithPlateID.GCL.r}
\title{Combine QC Conflict Reports and Add PLATE_ID Column}
\usage{
combine_conflicts(files)
}
\arguments{
\item{files}{A vector of CSV files that you want to combine, including the file extension.}
}
\description{
This function combines QC conflict reports and adds a column of PLATE_IDs associated with each individual conflict.  The function requires "*.gcl" objects and the "ProjectSillys" vector created by ReadProjectloki2r.
}
\examples{
combine_conflicts(files = c("CM31 qc 01 Import Results.csv", "CM31 qc 02 Import Results.csv", "CM31 qc Plate3 Imports.csv"))

}
\author{
awbarclay
}
\concept{QA scripts}
