% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_loci.r
\name{combine_loci}
\alias{combine_loci}
\title{Combine Loci}
\usage{
combine_loci(sillyvec, markerset, update = TRUE, delim = c(".", "_")[1])
}
\arguments{
\item{sillyvec}{A character vector of silly codes.}

\item{markerset}{A vector of the set of loci you wish to combine.}

\item{update}{Logical switch. If TRUE, the "LocusControl" object is updated and all "\emph{.gcl" objects in "sillyvec" will be updated with the new marker. If FALSE, the "LocusControl" object is not updated and a temporary object called "}.temp.gcl" with the updated data is created.}

\item{delim}{Specifies the separator between combined loci, either a period (.) which is the default or an underscore (_) so locus names will work in SPAM.}
}
\description{
This function combines a set of markers into a single marker.
}
\details{
This function requires a LocusControl object. Run \code{\link[=create_locuscontrol]{create_locuscontrol()}} prior to this function. This function requires dplyr version 1.0.0 or higher.
}
\examples{
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

markerset <- c("One_GPDH", "One_GPDH")

LocusControl <- GCLr::ex_LocusControl

GCLr::combine_loci(sillyvec, markerset, update = FALSE, delim = c(".", "_")[1])

}
