% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineLoci.GCL.R
\name{CombineLoci.GCL}
\alias{CombineLoci.GCL}
\title{This function combines a set of markers into a single marker set.}
\usage{
CombineLoci.GCL(sillyvec, markerset, update = TRUE, delim = c(".", "_")[1])
}
\arguments{
\item{sillyvec}{a character vector of silly codes (e.g. sillyvec <- c("KQUART06","KQUART08","KQUART09"))}

\item{markerset}{a vector of the set of loci you wish to combine (e.g., c("Ots_vatf-251", "Ots_ZR-575"))}

\item{update}{logical switch. If TRUE, the "LocusControl" object is updated and all ".gcl" objects in "sillyvec" will be updated with the new marker. If FALSE, the "LocusControl" object is not updated and a temporary object called "*.temp.gcl" with the updated data is created.}

\item{delim}{the separator in the combined locus name, either a period (.) which is the default or an underscore (_)}
}
\description{
The combined marker will be added to supplied collections (silly codes) in sillyvec and the LocusControl object will be updated if update = TRUE
}
\note{
This function requires a LocusControl object. Run CreateLocusControl.GCL prior to this function. Also requires dplyr version >= 1.0.0
}
\examples{

CreateLocusControl.GCL(markersuite = "Sockeye2011_96SNPs", username = "awbarclay", password = "mypassword"

LOKI2R.GCL(sillyvec = c("SLARS11O", "SSKWEN07", "SSKK594L", "SMOOT92"), username = "awbarclay", password = "mypassword")

CombineLoci.GCL(sillyvec = c("SLARS11O", "SSKWEN07", "SSKK594L", "SMOOT92"), markerset = c("One_Cytb_17", "One_CO1","One_Cytb_26"), update = TRUE, delim = c(".","_")[1])

CombineLoci.GCL(sillyvec = c("SLARS11O", "SSKWEN07", "SSKK594L", "SMOOT92"), markerset = c("One_MHC2_251", "One_MHC2_190"), update = TRUE, delim = c(".","_")[1])

}
