% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_comps_between.r
\name{compare_comps_between}
\alias{compare_comps_between}
\title{Compare Compositions Between Mixtures}
\usage{
compare_comps_between(
  mixnames,
  groupnames,
  mixdir,
  diffs = seq(10)/100,
  nchains = 5,
  burn = 1/2
)
}
\arguments{
\item{mixnames}{A character vector of length 2 containing the mixture names to compare.}

\item{groupnames}{A character vector giving the names of the stock groups resolved by the mixtures, with length equal to the number columns in the RGN files.}

\item{mixdir}{An atomic string giving the folder path where the mixture sub-folders are located.}

\item{diffs}{Differences to check.}

\item{nchains}{The number of MCMC chains for summarizing the mixture results; default is 5.}

\item{burn}{The proportion of iterations to burn prior to summarizing results; default is 1/2.}
}
\value{
A list with two elements: \code{one.sided} and \code{two.sided}.
\describe{
\item{\code{one.sided}}{A tibble with one-sided p-values for each group.}
\item{\code{two.sided}}{A tibble with two-sided p-values for each group at different \code{diffs}.}
}
}
\description{
This function compares the difference in stock compositions between two mixtures.
\strong{This needs a better explanation of what it does}
}
\examples{

\dontrun{
load("V:/Analysis/2_Central/Sockeye/Cook Inlet/2012 Baseline/Mixture/2013 UCIfisheryMixtures/2013UCIfisheryMixtureAnalysis.RData")
mixdir <- "V:/Analysis/2_Central/Sockeye/Cook Inlet/2012 Baseline/Mixture/2013 UCIfisheryMixtures/BAYES/Output"
groups <- c("group1", "group2", "group3")  # Replace with actual group names
compare_comps_between(mixnames = c("DriftExpCorr.Jul11", "Drift.Jul8"), groupnames = groups, mixdir = mixdir, diffs = seq(10)/100, nchains = 5, burn = 1/2)
}

}
