% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.r
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Create confusion matrices}
\usage{
confusion_matrix(
  reference,
  gen_start_col = 5,
  output = c("group_group", "pop_group", "pop_pop")[1]
)
}
\arguments{
\item{reference}{A rubias baseline object created by create_rubias_base().}

\item{gen_start_col}{The first column containing genetic data in the baseline object. This will usually be the default unless the baseline object is modified with additional variables.}

\item{output}{A character vector of summary tibbles you want to produce. Must be one or more of the following: "pop_pop", "pop_group", "group_group". The default is "group_group".}
}
\value{
Named list of tibble(s) in long format:
\itemize{
\item \code{group_group}: A tibble containing the following variables:
\itemize{
\item \code{repunit}: The known repunit (aka reporting group).
\item \code{inferred_repunit}: The inferred repunit.
\item \code{mean_group_group_scaled_like}: The average probability of each individual from a group originating from a pop in that group.
}
\item \code{pop_group}: A tibble containing the following variables:
\itemize{
\item \code{collection}: The known population.
\item \code{inferred_repunit}: The inferred repunit.
\item \code{mean_pop_group_scaled_like}: The average probability of each individual from a pop originating from a pop in that group.
}
\item \code{pop_pop}: A tibble containing the following variables:
\itemize{
\item \code{pop}: The known population.
\item \code{inferred_pop}: The inferred repunit.
\item \code{mean_pop_pop_scaled_like}: The average probability of each individual from a pop originating from a pop.
}
}
}
\description{
This function is a wrapper for rubias::self_assign(). It takes the genetic likelihood profile produced by rubias::self_assign() and summarizes it into tibbles at the group by group, group by pop, and pop by pop level. This genetic likelihood profile is intended to show how well populations/reporting groups are differentiated by the markers at hand. These probabilities can give an idea of potential misallocation and where. This function can aid in marker selection.
}
\examples{
baseline <- read_csv("V:/Analysis/5_Coastwide/Chum/NPen2WA_Chum_baseline/rubias/baseline/NPen2Wa_Chum_227pops_91loci_base.csv")
ConfusionMatrices_out <- confusion_matrix(reference = baseline , gen_start_col = 5, output = c("group_group", "pop_group", "pop_pop"))

}
