% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bayes_base.r
\name{create_bayes_base}
\alias{create_bayes_base}
\title{Create Bayes Baseline File}
\usage{
create_bayes_base(sillyvec, loci, dir, baseline_name, ncores = 4)
}
\arguments{
\item{sillyvec}{Character vector of populations in the baseline.}

\item{loci}{Character vector of the loci you wish to include.}

\item{dir}{Character vector of where to save the ".bse" file.}

\item{baseline_name}{Character vector of what to name the baseline.bse, not including the .bse extension.}

\item{ncores}{The number of cores to use in a foreach \%dopar\% loop. If the number of cores exceeds the number on your device, then ncores defaults to \code{\link[=detectCores]{detectCores()}}.}
}
\value{
The fortran format of the baseline file - this object is needed for \code{\link[=create_bayes_ctl]{create_bayes_ctl()}}.
}
\description{
This function creates a baseline (.bse) file needed for the \code{BAYES}. It takes a character vector of populations
in the baseline, a character vector of the loci to include, the directory where the ".bse" file will be saved,
the desired name for the baseline (excluding the ".bse" extension), and the number of cores to use for parallel processing.
The function calculates the frequency of alleles for each population and locus, and saves the baseline as a .bse file.
It returns the Fortran format of the baseline file, which is needed for creating a control file for \code{BAYES}.
}
\examples{
\dontrun{
load("V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/Susitna_Chinook_baseline_2020.Rdata")
sillyvec <- Final_Pops$silly
loci <- loci82
base_fortran <- create_bayes_base(sillyvec = sillyvec, loci = loci, dir = getwd(), baseline_name = "SusitnaChinook31pops82loci", ncores = 8)
}

}
