% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bayes_mix.r
\name{create_bayes_mix}
\alias{create_bayes_mix}
\title{Create BAYES mixture files}
\usage{
create_bayes_mix(mixvec, loci, dir, ncores = 4, LocusCtl = LocusControl)
}
\arguments{
\item{mixvec}{Character vector of ".gcl" objects you want to produce mixture files for.}

\item{loci}{Character vector of the loci you wish to include.}

\item{dir}{Character vector of where to save the ".mix" file.}

\item{ncores}{Number of cores to use for parallel processing.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
Returns the FORTRAN format of the mixture file, which is needed for \code{\link[=create_bayes_ctl]{create_bayes_ctl()}}.
It also saves each mixture in \code{mixvec} as a .mix file.
}
\description{
This function creates a BAYES mixture (.mix) file for each silly in mixvec.
}
\note{
If you want to analyze more than one silly as a mixture, use \code{\link[=pool_collections]{pool_collections()}} to combine them into a new silly.gcl
}
\examples{
\dontrun{
load("V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/Susitna_Chinook_baseline_2020.Rdata")
loki2r(sillyvec = c("KSUSC18FW", "KSUSCN18", "KSUSC19FW", "KSUSCN19"), username = "awbarclay", password = password)
combine_loci(sillyvec = c("KSUSC18FW", "KSUSCN18", "KSUSC19FW", "KSUSCN19"), markerset = c("Ots_MHC1", "Ots_MHC2"))
pool_collections(collections = c("KSUSC18FW", "KSUSCN18"), newname = "Susitna2018")
pool_collections(collections = c("KSUSC19FW", "KSUSCN19"), newname = "Susitna2019")
loci <- c(loci82, "Ots_MHC1.Ots_MHC2")
mix_fortran <- create_bayes_mix(mixvec = c("Susitna2018", "Susitna2019"), loci = loci, dir = getwd(), ncores = 8)
}

}
