% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hwler_ctl.r
\name{create_hwler_ctl}
\alias{create_hwler_ctl}
\title{Create HWLER Control Files}
\usage{
create_hwler_ctl(
  sillyvec = NULL,
  loci,
  input,
  nsamples = c(1000, 2, 5),
  nchains = 5,
  dir,
  initval = "1.0",
  seeds = matrix(sample(seq(10000), 3 * nchains), nrow = 3),
  thin = c(1, 1, 1),
  inputfortran,
  switches = "T T T F T T T T T",
  LocusCtl = LocusControl
)
}
\arguments{
\item{sillyvec}{a character vector of baseline sillys included in the input file. Leave this \code{NULL} if no baseline individuals are present in the input file.}

\item{loci}{character vector of the loci used to produce the baseline and mixture files.}

\item{input}{The mixture and/or baseline input file name without the extension. This file can contain mixture data only, baseline data only, or a combination mixture and baseline.}

\item{nsamples}{a vector specifying the number of samples for each chain}

\item{nchains}{the number of MCMC chains to analyze the mixtures}

\item{dir}{the directory path where the control files will be saved}

\item{initval}{the initial starting value for each MCMC chain}

\item{seeds}{a matrix of random seeds containing 3 seeds per chain, where \code{nrow(seeds) = 3} and \code{ncol(seeds) = nchains}.}

\item{thin}{thinning intervals for MCMC sample of 1) stock proportions, 2) baseline allele or type relative frequencies, and 3) stock assignments of each mixture individual.}

\item{inputfortran}{the FORTRAN format of the input created by \code{\link[=create_hwler_input]{create_hwler_input()}}}

\item{switches}{a character string of logical switches, with default value "T T T F T T T T T" (see details).}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
Writes out HWLER control (.ctl) files.
}
\description{
This function creates HWLER (Hardy-Weinberg and Linkage Equilibrium Sampler) control files for each chain based on the specified parameters.
}
\details{
The HWLER input file can contain only mixture individuals, only baseline individuals, or a combination of mixture and baseline individuals depending on the type of analysis you are doing.
The \code{switches} argument has 9 program options turned “on” with “T” for true or “off” with “F” for false:
\enumerate{
\item Baseline printed
\item Mixture printed
\item Output stock assignments of individuals (baseline & mixture) or snapshots of cluster assignments of individuals (baseline or mixture only)
\item Suppress the printing of MCMC samples of stock composition (baseline & mixture) or snapshots of numbers of individuals assigned to clusters (baseline or mixture only)
\item Baseline provided
\item Mixture provided
\item Individual identification provided in data file
\item Output cluster information for the binary tree
\item Sample the Dirichlet mass parameter
}
}
\examples{
\dontrun{
load("V:/Analysis/2_Central/Sockeye/Cook Inlet/Missing Baseline Analysis/Missing Baseline Analysis.RData")
HWLERControlFile.GCL(sillyvec = PopNames69, loci = loci34, input = "myHWLER.input", nsamples = c(22000, 2, 5), nchains = 1, dir = "HWLER/control", initval = "1.0", seeds = matrix(sample(seq(2000000000), 3*nchains), nrow = 3), thin = c(1, 100, 4), inputfortran = inputfortran, switches = "F F T F T T T T F")
}
}
\seealso{
For additional details, run the following code in your console to bring up the HWLER manual:
\code{utils::browseURL(system.file("HWLER", "HWLER_manual.doc", package = "GCLr"))}
}
