% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_locuscontrol.r
\name{create_locuscontrol}
\alias{create_locuscontrol}
\title{Create LocusControl Object}
\usage{
create_locuscontrol(markersuite = NULL, locusnames = NULL, username, password)
}
\arguments{
\item{markersuite}{The pre-defined name in LOKI for the set of markers for which you want genotypes (e.g., markersuite="KenaiChinook2010_40SNPs"). This set must be pre-defined in LOKI.}

\item{locusnames}{a character vector of locus names spelled exactly the way they are in LOKI. Use "Locus Report Brief" in OceanAK to get the correct locus names.}

\item{username}{your state user name}

\item{password}{your password used to access LOKI - contact Eric Lardizabal if you don't have a password for LOKI}
}
\value{
This function assigns a tibble with the following columns:
\itemize{
\item \code{Markersuite} (character): markersuite name.
\item \code{locusnames} (character): the locus name used by the GCL.
\item \code{Publishedlocusnames} (character): the locus name used in publications.
\item \code{nalleles} (integer): the number of alleles for each locus.
\item \code{ploidy} (integer): the ploidy of the locus (2 = diploid, 1 = haploid).
\item \code{alleles} (named list): named list of tibbles containing the alleles for each locus.
}
The tibble will be named "LocusControl" and assigned to your current workspace.
}
\description{
This function connects to LOKI (GCL database) and creates a "LocusControl" object.
}
\examples{
password <- "************"
create_locuscontrol(markersuite = "UCI_Chinook_GTSeq_557SNPs", locusnames = NULL, username = "awbarclay", password = password)

}
