% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rubias_mix.r
\name{create_rubias_mix}
\alias{create_rubias_mix}
\title{Create rubias Mixture}
\usage{
create_rubias_mix(sillyvec, loci, path = "rubias/mixture")
}
\arguments{
\item{sillyvec}{A character vector of the mixture silly codes, each silly code is treated as its own mixture.}

\item{loci}{A character vector of the loci to include.}

\item{path}{A character vector of where to save each mixture as a ".csv" (see details).}
}
\value{
A dataframe in \pkg{rubias} mixture format and writes out a ".csv" file of each mixture.
}
\description{
This function creates the mixture dataframe needed for \pkg{rubias}.
}
\details{
This function saves each mixture as its own ".csv" file for posterity. To read in these ".csv" files, use \code{readr::read_csv(file = file, col_types = cols(.default = "c")} to make sure all columns are character vectors (if homozygous for T, it will become a logical vector).
}
\examples{
\dontrun{
 
 mixvec <- c("S1_2018", "S2_2018", "S3_2018", "S4_2018", "S5_2018")
 
 loci <- LocusControl$locusnames
 
 GCLr::create_rubias_mix(sillyvec = mixvec, loci = loci, path = "rubias/mixture")
 
}

}
\seealso{
\link{ex_mixtures}
}
