% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLocusControl.GCL.R
\name{CreateLocusControl.GCL}
\alias{CreateLocusControl.GCL}
\title{This function connects to the Gene Conservation Lab's Oracle database Loki and creates a "LocusControl" object}
\usage{
CreateLocusControl.GCL(
  markersuite = NULL,
  locusnames = NULL,
  username,
  password
)
}
\arguments{
\item{markersuite}{the pre-defined name in LOKI for the set of markers for which you want genotypes (e.g. markersuite="KenaiChinook2010_40SNPs"). This set must be pre-defined in Loki (see Eric Lardizabal).}

\item{locusnames}{a character vector of locus names spelled exactly the way they are in Loki. Use Locus Report Brief in OceanAK to get the correct locus names.}

\item{username}{your Oracle user name}

\item{password}{your Oracle password used to access Loki (see Eric Lardizabal if you don't have a username or password)}
}
\value{
A tibble with 6 columns: Markersuite, locusnames, Publishedlocusnames, nalleles, ploidy (2 = diploid, 1 = haploid), alleles. The tibble will be named "LocusControl"
}
\description{
This function connects to the Gene Conservation Lab's Oracle database Loki and creates a "LocusControl" object
}
\examples{

CreateLocusControl.GCL(markersuite = "UCI_Chinook_GTSeq_557SNPs", locusnames = NULL, username = "awbarclay", password = "mypassword")


}
