% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_comb_bayes_output.r
\name{custom_comb_bayes_output}
\alias{custom_comb_bayes_output}
\title{Custom Combine BAYES Output}
\usage{
custom_comb_bayes_output(
  groupvec,
  group_names,
  maindir,
  mixvec,
  ext = "RGN",
  nchains = 5,
  burn = 0.5,
  alpha = 0.1,
  threshhold = 5e-07,
  plot_trace = FALSE,
  ncores = 4
)
}
\arguments{
\item{groupvec}{Numeric vector indicating the reporting group affiliation.
If summarizing for the original reporting groups using the RGN output, groupvec is seq(length(group_names)).
If summarizing for new groups using the BOT output, groupvec is the length of the number of populations in the baseline.}

\item{group_names}{Character vector of group names where length(group_names) == max(groupvec).}

\item{maindir}{The directory where the mixture folders are located.
\strong{Note: the results for each mixture must be in their own folder with the same name as the mixture.}}

\item{mixvec}{Character vector of mixture sillys, used to read in the BAYES output (.RGN or .BOT).}

\item{ext}{The extension of the BAYES output file you want to read in and summarize, can either be "RGN" (reporting group estimates) or "BOT" (population estimates).
\strong{Note: use "BOT" when resummarizing.}}

\item{nchains}{The number of chains to summarize (there should be output files for each chain).}

\item{burn}{The proportion of iterations to drop from the beginning of each chain.
For example, for 40,000 iterations setting burn = 0.5 will drop the first 20,000 iterations.}

\item{alpha}{Numeric constant specifying credibility intervals, default is 0.1, which gives 90\% CIs (i.e. 5\% and 95\%).}

\item{plot_trace}{Logical switch, when TRUE will create a trace plot for each mixture chain and repunit (reporting group).
The trace plots can take a while to render depending on the number of mixtures, reporting groups, chains, and iterations you are summarizing.}

\item{ncores}{A numeric vector of length one indicating the number of cores to use.}

\item{threshold}{Numeric constant specifying how low stock comp is before assuming 0, used for \code{P=0} calculation.}
}
\value{
Returns a tibble with 8 fields for each mixture and repunit (reporting group):
\itemize{
\item mixture_collection: Factor of mixtures (only a factor for ordering, plotting purposes).
\item repunit: Factor of reporting groups (only a factor for ordering, plotting purposes).
\item mean: Mean stock composition.
\item sd: Standard deviation.
\item median: Median stock composition.
\item loCI: Lower bound of credibility interval.
\item hiCI: Upper bound of credibility interval.
\item P=0: The proportion of the stock comp distribution that was below \code{threshold} (i.e., posterior probability that stock comp = 0).
\item gr: The Gelman-Ruban shrink factor for checking among-chain convergence.
}

Also returns a trace plot for each mixture and repunit if \code{plot_trace = TRUE}.
}
\description{
This function computes summary statistics from \code{BAYES} output, similar to custom_comb_rubias_output().
}
\examples{
\dontrun{

  setwd("V:/Analysis/2_Central/Sockeye/Cook Inlet/2012 Baseline/Mixture/2016 UCIfisheryMixtures")
  attach("2016UCIfisheryMixtureAnalysis.RData")
  groupvec <- groupvec
  group_names <- groups
  mixvec <- mixvec
  detach()
  
  maindir <- "BAYES/Output"

  results <- custom_comb_bayes_output(groupvec = groupvec, group_names = group_names, maindir = maindir, mixvec = mixvec, ext = "BOT", nchains = 5, burn = 0.5, alpha = 0.1, threshold = 5e-7, plot_trace = TRUE, ncores = 8)

}

}
\seealso{
custom_comb_rubias_output()
}
