% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_comb_hwler_output.r
\name{custom_comb_hwler_output}
\alias{custom_comb_hwler_output}
\title{Custom Combine HWLER Output}
\usage{
custom_comb_hwler_output(
  groupvec,
  groupnames,
  maindir,
  mixvec,
  ext = "BOT",
  burn = 0.5,
  alpha = 0.1,
  PosteriorOutput = TRUE
)
}
\arguments{
\item{groupvec}{A numeric vector indicating the reporting group affiliation for each population in the baseline, where \code{groupvec} is the same length as the number of populations in the known baseline..}

\item{groupnames}{A character vector of group names corresponding to the groups in the baseline, where \code{length(groupnames) == max(groupvec)}.}

\item{maindir}{Directory where the mixture folders are located, where the results for each mixture must be in their own folder with the same name as the mixture.}

\item{mixvec}{A character vector of mixture silly codes, used to read in the HWLER output.}

\item{ext}{Extension of the HWLER output file you want to read in and summarize, can either be "RGN" (reporting group estimates) or "BOT" (population estimates). Use "BOT" when resummarizing.}

\item{burn}{Proportion of iterations to drop from the beginning of each chain. For example, for 40,000 iterations setting \code{burn = 0.5} (default) will drop the first 20,000 iterations.}

\item{alpha}{Numeric constant specifying credibility intervals; default is 0.1, which gives 90\% CIs (i.e., 5\% and 95\%).}

\item{PosteriorOutput}{A logical value indicating whether to output the posterior values.}
}
\value{
A list of combined HWLER output for each mixture, including summary statistics and posterior output (if requested).
}
\description{
This function computes summary statistics from HWLER (Hardy-Weinberg and Linkage Equilibrium Sampler) output.
}
\details{
This function makes all accommodations for the extra baseline group.
}
\examples{
GCLr::custom_comb_hwler_output()

}
