% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\docType{data}
\name{drv}
\alias{drv}
\title{Oracle Driver for RJDBC}
\format{
An object of class \code{JDBCDriver} of length 1.
}
\usage{
drv
}
\description{
This is the oracle driver that is used by all \code{GCLr} functions that use the \code{RJDBC} package to pull data rom Loki.
}
\details{
This object was created from the ojdbc8.jar file and saved as a data object in the \code{GCLr} package.
}
\keyword{datasets}
