% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_baseline.R
\docType{data}
\name{ex_baseline}
\alias{ex_baseline}
\title{Example Rubias Baseline Data}
\format{
A tibble data frame with 2,443 rows and 190 columns:
\describe{
\item{sample_type}{sample type = reference}
\item{repunit}{reporting unit (aka reporting group)}
\item{collection}{baseline collection name}
\item{indiv}{individual ID}
\item{One_ACBP-79 to One_MHC2_190.One_MHC2_251.1}{genotype data for 93 SNP loci}
}
}
\usage{
ex_baseline
}
\description{
This is an example \code{rubias} baseline (aka reference) dataset containing genotypes for 14 baseline collections.
}
\keyword{datasets}
