% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_mixtures.R
\docType{data}
\name{ex_mixtures}
\alias{ex_mixtures}
\title{Example \code{rubias} Mixture Data}
\format{
A tibble data frame with 1,517 rows and 190 columns:
\describe{
\item{sample_type}{sample type = mixture}
\item{repunit}{reporting unit (aka reporting group)}
\item{collection}{mixture collection name}
\item{indiv}{individual ID}
\item{One_ACBP-79 to One_MHC2_190.One_MHC2_251.1}{genotype data for 93 SNP loci}
}
}
\usage{
ex_mixtures
}
\description{
This is an example \code{rubias} mixture dataset conatining genotypes for 16 mixtures.
}
\keyword{datasets}
