% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_pairwise_fst.R
\docType{data}
\name{ex_pairwise_fst}
\alias{ex_pairwise_fst}
\title{Pairwise Fst Matrix}
\format{
A named matrix with 14 rows and 14 columns
}
\usage{
ex_pairwise_fst
}
\description{
This is an example pairwise Fst matrix for creating trees and mds plots
}
\keyword{datasets}
