% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_photo_data.r
\name{extract_photo_data}
\alias{extract_photo_data}
\title{Extract photo data from photos taken on a GPS enabled device and plot on map.}
\usage{
extract_photo_data(folder, data.file = NULL, plot.map = TRUE, map.file = NULL)
}
\arguments{
\item{folder}{The path to where the photos are located. The folder must contain files that have ".jpg" extensions.}

\item{data.file}{The name of the output data file including .csv extension. If set to NULL, no file is written.}

\item{plot.map}{If TRUE, then an interactive map will appear in the plot window (see details).}

\item{map.file}{The name of the output map widget file including .html extension. If set to NULL, then no file is written.
If you do not supply the full file path, the file will be saved to the supplied photos folder (see details).}
}
\value{
A tibble containing the following variables: FileName, latitude, longitude, date, time.
}
\description{
This function extracts file name, latitude, longitude, and date and time from photos taken on a digital device with GPS capability. If the device used to take the photos did not have GPS capability, then only the file name and data and time take will be extracted and a message will print in the console.
The interactive map contains points for each photo containing GPS coordinates. When you hover over the point, the image file name is displayed, and if you click on a point, the image appears in a popup window.
}
\details{
Most smart phones record GPS coordinates when taking photos and a lot of newer digital cameras also have GPS capability. However, most devices have to be set up to record this information. For instance, on an iphone you need to have location services turned on or it won't record the location of a photo.
One issue with this function is that the rotation of the popup image on the map is not always correct in the plot window; however, it appears to be correct in the .html file. Not sure why this happens or how to fix this issue.
Another thing to note is that \code{plot.map} and \code{map.file} work independently, so you can produce a map file without plotting in R or plot in R and don't produce a file.
}
\examples{
extract_photo_data(folder = "V:\\\\Library\\\\Media\\\\Photos\\\\2018 Trips\\\\Coho Baseline\\\\Glacier Cr", 
                   data.file = "PhotoData.csv", 
                   plot.map = TRUE, 
                   map.file = "PhotoData_map.html")

}
