% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_alt_species.r
\name{find_alt_species}
\alias{find_alt_species}
\title{Identify Alternate Species}
\usage{
find_alt_species(sillyvec, species = "chum", LocusCtl = LocusControl)
}
\arguments{
\item{sillyvec}{A character vector of silly codes (e.g., c("SYEHR07", "SWENA98", "STATS05")).}

\item{species}{The species you are working with; options = sockeye, chum; default = chum.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
A plot of Failed vs Alternate and a tibble containing:
\itemize{
\item silly_fish: The fish ID.
\item prop_alter: The proportion of alternate markers.
\item prop_fail: The proportion of failed markers.
\item num_alter: The number of non-missing alternate markers used for analyses.
}
}
\description{
This function identifies the likelihood of incorrect species based on a few identifiable markers. Specifically, it works with chum and sockeye.
}
\details{
This function requires a \code{LocusControl} object and ."gcl" objects. Run [GCLr::create_locuscontrol[] and \link{loki2r} prior to this function.
The Alternate and Failed marker lists were text files in a random folder on a random network drive; marker lists were hard-coded into function so it doesn't relying on hidden files. In the future, if applicable, we can simply add more markers here.
}
\examples{
\dontrun{

wrong_spp <- GCLr::find_alt_species(sillyvec = sillys, species = "sockeye")

}

}
