% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers_test_combinations.r
\name{fisher_combinations}
\alias{fisher_combinations}
\title{Generate pooling tests}
\usage{
fisher_combinations(sillys, min_length = 2, max_length)
}
\arguments{
\item{sillys}{vector of "silly" elements.}

\item{min_length}{The minimum length of combinations to generate. Default is 2 for pairs of sillys.}

\item{max_length}{The maximum length of combinations to generate.}
}
\value{
a list item containing each of the tests
}
\description{
This function generates all possible pooling tests for a list of silly codes. The output is plugged into fishers_test(freq = x, loci = x, tests = combos)
}
\examples{
silly_elements <- c("silly1", "silly2", "silly3", "silly4", "silly5", "silly6")
min_length <- 2
max_length <- 6
combos <- fisher_combinations(silly_elements, min_length, max_length)

}
