% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCLr-package.R
\docType{package}
\name{GCLr-package}
\alias{GCLr}
\alias{GCLr-package}
\title{GCLr: Genetic Analysis Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Genetic analysis tools to streamline genetic data analyses performed by Gene Conservation Lab (GCL) staff.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/commfish/GCLr}
}

}
\author{
\strong{Maintainer}: Andy Barclay \email{andy.barclay@alaska.gov}

Authors:
\itemize{
  \item Andy Barclay \email{andy.barclay@alaska.gov}
}

}
\keyword{internal}
