% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop2gcl.r
\name{genepop2gcl}
\alias{genepop2gcl}
\title{Convert GENEPOP File to ".gcl" Objects}
\usage{
genepop2gcl(filename)
}
\arguments{
\item{filename}{The full file path of the GENEPOP file.}
}
\value{
Assigns ".gcl" and LocusControl objects to the current workspace and returns a vector of the ".gcl" objects created without the ".gcl" extension (i.e., \code{sillyvec}).
}
\description{
Create ".gcl" and \code{LocusControl} objects from a GENEPOP file.
}
\details{
An allele conversion argument may be added in the future if there is a need (mainly for SNPs). This would require a conversion input file from the user to define the corresponding alleles for each locus (e.g., 1 = A, 2 = C, 3 = G, 4 = T).
}
\examples{
GCLr::genepop2gcl(filename = system.file("genepop", "ex_genepop.txt", package = "GCLr"))

}
