% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_haplotypes_for_loki.R
\name{generate_haplotypes_for_loki}
\alias{generate_haplotypes_for_loki}
\title{Generate all possible haplotypes for GT-seq amplicons}
\usage{
generate_haplotypes_for_loki(input, output)
}
\arguments{
\item{input:}{A data frame containing the following columns:
\itemize{
\item SNP (optional)
\item LOCUS_NAME: The identifier/name for each locus.
\item SNPpos: Used for sorting, first SNP in the amplicon, then next, etc.
\item Allele1: The first allele at a given SNP, already alphabetical
\item Allele2: The second allele at a given SNP, already alphabetical
}}

\item{output:}{Path to the output CSV file. Note that it must contain the
filename and extension as well (e.g., ~/../Desktop/output_haplos.csv)}
}
\value{
A data frame containing the following columns:
\itemize{
\item LOCUS: Locus name
\item VALUE: Combined alleles for each haplotype
}
}
\description{
This function takes a data frame containing SNP information and
generates all possible haplotype combinations for GT-seq amplicons.
}
\examples{
# Example Usage
my_haplotypes <- generate_haplotypes_for_loki(input = "path/to/input.csv", output = "path/to/output.csv")
}
