% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asl_data.r
\name{get_asl_data}
\alias{get_asl_data}
\title{Get ASL Data From Loki}
\usage{
get_asl_data(sillyvec, username, password, file = NULL, import.vars = TRUE)
}
\arguments{
\item{sillyvec}{a character vector of silly codes}

\item{username}{your state user name}

\item{password}{your password used to access Loki. Contact Eric Lardizabal if you don't have a password for Loki.}

\item{file}{The file path, including .csv extension, for writing out a csv file of the output. If set to NULL, no file will be saved.}

\item{import.vars}{Do you only want the variables allowed for the ASL importer? (default: TRUE)}
}
\value{
If \code{import.vars = TRUE}, the function outputs a tibble containing the following variables:
\itemize{
\item \code{Collection ID}
\item \code{Fish ID}
\item \code{Freshwater Age}
\item \code{Ocean Age}
\item \code{Sex}
\item \code{Length}
\item \code{Weight}
\item \code{Scale Card Number}
\item \code{Scale Card Position}
\item \code{ASL Number}
}
If \code{import.vars = FALSE}, a variable for \verb{silly code} is added to the output tibble.
}
\description{
This function connects to Loki and pulls the fish information (ASL data) for each silly in sillyvec. This pulls the same information as the "ASL Import" report in OceanAK.
}
\details{
The output of this function can be used for creating an import file for the Loki ASL Data Importer. The the importer requires that all columns are in the correct order and spelled correctly. The importer will not work if you include silly code in the import file.  When writing the import file, make sure it does not contain NAs When writing the import file, make sure it does not contain NAs. (see example below) # Also, when using readr::write_csv to write out the tissue import file, make sure to change the eol argument to from the default \verb{\\n} to \verb{\\r\\n} or the importer
Also, if you use \code{\link[readr:write_csv]{readr::write_csv()}} to write out the tissue import file, make sure to change the eol argument to from the default \verb{\n} to \verb{\r\n} or the importer will give you an error message about the header names.  e.g.,  \verb{import_file \%>\% write_csv(file = "ImportFile.csv", na = "", eol = "\r\n")}
}
\examples{
\dontrun{

 GCLr::get_asl_data(sillyvec = c("PPORTCH21", "PBARABCR21"), username = "awbarclay", password = password, import.vars = FALSE, file = path.expand("~/test_fish_table.csv"))

}

}
