% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fish_inventory.R
\name{get_fish_inventory}
\alias{get_fish_inventory}
\title{Get Sample Inventory Plots}
\usage{
get_fish_inventory(years, dir, username, password)
}
\arguments{
\item{dir}{the directory where you want the plot .png files to be saved. \strong{Note: the files are named automatically, so just the path to the folder here.}}

\item{username}{your state user name}

\item{password}{your password used to access LOKI; see Eric Lardizabal if you need to set up a password}

\item{year}{the years you want to look at. Years must be greater than 2016 or the function will give an error (see details).}
}
\description{
This function pulls a sample inventory from Loki and produces 3 plots.
}
\details{
This function requires a build time stamp for for each collection to work. Build time stamps were not added to Loki until 2017, so this function will only work for years >= 2017.
This function creates three plots:

\describe{

\item{Plot 1}{A barplot of the number of samples in Loki by species and year}
\item{Plot 2}{A barplot of the overall number of samples in Loki by species for the years supplied}
\item{Plot 3}{A heatmap of the number of sample by month and year, with the overall average by month for the years supplied}

}
}
\examples{
\dontrun{

get_fish_inventory(years = 2018:2023, username = "myusername", password = "mypassword")

}
}
