% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geno.r
\name{get_geno}
\alias{get_geno}
\title{Get Genotypes}
\usage{
get_geno(
  project_name = NULL,
  sillyvec = NULL,
  loci = NULL,
  file,
  username,
  password,
  open.file = FALSE
)
}
\arguments{
\item{project_name}{A character vector of GCL project names to pull from LOKI (default = NULL).}

\item{sillyvec}{A character vector of silly codes without the ".gcl" extension (default = NULL).}

\item{loci}{A character vector of locus names as spelled in LOKI (default = NULL).}

\item{file}{The full file path with .csv extension for writing output genotypes file; with "\\" or "/" separator between folder.}

\item{username}{Your user name for accessing LOKI.}

\item{password}{Your password for accessing LOKI.}

\item{open.file}{Logical, if set to TRUE, the .csv file will open after it has been written (default = FALSE).}
}
\value{
A slim, tibble and a UTF-8 coded .csv file with single column genotypes in a wide format (1 column per locus):
\itemize{
\item \code{LAB_PROJECT_NAME}: the lab project name
\item \code{FK_COLLECTION_ID}: the collection ID (internal to LOKI)
\item \code{SILLY_CODE}: the silly code
\item \code{FK_FISH_ID}: the fish ID
\item \code{DNA_PLATE_ID}: the plate ID that was genotyped (only if \code{project_name} is given)
\item \code{Locus_1}: genotype for the first locus, 1 column (i.e., all alleles, separated by "/")
\item \code{Locus_n}: genotype for the nth locus, 1 column (i.e., all alleles, separated by "/")
}
}
\description{
This function pulls a genotypes report from LOKI and writes the data to a UTF-8 coded .csv file.
}
\details{
This function requires the R package "RJDBC" for connecting to LOKI.
It connects to LOKI using the provided \code{username} and \code{password}, and retrieves genotypes and and slim attributes.
The genotypes report will only contain DNA plate IDs if project_name is supplied.
}
\examples{
\dontrun{
Geno_locisilly <- get_geno(sillyvec = sillyvec, loci = loci, file = "TestGenotypesReport_sillyloci.csv", username = username, password = password, project_name = NULL)
Geno_silly <- get_geno(sillyvec = sillyvec, loci = NULL, file = "TestGenotypesReport_silly.csv", username = username, password = password, project_name = NULL)
Geno_proj <- get_geno(project_name = "K205", file = "TestGenotypesReport_proj.csv", username = username, password = password, open.file = TRUE)
}

}
