% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtseq_metadata.r
\name{get_gtseq_metadata}
\alias{get_gtseq_metadata}
\title{Get metadata from GT-seq projects}
\usage{
get_gtseq_metadata(
  project_name,
  dir,
  file_name,
  username,
  password,
  open.file = FALSE
)
}
\arguments{
\item{project_name}{a character vector of one or more GT-seq project names as they are spelled in LOKI. e.g., c("P014","P015","P016") (see details)}

\item{dir}{the directory for writing output file with "\\" separator between folders. e.g., "V:\\Analysis\\Staff\\Andy Barclay\\R"}

\item{file_name}{the name of the csv file. (see details)}

\item{username}{your user name for accessing LOKI through R, a.k.a your state username}

\item{password}{our password for accessing LOKI through R.}

\item{open.file}{logical,if set to TRUE the CSV file will open after it has been written. Default is FALSE.}
}
\value{
a tibble of metadata by locus (crosstab/wide format) and writes out to one or more .csv files (see details)
}
\description{
This function pulls a GT-seq project metadata report from Loki
}
\details{
This function is only intended for pulling metadata for GT-seq projects. If a project name is supplied that is not a GT-seq project, the function will still produce files, but there will be NA's in most of the fields.
Reports that exceed the maximum number of rows for Excel (1048575 rows max) will be split into multiple files with a number added to the file name (e.g. test_metadat_1.csv, test_metadat_2.csv)
}
\examples{
\dontrun{
  metadata <- get_gtseq_metadata(project_name = "K158", dir = path.expand("~"), file_name = "test_metadat", username = "awbarclay", password = "************", open.file = FALSE)
}

}
