% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtseq_sample_sheet.R
\name{get_gtseq_sample_sheet}
\alias{get_gtseq_sample_sheet}
\title{Get a GTSeq Project Sample Sheet}
\usage{
get_gtseq_sample_sheet(project, i7_NAME_start, dir, username, password)
}
\arguments{
\item{project}{the project name from Loki iStrategy (e.g., "P015")}

\item{i7_NAME_start}{a character string with three characters (i.e., \code{nchar(i7_NAME_start) = 3}) and leading zeros.}

\item{dir}{the directory (folder) where you want to write out the sample sheet .csv file. \emph{Note: The file will be named automatically using the supplied project name.}}

\item{username}{your state user name}

\item{password}{your password used to access LOKI; see Eric Lardizabal if you need to set up a password}
}
\value{
This function writes out a .csv file containing the following fields:
\itemize{
\item \strong{Project}: project name
\item \strong{SILLY_FISH_ID}: silly code and fish number
\item \strong{Plate_ID}: extraction plate ID number
\item \strong{i7name}: i7 barcode name for each plate
\item \strong{i7sequence}: i7 barcode sequence
\item \strong{WellID}: extraction plate well ID (e.g., A1, B1, C1, ....., H12)
\item \strong{i5name}: i5 barcode name for each fish
\item \strong{i5sequence}: i5 barcode sequence
\item \strong{Panel}: the marker panel name (marker suite)
}
}
\description{
This function connects to Loki and gets the GTSeq sample sheet for a given project.
i7 barcodes are added to the sheet starting with the supplied \code{i7_NAME_start}. (see details)
}
\details{
For this function to work, the project information must be entered into \link{iStrategy} first.
i7 barcodes come in individual vials and get used up over time.
To make sure the vials are used evenly (i.e., run out at the same time), each project will start with the next i7_NAME after the last i7 name that was use in the previous project.
To determine which what i7 barcode that was used last, look 'NextSeq Flow Cell Log.xlsx' located here \link[="S:\\GCL\\Lab\\Genotyping\\SNP Projects"]{"S:\GCL\Lab\Genotyping\SNP Projects"}("S:/GCL/Lab/Genotyping/SNP Projects").
The last project should be at the bottom of the sheet if the "Run Date" field is sorted lowest to highest. The "i7s" field will contain the range of i7s used in the project.
}
\examples{
\dontrun{

  get_gtseq_sample_sheet (project = "P015", i7_NAME_start = "001", dir = path.expand("~"), username = "awbarclay", password = scan(path.expand("~/R/usr.pw"), what = "")[2])

}
           
}
