% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenotypes.r
\name{get_phenotypes}
\alias{get_phenotypes}
\title{Get All Combinations of Phenotypes}
\usage{
get_phenotypes(markerset, LocusCtl = LocusControl)
}
\arguments{
\item{markerset}{is a vector of loci (e.g., c("Ots_vatf-251", "Ots_ZR-575"))}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}}
}
\value{
a vector of phenotypes
}
\description{
This function gets all possible combinations of phenotypes for set of loci.
}
\note{
this function is called on by \code{\link[=combine_loci]{combine_loci()}} and requires a LocusControl object. Run \code{\link[=create_locuscontrol]{create_locuscontrol()}} prior to this function.
}
\examples{

GCLr::get_phenotypes(markerset = c("One_vatf-214", "One_ZNF-61"), LocusCtl = GCLr::ex_LocusControl)

}
