% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_locations.r
\name{get_tissue_locations}
\alias{get_tissue_locations}
\title{Get Tissue Data from Loki}
\usage{
get_tissue_locations(
  unit,
  username,
  password,
  bad_locations = TRUE,
  all_tissues = TRUE
)
}
\arguments{
\item{unit}{the storage unit(s) you're interested in; can be single or multiple units (e.g., "99" or c("WA","WB","WC")) (see details)}

\item{username}{your Loki username}

\item{password}{your Loki password}

\item{bad_locations}{TRUE or FALSE, do you want a list of tissues with incorrect location information (default = FALSE)}

\item{all_tissues}{TRUE or FALSE, do you want a list of ALL tissues (default = FALSE) (see details)}
}
\value{
this function assigns the following objects to your workspace:
\itemize{
\item \code{tissuemap}: an object mapping out all the tissues for a storage unit
\item \code{all_tissues}: an object consisting of all tissues Loki, within a storage unit
\item \code{bad_locations}: an object consisting of all tissues within Loki, containing unexpected, missing, or otherwise incorrect location information (i.e., SHELF_RACK and/or SLOT)
}
}
\description{
This function pulls tissue data from Loki, for a given storage location, and updates the tissue location maps.
}
\details{
This function best when run with a single type of unit (e.g., all of B7, all of Warehouse).
You can combine but the output could get messy due to different shelf naming conventions.
If you set \code{all_tissues = TRUE} the resulting output will be very large!
}
\examples{
\dontrun{
 username <- "awesomeuser"
 password <- "awesomepassword1"
 unit <- c("A", "B", "C", "D", "E", "F") # We want these units in B7
 bad_locations <- TRUE # yes, I want to identify all wrong/missing locations
 all_tissues <- TRUE # yes, I want all the data
 
 TissueLocations_2R.GCL(unit = unit, username = username, password = password, bad_locations = bad_location, all_data = all_data)
}
}
