% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierfstat_global_stats.r
\name{hierfstat_global_stats}
\alias{hierfstat_global_stats}
\alias{HierFstats_global.GCL}
\title{Compute global hierarchical F-statistics and variance components.}
\usage{
hierfstat_global_stats(levels, genotypes)
}
\arguments{
\item{levels}{A data frame containing the different levels (factors) from the outermost (e.g., region) to the innermost before the individual.}

\item{genotypes}{A data frame containing the genotypes in single-column numeric format (e.g., 101, 102, 202).}
}
\value{
A list containing 3 tibbles:
\itemize{
\item \code{loci} - variance components by locus,
\item \code{overall} - variance components summed over all loci,
\item \code{F} - hierarchical F-statistics (see \code{\link[hierfstat:varcomp.glob]{hierfstat::varcomp.glob()}} documentation).
}
}
\description{
This is a wrapper function for \code{\link[hierfstat:varcomp.glob]{hierfstat::varcomp.glob()}}. This wrapper will compute variance components and hierarchical F-statistics for haploid and diploid loci and combine the results into a single output.
}
\examples{
sillyvec <- GCLr::base2gcl(GCLr::baseline)

region <- GCLr::baseline \%>\%
  dplyr::group_by(collection) \%>\%
  dplyr::filter(dplyr::row_number()==1) \%>\%
  dplyr::pull(repunit) \%>\%
  factor() \%>\%
  as.numeric()

loci <- GCLr::LocusControl$locusnames

fstat.dat <- GCLr::create_hierfstat_data(sillyvec = sillyvec, region = region, pop = seq_along(sillyvec), loci =  GCLr::LocusControl$locusnames, ncores = 4)

hierfstat_global_stats(levels = fstat.dat[ ,1:3], genotypes = fstat.dat[ ,-c(1:3)])

}
