% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_pca.R
\name{indiv_pca}
\alias{indiv_pca}
\title{Individual Principal Component Analysis (PCA)}
\usage{
indiv_pca(file, groupvec, groups, ncomp = 3)
}
\arguments{
\item{file}{the file path of a GENEPOP file with .gen extension containing individual genotypes. Make sure that each individual in the file has a unique identifier (e.g., popA_1, popA_2, popA_3). If using \code{\link[=gcl2genepop]{gcl2genepop()}} to create the GENEPOP file, set \code{VialNums = TRUE}.}

\item{groupvec}{A numeric vector indicating the group affiliation of each collection in the GENEPOP file.}

\item{groups}{A character vector of group names with a length equal to the maximum value in \code{groupvec}}

\item{ncomp}{The number of principal coordinates to include in the output (default = 3).}
}
\value{
A named list with three elements:
\itemize{
\item \code{pca.raw}: the raw output from \code{\link[ade4:dudi.pca]{ade4::dudi.pca()}}
\item \code{pca.df}: a tibble containing group, collection, indiv (unique individual identifier), and coordinates for each principal component.
\item \code{var_exp}: the variance explained for all principal components (calculated from eigen values in pca.raw)
}
}
\description{
This function reads in a GENEPOP file and performs an individual Principal Component Analysis.
}
\details{
This function requires the adegenet and ade4 packages.
For this function to work properly, the GENEPOP file must have a ".gen" extension (required by \code{\link[adegenet:read.genepop]{adegenet::read.genepop()}}), and
each individual must have a unique identifier (see "file" argument description). The output of this function can be supplied to \code{\link[=plot_indiv_pca]{plot_indiv_pca()}} to plot the first 3 principal components of the analysis in an interactive 3-dimensional scatter plot.
The variance explained by each principal coordinate are calculated from the eigen values in the raw PCA output (i.e., (pca.raw$eig)/sum(pca.raw$eig)).
}
\examples{

group_info <- GCLr::ex_baseline \%>\%
  dplyr::mutate(collection = factor(collection, levels = unique(collection))) \%>\% 
  dplyr::group_by(repunit, collection) \%>\%
  dplyr::summarise(group = unique(repunit) \%>\% factor()) \%>\% 
  dplyr::arrange(collection)

file <- system.file("genepop", "ex_genepop.gen", package = "GCLr")

groups <- group_info$group \%>\% unique()

groupvec <- group_info$group \%>\%
  factor(levels = groups) \%>\%
  as.numeric()

indiv_pca(file = file, groupvec = groupvec, groups = groups, ncomp = 3)

}
\seealso{
\code{\link[ade4:dudi.pca]{ade4::dudi.pca()}}

\code{\link[=plot_indiv_pca]{plot_indiv_pca()}}
}
