% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOKI2R.GCL.R
\name{LOKI2R.GCL}
\alias{LOKI2R.GCL}
\title{This function connects to the Gene Conservation Lab Oracle database Loki and creates a "gcl" object containing genotypes and paired data for each sample.}
\usage{
LOKI2R.GCL(
  sillyvec,
  username,
  password,
  test_type = c("SNP", "GTSNP", "MSAT")[1],
  include_missing = FALSE
)
}
\arguments{
\item{sillyvec}{a character vector of silly codes (e.g. sillyvec <- c("KQUART06","KQUART08","KQUART09"))}

\item{username}{your Oracle user name}

\item{password}{your Oracle password (contact Eric Lardizabal if you don't have a username or password)}

\item{test_type}{the test type ("SNP", "GTSNP", "MSAT) you would like to pull from Loki - default is test_type = "SNP". This argument is a workaround. Some collections in Loki contain data for both SNP and GTSNP test types. If genotypes for both test types were pulled at the same time for the same individuals and loci, this would cause the function to stop and throw an error message. If you want data for both test types in your workspace, you will need to run this function separately for each test type. If you pull data for both test types for the same silly code, you will need to rename your '.gcl' object for the first test type before pulling data for the second test type, otherwise the first '.gcl' object will be overwritten.}

\item{include_missing}{logical statement whether to include all samples even if they were never genotyped for all loci in LocusControl.}
}
\value{
This function returns a tibble of two-column genetic data with 19 columns of sample attributes preceding it.
}
\description{
An object, similar to the rubias input format, is created for each collection (a.k.a silly) in sillyvec containing genotypes for each locus in LocusControl$locusnames. The default for this function is to only include fish that have been genotyped for all loci in LocusControl$locusnames; however, the function will include all fish when include_missing = TRUE, which is not recommended or needed for most analyses.
}
\note{
This function requires a LocusControl object. Run CreateLocusControl.GCL prior to this function.
}
\examples{

CreateLocusControl.GCL(markersuite = "Sockeye2011_96SNPs", username ="awbarclay", password = "mypassword")

sillyvec <- c("SUCIWS06", "SUCIWS07", "SUCIWS08", "SUCIWS09", "SUCIWS10", "SUCIWS11", "SUCIWS12", "SUCIWS13", "SCIMA22")

LOKI2R.GCL(sillyvec = sillyvec, username = "awbarclay", password = .password, test_type = "SNP", include_missing = TRUE)

}
