% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loki2r.r
\name{loki2r}
\alias{loki2r}
\alias{LOKI2R.GCL}
\title{Pull Genotypes by Collection}
\usage{
loki2r(
  sillyvec,
  username,
  password,
  test_type = c("SNP", "GTSNP", "MSAT")[1],
  include_missing = FALSE
)

LOKI2R.GCL(
  sillyvec,
  username,
  password,
  test_type = c("SNP", "GTSNP", "MSAT")[1],
  include_missing = FALSE
)
}
\arguments{
\item{sillyvec}{a character vector of silly codes without the .gcl extension.}

\item{username}{your state user name}

\item{password}{your password used to access LOKI; see Eric Lardizabal if you need to set up a password}

\item{test_type}{the test type ("SNP" or "GTSNP") you would like to pull from Loki. (default = "SNP")}

\item{include_missing}{whether to include all fish even if they were never genotyped for all loci in LocusControl (default = FALSE)}
}
\value{
This function assigns a tibble with the following columns for each silly:

\itemize{
\item \verb{Columns 1-19:}
\itemize{
\item \verb{FK_FISH_ID (double): fish ID numbers for each individual}
\item \verb{COLLECTION_ID (double): the unique collection ID number for each individual}
\item \verb{SILLY_CODE (character): the silly code for each individual}
\item \verb{PLATE_ID (character): the extraction plate ID for each individual}
\item \verb{PK_TISSUE_TYPE (character): the tissue type extracted for DNA for each individual}
\item \verb{CAPTURE_LOCATION (character): the location where each individual was captured for sampling}
\item \verb{CAPTURE_DATE (date): the date each individual was captured (e.g., May 5, 2020 = "2020-05-05")}
\item \verb{END_CAPTURE_DATE (date): the last collection date for a silly (e.g., May 5, 2020 = "2020-05-05")}
\item \verb{MESH_SIZE (character): the mesh size of the net used to capture (harvest) each individual}
\item \verb{MESH_SIZE_COMMENT (character): comments about mesh size}
\item \verb{LATITUDE (double): the latitude where each individual was captured in decimal degrees}
\item \verb{LONGITUDE (double): the longitude where each individual was captured in decimal degrees}
\item \verb{AGENCY (character): the name of the agency or organization that collected each individual}
\item \verb{VIAL_BARCODE (character): the barcode on the collection vial}
\item \verb{DNA_TRAY_CODE (character): the barcode on the collection tray/card}
\item \verb{DNA_TRAY_WELL_CODE (double): the unique number assigned to each position in the collection tray/card for each individual (e.g., positions A1-A10 = codes 1-10)}
\item \verb{DNA_TRAY_WELL_POS (character): the position in the collection tray/card (e.g., A1, A2, B1, B2, etc.)}
\item \verb{CONTAINER_ARRAY_TYPE_ID (double): the number code for the collection container (e.g., tray or card)}
\item \verb{SillySource (double): the original silly code and fish ID for each individual (e.g., KQUART06_1). When pulled from loki this will be the SILLY_CODE and FK_FISH_ID}
}
\item \verb{Columns 20+}
\itemize
\item \verb{genotypes with a column for each alelle for each locus}
}
}
\description{
This function connects to LOKI and creates a .gcl object for each silly in \code{sillyvec} containing genotypes for each locus in LocusControl$locusnames.The default for this function is to only include fish that have been genotyped for all loci in \code{LocusControl$locusnames}; however, the function will include all fish when include_missing = TRUE, which is not recommended or needed for most analyses. If a silly has no fish with genotypes for the loci in \code{LocusControl}, no object will be created and a message will appear listing the sillys that had no data.
}
\note{
This function requires a LocusControl object. Run \code{\link[=create_locuscontrol]{create_locuscontrol()}} prior to this function.
}
\examples{
  create_locuscontrol(markersuite = "Sockeye2011_96SNPs", username ="awbarclay", password = .password)#Locus control
  sillyvec <- c("SUCIWS06", "SUCIWS07", "SUCIWS08", "SUCIWS09", "SUCIWS10", "SUCIWS11", "SUCIWS12", "SUCIWS13", "SCIMA22")
  loki2r(sillyvec = sillyvec, username = "awbarclay", password = .password, test_type = "SNP", include_missing = TRUE)
   
}
