% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loki_url.r
\name{loki_url}
\alias{loki_url}
\alias{LOKI_URL.GCL}
\title{Build URL to connect to LOKI}
\usage{
loki_url(test.db = FALSE)
}
\value{
The URL string used by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} to connect to the Oracle cloud database (LOKI)
}
\description{
This function builds the URL to connect to LOKI and is called by other GCL functions that connect to LOKI via \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
}
\note{
this function is used in the background
}
\keyword{internal}
