% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new2old_gcl.r
\name{new2old_gcl}
\alias{new2old_gcl}
\alias{new2old_gcl.GCL}
\title{Convert New Style "*.gcl" objects to Old Style}
\usage{
new2old_gcl(sillyvec, save_new = FALSE, ncores = 4)

new2old_gcl.GCL(sillyvec, save_new = FALSE, ncores = 4)
}
\arguments{
\item{sillyvec}{A character vector of silly codes with associated *.gcl objects that need to be converted}

\item{save_new}{whether you want the new objects overwritten without saving (TRUE) or assign the new objects to ".gcl_new" before overwriting (FALSE)}

\item{ncores}{a numeric vector of length one indicating the number of cores to use}
}
\description{
This function converts new style ".gcl" objects from tibbles to a list of arrays. The new style objects can be saved as ".gcl_new" before they are overwritten.
}
\details{
This function checks if all the required objects exist in the environment and converts the new style tibble "*.gcl" objects to the old style list format. It performs various checks and transformations on the objects, and assigns the converted objects to the workspace.
}
\examples{
create_locuscontrol(markersuite = "Sockeye2011_96SNPs", username = "awbarclay", password = .password) 

sillyvec <- c("SGILL20D11", "SGILL20D6", "SGILL20D8")
loki2r(sillyvec = sillyvec, username = "awbarclay", password = .password)
combine_loci(sillyvec = sillyvec, markerset = c("One_CO1","One_Cytb_17","One_Cytb_26"), update = TRUE)
new2old.GCL(sillyvec = sillyvec, save_new = TRUE, ncores = 8)

}
