% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old2new_locuscontrol.r
\name{old2new_locuscontrol}
\alias{old2new_locuscontrol}
\title{Convert Old LocusControl to New LocusControl Tibble}
\usage{
old2new_locuscontrol(LocCtrl = LocusControl, save_old = FALSE)
}
\arguments{
\item{LocCtrl}{An old LocusControl object (list) to be converted. If the
object is already a tibble, an error will be thrown.}

\item{save_old}{Logical, indicating whether to save the old LocusControl
object as "LocusControl_old" in the global environment. Default is FALSE.}
}
\value{
The function does not return a value explicitly, but it creates a new
LocusControl tibble with updated format and saves it in the global
environment as "LocusControl".
}
\description{
This function converts an old LocusControl object to a new LocusControl tibble
to accommodate changes in the data structure. The old LocusControl object is
expected to be a list, and the new LocusControl tibble will have a more
standardized and tidy format.
}
\details{
This function takes an old LocusControl object and converts it to a new
LocusControl tibble, which consists of two parts. The first part, locus_info,
contains the following columns:

\describe{
\item{MarkerSuite}{The name of the MarkerSuite.}
\item{locusnames}{A character vector of locus names.}
\item{Publishedlocusnames}{A character vector of published locus names.
If the old LocusControl object has a shorter Publishedlocusnames vector
than locusnames, NA values will be used to make them the same length.}
\item{nalleles}{The number of alleles for each locus.}
\item{ploidy}{The ploidy level for each locus.}
}

The second part, alleles, contains the alleles information for each locus.
It has two columns:

\describe{
\item{alleles}{A list column containing the allele numbers for each locus.}
\item{call}{A list column containing the alleles calls for each locus.}
}

Note that this function will save the old LocusControl object as
"LocusControl_old" in the global environment if the \code{save_old} argument
is set to TRUE.
}
\examples{
\dontrun{
load("V:/Analysis/2_Central/Chinook/Cook Inlet/2019/2019_UCI_Chinook_baseline_hap_data/2019_UCI_Chinook_baseline_hap_data.RData")

GCLr::old2new_locuscontrol(LocCtrl = LocusControl, save_old = TRUE)
}

}
\seealso{
\code{\link{LocusControl}}, \code{\link{combine_loci}}
}
