% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_fst.r
\name{pairwise_fst}
\alias{pairwise_fst}
\title{Create Pairwise Fst Matrix}
\usage{
pairwise_fst(sillyvec, loci, inputfile, popnames = NULL, ncores = 4)
}
\arguments{
\item{sillyvec}{a vector of silly codes without the ".gcl" extension (e.g., sillyvec <- c("KQUART06", "KQUART08", "KQUART10"))}

\item{loci}{a character vector of locus names}

\item{inputfile}{the file path of the \pkg{genepop} input file including the .txt extension}

\item{popnames}{optional vector of population names corresponding to \code{sillyvec} to add as the dimnames of the output matrix. If \code{popnames} is not supplied, \code{sillyvec} will be used as the dimnames.}

\item{ncores}{a numeric vector of length one indicating the number of cores to use}
}
\value{
a matrix of pairwise Fst values.
}
\description{
This function generates a matrix of pairwise Fst values using the \pkg{genepop} package.
}
\examples{

sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

loci <- GCLr::ex_baseline[,-c(1:5)] \%>\%
  names() \%>\%
  gsub(pattern = "*\\\\.1", x = ., replacement = "") \%>\%
  unique()

GCLr::pairwise_fst(sillyvec = sillyvec, loci = loci, inputfile = system.file("genepop", "ex_genepop.txt", package = "GCLr"), ncores = parallel::detectCores())

}
