% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allele_freq.r
\name{plot_allele_freq}
\alias{plot_allele_freq}
\title{Plot Allele Frequency Bubble Plots}
\usage{
plot_allele_freq(
  freq,
  file,
  sillyvec = NULL,
  groupvec = NULL,
  loci = NULL,
  group_col = NULL,
  popnames = NULL,
  xlab.cex = 6,
  ylab.cex = 12,
  xtick.int = 1
)
}
\arguments{
\item{freq}{A tibble of allele frequencies generated by \code{\link[=calc_freq_pop]{calc_freq_pop()}}.}

\item{file}{A character vector of length 1 with the full file path including ".pdf" extension where plots will be saved.}

\item{sillyvec}{An optional character vector of silly codes without the ".gcl" extension (default = \code{NULL}).
If \code{NULL}, \code{sillyvec} will come directly from \code{freq}.}

\item{groupvec}{An optional numeric vector indicating the group affiliation of each silly in \code{sillyvec},
must be the same length as \code{sillyvec} (default = \code{NULL}).}

\item{loci}{An optional character vector of locus names (default = \code{NULL}).
If \code{NULL}, \code{loci} will come directly from \code{freq}.}

\item{group_col}{An optional character vector of colors corresponding to each group in \code{groupvec} (default = \code{NULL}).}

\item{popnames}{An optional character vector of collection/population names for each silly in \code{sillyvec},
must be the same length as \code{sillyvec} (default = \code{NULL}).
If \code{NULL}, \code{popnames} will be \code{1:length(sillyvec)}.
Setting \code{popnames} for a large number of pops may cause the x-axis tick labels to become unreadable.}

\item{xlab.cex}{An optional numeric vector of length 1 indicating the x-axis label font size, (default = 6).}

\item{ylab.cex}{An optional numeric vector of length 1 indicating the y-axis label font size, (default = 12).}

\item{xtick.int}{An optional numeric vector of length 1 indicating the x-axis tick interval, (default = 1).
This argument only works when \code{popnames = NULL}.}
}
\value{
A ("*.pdf") file containing allele frequency bubble plots, 1 page per locus
}
\description{
This function plots the results from \code{\link[=calc_freq_pop]{calc_freq_pop()}}. It creates a ("*.pdf") of bubble plots of allele frequencies for each locus.
}
\details{
Allele frequency bubble plots are most useful for loci with multiple alleles (e.g. uSATs or microhaplotypes).
For visualizing allele frequencies for SNPs, use \code{\link[=plot_freq_fis_4snps]{plot_freq_fis_4snps()}}.
}
\examples{
\dontrun{
freqs <- GCLr::calc_freq_pop(sillyvec = sillyvec)
GCLr::plot_allele_freq(freq = freqs, file = "~/R/test_allele_freq.pdf")
}

}
\seealso{
\code{\link[=calc_freq_pop]{calc_freq_pop()}}
\code{\link[=plot_freq_fis_4snps]{plot_freq_fis_4snps()}}
}
