% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_color_tree.r
\name{plot_color_tree}
\alias{plot_color_tree}
\alias{PlotColorTree.GCL}
\title{Plot a Colored Phylogenetic Tree}
\usage{
plot_color_tree(
  color.tree,
  rm.neg.branch = TRUE,
  type = "phylogram",
  scale = TRUE,
  show.tip.label = TRUE,
  edge.width = 3,
  adj = NULL,
  cex = 1,
  font = 1,
  label.offset = 0,
  gen.dist = "FST",
  ...
)
}
\arguments{
\item{color.tree}{a color tree object created with \code{\link[=add_tree_color]{add_tree_color()}}.}

\item{rm.neg.branch}{logical; if TRUE (default), negative branch lengths are removed.}

\item{type}{a character string specifying the type of phylogeny to be drawn. It must be one of "phylogram" (the default), "cladogram", "fan", "unrooted", "radial", or any unambiguous abbreviation of these.}

\item{scale}{logical; whether to add a scale to the tree plot.}

\item{show.tip.label}{logical; if TRUE (default), labels are placed at the tip of each branch.}

\item{edge.width}{a numeric vector giving the width of the branches of the plotted phylogeny. If fewer widths are given than the length of edge, then these are recycled.}

\item{adj}{a numeric specifying the justification of the text strings of the labels: 0 (left-justification), 0.5 (centering), or 1 (right-justification). This option has no effect if type = "unrooted". If NULL (the default), the value is set with respect to the direction (see details in \code{\link[ape:plot.phylo]{ape::plot.phylo()}} documentation).}

\item{cex}{a numeric value giving the factor scaling of the tip labels (Character Expansion).}

\item{font}{an integer specifying the type of font for the labels: 1 (plain text), 2 (bold), 3 (italic, the default), or 4 (bold italic)}

\item{label.offset}{a numeric giving the space between the nodes and the tips of the phylogeny and their corresponding labels. This option has no effect if type = "unrooted".}

\item{gen.dist}{the genetic distance used to produce the tree. If "fst", "FST", or "Fst" is supplied, the "F" will be italicized and the "ST" will be subscripted.}

\item{...}{further arguments to be passed to \code{\link[ape:plot.phylo]{ape::plot.phylo()}}}
}
\value{
This function plots a phylogenetic tree.
}
\description{
This function is a wrapper for \code{\link[ape:plot.phylo]{ape::plot.phylo()}}. It takes a tree object created by \code{\link[=add_tree_color]{add_tree_color()}} and plots it with an optional scale.
}
\examples{
load("V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/Susitna_Chinook_baseline_2020.Rdata")
Fst36 <- read_genepop_pwfst(file = "GENEPOP/Sustina36pops84loci.MIG", popnames = Pooling_1_pops$location)
FstTree <- ape::nj(Fst36)
groupvec <- factor(Pooling_1_pops$group, unique(Pooling_1_pops$group)) \%>\% as.numeric()
colortree <- add_tree_color(tree = FstTree, currentnames = FstTree$tip.label, treenames = FstTree$tip.label, groupvec = groupvec, regioncol = grcol)
plot_color_tree(color.tree = colortree, rm.neg.branch = TRUE, type = "phylogram", scale = TRUE, show.tip.label = TRUE, edge.width = 3, adj = 0.02, cex = 0.7, font = 1, label.offset = 0.001, gen.dist = "FST")

}
