% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confusion_matrix.r
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot Confusion Matrix}
\usage{
plot_confusion_matrix(
  confusion_matrix,
  plot = c("group_group", "pop_group", "pop_pop")[1],
  group_names = NULL,
  pop_names = NULL,
  high_color = "#132B43",
  low_color = "#56B1F7",
  text_color = "yellow"
)
}
\arguments{
\item{confusion_matrix}{a list of confusion matrix tibble(s) produced by \code{\link[=confusion_matrix]{confusion_matrix()}}}

\item{plot}{a character vector of length 1 indicating the type of matrix you want to plot}

\item{group_names}{an optional vector of group names used to keep groups in order; if NULL, the groups will be ordered alphabetically}

\item{pop_names}{an optional vector of population names used to keep pops in order; if NULL, the pops will be ordered alphabetically.}

\item{high_color}{the color used for high "mean scale likelihood" values by \code{\link[ggplot2:scale_fill_gradient]{ggplot2::scale_fill_gradient()}}. Either supply hexadecimal or R color name from \code{\link[=colors]{colors()}}.}

\item{low_color}{The color used for low values by \code{\link[ggplot2:scale_fill_gradient]{ggplot2::scale_fill_gradient()}}. Either supply hexadecimal or color name from \code{\link[=colors]{colors()}}.}

\item{text_color}{The color used by \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} to color the Mean Scaled Likelihood values on the plot.}
}
\value{
A heatmap of mean scaled likelihood.
}
\description{
This function takes the output object from \code{\link[=confusion_matrix]{confusion_matrix()}} and plots a confusion matrix in a heatmap using \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}.
}
\examples{
baseline <- GCLr::ex_baseline
ConfusionMatrices_out <- confusion_matrix(reference = baseline , gen_start_col = 5, output = c("group_group", "pop_group", "pop_pop"))
groups <- baseline$repunit \%>\% unique()
pops <- baseline$collection \%>\% unique()
plot_confusion_matrix(confusion_matrix = ConfusionMatrices_out, group_names = groups, pop_names = NULL, plot = c("group_group", "pop_group", "pop_pop")[1], high_color = "darkblue", low_color = "lightblue", text_color = "yellow")

}
