% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genepop_hwe.r
\name{plot_genepop_hwe}
\alias{plot_genepop_hwe}
\title{Plot Genepop HWE Output}
\usage{
plot_genepop_hwe(
  GenepopHWE_report,
  sillyvec = NULL,
  plot_type = c("silly", "locus")[1]
)
}
\arguments{
\item{GenepopHWE_report}{Raw output from \code{\link[=read_genepop_hwe]{read_genepop_hwe()}}.}

\item{sillyvec}{An optional character vector of silly codes without the ".gcl" extension used to filter results when \code{plot_type = "silly"} (default = \code{NULL}).
If \code{NULL}, \code{sillyvec} will be inherited from \code{GenepopHWE_report}. This argument does not filter results when \code{plot_type = "locus"}.}

\item{plot_type}{A Character vector of length 1 specifying whether to plot p-values by "silly" or "locus" (default = "silly").
If "locus", only loci with an overall p-value across populations < 0.1 will be shown.}
}
\value{
A plot faceted by each population/locus Each facet shows:
\itemize{
\item a histogram of locus- or population-specific p-values
\item the overall p-value across loci/populations printed in red
\item a horizontal red line showing the expected number of loci/populations within each p-value bin by chance
}
}
\description{
This function plots the results from \code{\link[=read_genepop_hwe]{read_genepop_hwe()}}. It displays a histogram of locus- or population-specific p-values and
the overall p-value across loci/populations, faceted by each locus/population.
}
\details{
This function is used to visualize the results from \code{\link[=read_genepop_hwe]{read_genepop_hwe()}}, which is used to test whether collections/populations have significant
deviations from Hardy-Weinberg expectations (HWE; overall p-value < 0.05 with Bonferroni correction for the number of collections in the baseline).
When using this function to test loci for conformance to HWE, we recommend removing loci when the overall p-value < 0.01. However, in both cases
it can also be helpful to see the distribution of locus- or population-specific p-values to determine if there are just a few loci/populations driving the
overall p-value, or if there are several loci/populations that have low p-values. When populations or loci do not conform to HWE, users should investigate
patterns of Fis to determine the root cause.
}
\examples{
\dontrun{
genepop_hwe <- read_genepop_hwe(file = "~/R/test.txt.P", sillyvec = sillyvec)
plot_genepop_hwe(GenepopHWE_report = genepop_hwe, plot_type = "silly")
plot_genepop_hwe(GenepopHWE_report = genepop_hwe, plot_type = "locus")
}

}
\references{
The idea for the plots created by this function comes from:
\itemize{
\item Waples, R.S., 2015. Testing for Hardy–Weinberg proportions: have we lost the plot?. Journal of heredity, 106(1), pp.1-19. \url{https://academic.oup.com/jhered/article/106/1/1/2961870}
}
}
