% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indiv_pca.R
\name{plot_indiv_pca}
\alias{plot_indiv_pca}
\title{Plot a Principal Component Analysis (PCA)}
\usage{
plot_indiv_pca(pca, group_color, title = NULL)
}
\arguments{
\item{pca}{the raw output from \link{indiv_pca}}

\item{title}{the title to place at the top of the plot; if NULL (default) no title will be added to the plot}

\item{group_colors}{a character vector of colors corresponding to each group in the raw PCA output (groups = pca$pca.df$group \%>\% unique())}
}
\value{
A 3-dimensional scatter plot of PCs 1-3
}
\description{
This function takes the raw output from \link{indiv_pca} and plots the coordinates for the first 3 principal components in an interactive 3-dimensional scatter plot.
}
\examples{
\dontrun{
group_info <- GCLr::ex_baseline \%>\%
  dplyr::mutate(collection = factor(collection, levels = unique(collection))) \%>\% 
  dplyr::group_by(repunit, collection) \%>\%
  dplyr::summarise(group = unique(repunit) \%>\% factor()) \%>\% 
  dplyr::arrange(collection)

file <- system.file("genepop", "ex_genepop.gen", package = "GCLr")

file <- "C:/Users/awbarclay/Documents/R/GCLr/inst/genepop/ex_genepop.gen" #DELETE for final version

groups <- group_info$group \%>\% unique()

groupvec <- group_info$group \%>\%
  factor(levels = groups) \%>\%
  as.numeric()

pca_out <- indiv_pca(file = file, groupvec = groupvec, groups = groups, ncomp = 3)

plot_indiv_pca(pca = pca_out, group_color = c("blue", "green", "red"), title = NULL)

}

}
