% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loo_prec_rec.R
\name{plot_loo_prec_rec}
\alias{plot_loo_prec_rec}
\title{Produce a precision-recall curve plot}
\usage{
plot_loo_prec_rec(
  sa_input,
  rates = NULL,
  group_colors,
  by_group = TRUE,
  min_rec = 0.8,
  min_pre = 0.95,
  min_thres = 0.5,
  thres_levels = seq(0.01, 0.99, by = 0.01),
  plots = c("facet", "paired", "both")[3],
  ncores = parallel::detectCores()
)
}
\arguments{
\item{sa_input}{the unmodified output object from \link[rubias:self_assign]{rubias::self_assign}}

\item{rates}{optional; the unmodified output object from \link{loo_roc_rate_calc} created using the supplied sa_input. If left NULL, the rates will be calculated from sa_input. Supplying a rates object will make this function run a lot faster.}

\item{group_colors}{a vector of colors the same length as the number of groups in the input file (i.e., \code{input$repunit \%>\% unique()}).}

\item{by_group}{logical; whether you want the box plot by reporting group (TRUE) or population (FALSE); default = TRUE}

\item{min_rec}{the recall (true positive rate) cutoff for adding a vertical dashed line to the precision plot; default = 0.80}

\item{min_pre}{the precision cutoff for adding a horizontal dashed line to the precision plot; default = 0.95}

\item{min_thres}{minimum assignment likelihood threshold for adding a horizontal line at on the boxplot; default = 0.50}

\item{thres_levels}{a vector of assignment thresholds to include on precision plot (only used when \code{rates = NULL}).}

\item{plots}{the plots do you want the function to produce. There are three choices:
\itemize{
\item{\emph{facet}: }{An iteractive precision-recall curve plot with all reporting groups (i.e., faceted plot}
\item{\emph{paired}: }{Interative plots for each reporting group containing a box plot of assignement likelihoods on the left hand side and a precision-recall curve plot on the right hand side.}
\item{\emph{both}: }{Facet plot first, then paired plots.}
}}

\item{ncores}{the number of cores to use. Default is to use all available cores.}
}
\value{
When \code{plot = "facet"}, the function returns an interactive line plot of True positive rate (y-axis) and Precision (x-axis) faceted by reporting group (aka repunit), with dashed red lines indicating the cutoffs for precision (>= 0.95) and true positive rate (>= 0.80).
The red point on the line indicates where the assignment threshold = 50\%.
When \code{plot = "paired"}, the function returns a list of interactive plots for each reporting group in the leave-on-out analysis.
The left hand side is a box plot of the scaled likelihood of each baseline sample assigning to the reporting group, with a dashed red line indicating the 50\% assignment threshold.
The right hand side is a line plot of True positive rate (y-axis) and Precision (x-axis), with dashed red lines indicating the cutoffs for precision (>= 0.95) and true positive rate (>= 0.80).
The red point on the line indicates where the assignment threshold = 50\%.
When \code{plot = "both"}, the function returns a list of plots with the faceted TPR vs Precision plot as the first element and the paired plots for each reporting group for remaining elements.
When no rates are supplied, an object named 'rates' is assigned to your global environment.
}
\description{
This function takes the unmodified output from \link[rubias:self_assign]{rubias::self_assign}, plots the scaled likelihood for each baseline sample (y-axis) by reporting group or population in a box plot (optional), and plots recall (aka \emph{true positive rate}) (x-axis) by precision (y-axis)
in an interactive plot that can be used to assess reporting groups and assignment thresholds for individual assignment analyses.
}
\details{
Details about the leave-one-out analysis can be found on the \pkg{rubias} GitHub page: \url{https://github.com/eriqande/rubias} or
in Moran and Anderson (2018). Citation: Moran, B.M. and E.C. Anderson. 2019. Bayesian inference from the conditional genetic stock
identification model. Canadian Journal of Fisheries and Aquatic Sciences, 76(4): 551–560.
Recall and precision are defined as follows:
TP = number of true positives, FN = number of false negatives, and FP = number of false positives from the leave-one-out analysis.\itemize{
\item \eqn{Recall = TP / (TP + FN)} Recall is also know at the true positive rate.
\item \eqn{Precision = TP / (TP + FP)}
\item Where
\item \eqn{TP = number of true positives}
\item \eqn{FN = number of false negatives}
\item \eqn{FP = number of false positives}
}
}
\note{
The plots produced by this function are not intended for publications; however, the code within this function can be copied and modified to produce plots tailored for publications.
}
\examples{
sa_input <- rubias::self_assign(reference = GCLr::ex_baseline, gen_start_col = 5)

# Plots are returned as a list when more than one plot is produced 
plots <- plot_loo_tpr_prec(sa_input, group_colors = c("red", "green", "blue"), by_group = FALSE, min_rec = 0.80, min_pre = 0.95, min_thres = 0.50, thres_levels = seq(0.01, 0.99, by = 0.01), plots = c("facet", "paired", "both")[3], ncores = parallel::detectCores())

plots[[1]] # This is how you would look at the first plot in the list
     
}
\seealso{
\link{loo_roc_rate_calc}
}
