% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loo_tpr_prec.R
\name{plot_loo_tpr_prec}
\alias{plot_loo_tpr_prec}
\title{Plot True Positive Rate Vs Precision from Leave-One-Out Analysis}
\usage{
plot_loo_tpr_prec(input, file = NULL)
}
\arguments{
\item{input}{the unmodified output object from \code{GCLr::loo_roc_rate_calc()}}

\item{file}{the file name with .html extension for saving the plots.}
}
\value{
This function returns an interactive plot of True positive rate (y-axis) and Precision (x-axis) faceted by reporting group.
Dashed red lines indicate the cutoffs for precision (>= 0.95) and true positive rate (>= 0.80).
The red point on the line indicates where the assignment threshold = 50\%.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     If a file name is supplied, the function will save the plot to an .html file.
}\if{html}{\out{</div>}}
}
\description{
This function takes the unmodified output from \code{GCLr::loo_roc_rate_calc()},
calculates precision, and plots the true positive rate (y-axis) by precision (x-axis)
in an interactive plot that can be used to determine assignment thresholds for
individual assignment analyses.
}
\details{
\itemize{
\item TP = number of true positives, FN = number of false negatives, FP = number of false positives from the leave-one-out analysis.
\itemize{
\item \eqn{TPR = TP / (TP + FN)}
\item \eqn{Precision = TP / (TP + FP)}
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{   Details about the leave-one-out analysis can be found on the \pkg{rubias} GitHub page: \url{https://github.com/eriqande/rubias} or
   in Moran and Anderson (2018).
    
   citation: Moran, B.M. and E.C. Anderson. 2019. Bayesian inference from the conditional genetic stock 
             identification model. Canadian Journal of Fisheries and Aquatic Sciences, 76(4): 551–560. 
}\if{html}{\out{</div>}}
}
}
\examples{
group_names <- GCLr::ex_baseline$repunit \%>\% unique()

self_assign <- rubias::self_assign(reference = GCLr::ex_baseline, gen_start_col = 5)

input <- GCLr::loo_roc_rate_calc(data = self_assign, thres_levels = seq(0.01, .99, by = 0.005), group_names = group_names) # Note: make sure to include 0.50 in your threshold levels for 

plot_loo_tpr_prec(input, file = "~/example_plot.html") #Saves example plot to your documents folder
     
}
