% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rubias_IA_zscores.R
\name{plot_rubias_IA_zscores}
\alias{plot_rubias_IA_zscores}
\title{Plot distribution of \code{rubias} individual assignment z-scores.}
\usage{
plot_rubias_IA_zscores(path, mixnames, bins = 30, out_file = NULL)
}
\arguments{
\item{path}{Character vector of where to find output from each mixture as a .csv (created by \code{\link[=run_rubias_mix]{run_rubias_mix()}}).}

\item{mixnames}{A character vector of mixture names to include in the IA summary.}

\item{bins}{number of histogram bins (default = 30)}

\item{out_file}{Character vector of the file path to save the histograms as a pdf file including pdf extension. If NULL (default), no pdf will be produced.}
}
\value{
a histogram of z-scores for each mixture in \code{mixnames}
}
\description{
This function reads in a \code{rubias} individual posteriors csv file produced by \code{\link[=run_rubias_mix]{run_rubias_mix()}} and produces a histogram of the individual assignment (IA) z-scores. Significant outliers indicate mixture individuals not from any of the baseline populations (i.e., wrong species or missing baseline population).
}
\examples{
\dontrun{
path <- "V:/Analysis/2_Central/Sockeye/Cook Inlet/2012 Baseline/Mixture/UCI_sockeye_2020_postseason/rubias/output"
mixnames <- c("DriftDW_20", "DriftCorr_20", "UpperSub_20", "Kasilof600ft_20", "WestKalgin_20", "Eastern_20", "General_North_20", "General_South_20")
out_file <- "~/my_indiv_assign_zscore_plots.pdf"

GCLr::plot_rubias_IA_zscores(path = path, mixnames = mixnames, out_file = out_file)
}
}
