% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_collections.r
\name{pool_collections}
\alias{pool_collections}
\title{Combine "*.gcl" Objects}
\usage{
pool_collections(
  collections,
  loci = LocusControl$locusnames,
  IDs = NULL,
  newname = paste(collections, collapse = "."),
  LocusCtl = LocusControl
)
}
\arguments{
\item{collections}{A character vector of silly codes without the ".gcl" extension.}

\item{loci}{A character vector of locus names.}

\item{IDs}{A named list of FK_FISH_ID
These will be used to subset each collection before pooling. If no IDs are supplied, all individuals from each collection are used.}

\item{newname}{The name of the new "*.gcl" created. Do not provide the ".gcl" extension. If no name is supplied, then the newname defaults to
the collection names collapsed with a period between each name (e.g., "KQUART06.KQUART08.KQUART09").}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
assigns a new "pooled collection" to your global environment
}
\description{
This function combines "*.gcl" objects into a new one called "newname.gcl".
}
\details{
This function is useful for pooling collections from the same location into a single population ("pooled collection") and for producing "pooled mixture" objects for mixed stock analysis.
}
\examples{
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline, unpool = TRUE)

loci <- GCLr::ex_LocusControl$locusnames[-c(10, 12, 13, 32, 33)]

GCLr::pool_collections(collections = sillyvec[1:2], loci = loci, IDs = NULL, newname =  paste(sillyvec[1:2], collapse = "."), LocusCtl = GCLr::ex_LocusControl)

}
