% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prior_diff_allele_freq.r
\name{post_prior_diff_allele_freq}
\alias{post_prior_diff_allele_freq}
\title{Get Difference in Allele Frequencies Between Baseline and BAYES Posteriors}
\usage{
post_prior_diff_allele_freq(
  sillyvec,
  group_names,
  groupvec,
  loci,
  dir,
  nchains,
  mixname,
  burn = 0.5,
  ncores = 4
)
}
\arguments{
\item{sillyvec}{a character vector of population sillys used to create the baseline file.}

\item{group_names}{a character vector of reporting group names, where \code{length(group_names) == max(`groupvec`)}}

\item{groupvec}{a numeric vector indicating the reporting group affiliation of each population in \code{sillyvec}, where \code{length(groupvec) == length(sillyvec)}.}

\item{loci}{Character vector of the loci used to produce the baseline and mixture files.}

\item{dir}{the director where the BAYES mixture output files are located.  e.g., "BAYES/output" This directory should contain a folder named \code{mixname}, where the .FRQ files are saved.}

\item{nchains}{the number of MCMC chains analyzed for the mixture}

\item{mixname}{the name of the mixture. There should be a folder in \code{dir} named the same as the mixture}

\item{burn}{The proportion of iterations to drop from the beginning of each chain.
For example, for 40,000 iterations setting burn = 0.5 will drop the first 20,000 iterations.}

\item{ncores}{the number of cores to use when \code{\link[=calc_freq_pop]{calc_freq_pop()}} is called on (default = 4)}
}
\value{
a list containing 3 elements:
\itemize{
\item \code{MeanAbsDiffByPopLocus}: a tibble with 4 variables:
\describe{
\item{\code{chain}}{character; the chain number (e.g. Chain1, Chain2, Chain3, etc.)}
\item{\code{pop}}{factor; the population silly}
\item{\code{locus}}{factor; the locus name}
\item{\code{mean_abs_diff}}{double; the mean absolute difference between the original baseline frequency and the MCMC allele frequencies at each iteration}
}
\item \code{MeanAbsDiffByPop}: a tibble with 4 variables:
\describe{
\item{\code{chain}}{character; the chain number (e.g. Chain1, Chain2, Chain3, etc.)}
\item{\code{pop}}{factor; the population silly}
\item{\code{mean_abs_diff}}{double; the mean absolute difference between the original population baseline frequency and the group MCMC allele frequencies at each iteration over all loci}
\item{\code{outlier}}{logical; indicates whether the \code{mean_abs_diff} is an outlier}
}
\item \code{MeanAbsDiffByGroup}: a tibble with 4 variables:
\describe{
\item{\code{chain}}{character; the chain number (e.g. Chain1, Chain2, Chain3, etc.)}
\item{\code{group}}{character; the group name}
\item{\code{mean_abs_diff}}{double; the mean absolute difference between the original group baseline frequency and the group MCMC allele frequencies at each iteration over all loci}
\item{\code{outlier}}{logical; indicates whether the \code{mean_abs_diff} is an outlier}
}
}
}
\description{
This function reads in BAYES MCMC samples of baseline frequencies (.FRQ output files) for each population (one file per chain) and compares them to the original baseline allele frequencies (see details).
}
\details{
During a mixed stock analysis in BAYES, the original allele frequencies get updated at each MCMC iteration.
Looking at the difference between the average MCMC allele frequency and the original baseline frequencies can be useful when looking into reasons for non-convergence among chains (i.e., Gelman-Ruban shrink factor > 1.2)
Populations/reporting groups with large changes in allele frequencies for certain population/reporting groups can indicate that the mixture contains fish from a distinct population that is not represented in the baseline.
}
\note{
To run this function, you must have the baseline silly (.gcl objects) and LocusControl loaded in your current workspace and have BAYES frequency files for each chain of the mixture you are interested in.
}
\examples{
\dontrun{
dir <- "V:/Analysis/2_Central/Sockeye/Cook Inlet/Missing Baseline Analysis/BAYES/output"
load("V:/Analysis/2_Central/Sockeye/Cook Inlet/Missing Baseline Analysis/Missing Baseline Analysis.RData")
post_prior_diff_allele_freq(sillyvec = PopNames69, group_names = Groups, groupvec = groupvec69, loci = loci, dir = dir, nchains = 7, mixname = "Western06early")

}

}
\seealso{
For additional details, run the following code in your console to bring up the BAYES manual:
\code{utils::browseURL(system.file("BAYES", "MANUAL.DOC", package = "GCLr"))}
}
