% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_geno_stats.R
\name{proj_geno_stats}
\alias{proj_geno_stats}
\title{Calculate Project Genotype Statistics and Write to Excel Workbook}
\usage{
proj_geno_stats(report, file)
}
\arguments{
\item{report}{a report object created by \code{GCLr::get_geno()}}

\item{file}{the name of the excel workbook.}
}
\value{
This function writes an Excel workbook with three sheets:
\itemize{
\item \code{proj_Genotypes_Table}: a formatted genotypes table
\item \code{Calculations}: number of fish with success rates < 80\%, total number of project fish, and Proportion of fish with success rates greater than 80\%
\item \code{Avg success by plate}: a table with PlateID and Average Success Rate
}
}
\description{
This function takes the genotypes report object from \code{GCLr::get_geno()}, calculates genotypes statistics, and write them to a formatted Excel workbook.
}
\details{
This function is called on by \code{GCLr::get_geno()} when a \code{project_name} is supplied and \code{proj.stats = TRUE}.  However, this function can also be run separately if you have an genotypes object created by \code{GCLr::get_geno}.
}
\examples{
\dontrun{
Geno_proj <- GCLr::get_geno(project_name = "K205", file = NULL, username = "awbarclay", password = "password", open.file = TRUE) 

GCLr::proj_geno_stats(report = Geno_proj, file = path.expand("~/TestGenotypesReport_proj.xlsx"))
}   
  
}
\seealso{
\code{\link[=get_geno]{get_geno()}}
}
