% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.r
\name{qc}
\alias{qc}
\title{QC script}
\usage{
qc(
  dirqc,
  species,
  project,
  project_type,
  username,
  password,
  qcSummaryfile = paste("Project", project, "qc Summary R Script.xlsx"),
  conflict_rate = 0.1
)
}
\arguments{
\item{dirqc}{The directory where the QC files are located.}

\item{species}{The species for the project (e.g., "pink").}

\item{project}{The project name (e.g., "P014").}

\item{project_type}{The type of project ("Biomark", "uSat", or "GT-seq").}

\item{username}{The username for LOKI access.}

\item{password}{The password for LOKI access.}

\item{qcSummaryfile}{The file path for the QC summary file; This is automatically generated using \code{project} (don't change unless you know what you're doing).}

\item{conflict_rate}{The conflict rate threshold at which dupcheck between sillys occurs (default is 0.10).}
}
\value{
Returns summary tables and saves them in the QC summary file.
\itemize{
\item \code{text}:
\itemize{
\item DupCheck Results (if applicable)
}

\item \code{figures}:
\itemize{
\item Levelplot of genotyping success rate by silly and locus for ProjectSillys
\item Levelplot of genotyping success rate by silly and locus for qcSillys
\item Histogram of overall qc individual conflict rate
\item Individual histograms of duplicate rate for conflict individuals
}

\item \code{Excel}:
\itemize{
\item Summary by Silly
\item Conflicts by Silly
\item DupCheck Results (if applicable)
\item Conflicts by Locus
\item Conflicts by PlateID
\item Failure Rate by Silly
\item Failure Rate by Locus
\item Failure Rate by PlateID
\item Overall Failure Rate
\item Original Project Sample Size by Locus
\item Duplicates within Silly
}
}
}
\description{
Perform quality control for lab projects.
}
\note{
It is important to run this script in order, or some functions will not provide accurate results. This is by design, as this script only hits LOKI once to save time.
This is not a function, but a step wise operation that the user must run through to manually generate applicable files.
}
\examples{
\dontrun{
# Example usage
dirqc <- "V:/Lab/Genotyping/SNP Projects/Pink/Project P014 AHRP Parentage GTSeq Part 2/qc/"
species <- "pink"
project <- "P014"
project_type <- "Biomark"
username <- ""
password <- ""
qcSummaryfile <- paste("Project", project,"qc Summary R Script.xlsx")
conflict_rate <- 0.10
P014_qc(dirqc, species, project, project_type, username, password, qcSummaryfile, conflict_rate)
}

}
