% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_bounce.R
\name{qc_bounce}
\alias{qc_bounce}
\title{Get Quality Control Loki Bounce Summary and Write to Excel Workbook}
\usage{
qc_bounce(QC_directory, project, username, password)
}
\arguments{
\item{QC_directory}{the project QC directory}

\item{project}{the project name; must be spelled exactly the way it is in iStrategy.}

\item{username}{Your user name for accessing LOKI.}

\item{password}{Your password for accessing LOKI.}
}
\value{
This function writes an Excel workbook with five sheets:
\itemize{
\item \code{proj_Genotypes_Table}: a formatted genotypes table
\item \code{Calculations}: number of fish with success rates < 80\%, total number of project fish, and Proportion of fish with success rates greater than 80\%
\item \code{Avg success by plate}: a table with \emph{PlateID} and \emph{Average Success Rate} columns
\item \code{Conflicts by individual}: a table with \emph{SillySource} (aka fish ID), \emph{PlateID}, and \emph{success_rate} columns and associated columns for each type of conflict in the concordance file(s) (e.g., Homo-Het, Het-Homo, Homo-Homo), and a column of \emph{Total_conflicts}
\item \code{Conflicts by locus}: a table with \emph{locus}, \emph{PlateID}, and \emph{success_rate} columns and associated columns for each type of conflict in the concordance file(s) (e.g., Homo-Het, Het-Homo, Homo-Homo), and a column of \emph{Total_conflicts}
}

#' @examples
\dontrun{

QC_directory <- "V:/Lab/Genotyping/SNP Projects/Sockeye/Project S266 Port Moller Inseason 2024/QC"
project <- "S266"
username <- "awbarclay"
password <- scan("~/R/usr.pw", what = "")[2]

qc_bounce(QC_directory, project, username, password)

 }
}
\description{
This function produces an Excel file containing the project genotypes, the proportion of project fish with at least 80\% of loci with scores, average success rate by plate, conflicts and success rates for individual samples, and conflicts and success rates by locus
}
\details{
This function requires R package "RJDBC" for connecting to LOKI and calls on \code{GCLr::loki2r_proj()}, \code{GCLr::get_geno()},
and \code{GCLr::combine_conflicts()} to get the information it needs to create a QC bounce Excel Workbook. In order for this function to work,
there must be a folder named 'Conflict Reports' in the supplied \code{QC_directory} that contains Loki Concordance report(s) produced by the genotypes importer.
All concordance reports will be read in by the function as long as they include "Concordance" in their file names, so make sure that the folder only includes the reports you want summarized.
}
\seealso{
\code{\link[=loki2r_proj]{loki2r_proj()}}

\code{\link[=get_geno]{get_geno()}}

\code{\link[=combine_conflicts]{combine_conflicts()}}
}
