% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biomark_qc.r
\name{read_biomark_qc}
\alias{read_biomark_qc}
\title{Read Biomark QC Data}
\usage{
read_biomark_qc(qccsvFilepaths, skip = 15)
}
\arguments{
\item{qccsvFilepaths}{Character vector specifying the file paths of the QC CSV files.}

\item{skip}{Number of lines to skip while reading the CSV files. Default is 15.}
}
\value{
Returns a few silly objects to the global environment
\itemize{
\item \code{qc.gcl} list objects
\item \code{qcSillys}; a character vector of qc sillys
}
}
\description{
This function reads QC data from Biomark CSV files. It's called on by \code{\link[=qc]{qc()}}.
}
