% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genepop_pwfst.r
\name{read_genepop_pwfst}
\alias{read_genepop_pwfst}
\title{Read Pairwise Fst GENEPOP Output}
\usage{
read_genepop_pwfst(file, popnames, outfile = NULL)
}
\arguments{
\item{file}{the path to the GENEPOP .MIG file}

\item{popnames}{a character vector of population names corresponding to the populations in the GENEPOP file used to create the "*.MIG" file}

\item{outfile}{the path including file name to save a tab delimited text file of the pairwise matrix}
}
\value{
This function returns a named pairwise Fst matrix. If \code{!is.null(outfile)} the function produces a tab delimited text file of the pairwise matrix, otherwise no file is produced.
}
\description{
This function reads in the pairwise FST half matrix from a GENEPOP "*.MIG" file and produces a full matrix with dimnames. The matrix from this function can be used to create a pairwise FST tree using \code{\link[ape:nj]{ape::nj()}}
}
\examples{
GCLr::read_genepop_pwfst(file = system.file("genepop", "ex_genepop.txt.MIG", package = "GCLr"), popnames = paste0("Pop", seq(14)))

}
