% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proj_geno.r
\name{read_proj_geno}
\alias{read_proj_geno}
\title{Read Project Genotypes}
\usage{
read_proj_geno(
  project_name = NULL,
  sillyvec = NULL,
  loci = NULL,
  username,
  password
)
}
\arguments{
\item{project_name}{A character vector of one or more project names as spelled in LOKI.}

\item{sillyvec}{A character vector of SILLYs.}

\item{loci}{A character vector of locus names as they are spelled in LOKI.}

\item{username}{Your username for accessing LOKI through R.}

\item{password}{Your password for accessing LOKI through R.}
}
\value{
Returns \code{ProjectSillys}, a character vector of all sillys in the project (sillyvec).

Returns "slim" .gcl objects for each silly (slim = not all attributes table).

Returns \code{LocusControl} for all loci used in the project (including "loci", "nallales", "ploidy", "alleles").
}
\description{
This function is intended for use in the qc.R script. It pulls project genotypes to create "slim" .gcl objects for each silly and the LocusControl for all loci used in the project.
Warning: Genotypes (pre-October 2016) may not exist in the LOKI lookup table with a project name, so genotypes will have to be pulled by sillyvec and loci.
}
\details{
This function pulls project genotypes from LOKI database to create "slim" .gcl objects for each silly, along with the LocusControl for all loci used in the project.
The function checks the combination of arguments to ensure the correct usage.
}
\examples{
read_proj_geno(project_name = c("P014", "P015", "P016"), sillyvec = c("SCIMA18", "SCIMA17"), 
               loci = c("One_E2", "One_MHC2_251", "One_Cytb_17"), username = "awbarclay", password = "password")

}
