% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_usat_qc.r
\name{read_usat_qc}
\alias{read_usat_qc}
\title{Read uSat QC Genotypes}
\usage{
read_usat_qc(qccsvFilepaths)
}
\arguments{
\item{qccsvFilepaths}{A character vector with relative paths for QC genotype CSV files.}
}
\value{
Returns a few silly objects to the global environment
\code{qc.gcl} list objects
\code{qcSillys}; a character vector of qc sillys
}
\description{
This function reads in uSat QC genotypes from .csv files and creates .gcl objects in the global environment. It verifies that the QC sillys and loci are found in the project and assigns the \code{qcSillys} vector to the global environment. It's called on by \code{\link[=qc]{qc()}}.
}
\details{
The function reads in the QC genotype data from the CSV files, renames and formats columns, and creates scores and counts arrays for each silly. The resulting .gcl objects are assigned to the global environment with the names "sillyqc.gcl" (e.g., "silly1qc.gcl", "silly2qc.gcl").
}
