% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_alt_species.r
\name{remove_alt_species}
\alias{remove_alt_species}
\title{Removes wrong species}
\usage{
remove_alt_species(
  AlternateSpeciesReport,
  AlternateCutOff = 0.5,
  FailedCutOff = 0.5,
  NonmissingCutOff
)
}
\arguments{
\item{AlternateSpeciesReport}{The object created by \code{\link[=find_alt_species]{find_alt_species()}}}

\item{AlternateCutOff}{The minimum proportion of alternate alleles an individual must have in order to be considered the 'wrong species'. Default is 0.5}

\item{FailedCutOff}{The minimum proportion of failed loci an individual must have in order to be considered the 'wrong species'. Default is 0.5}

\item{NonmissingCutOff}{The minimum number of non-missing alternate loci an individual must have in order to be considered the 'wrong species'. Default is 3.}
}
\value{
Returns a tibble showing which fish were removed and their stats (alternate and failed markers)
}
\description{
This function removes fish that have been identified, via genetic markers, as the wrong species. It relies on output from \code{\link[=find_alt_species]{find_alt_species()}}.
}
\examples{
\dontrun{
 source(file = "Examples/qcExample.R")
}

}
