% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dups.r
\name{remove_dups}
\alias{remove_dups}
\title{Remove Duplicates based on Criteria}
\usage{
remove_dups(dupcheck, remove_both = FALSE)
}
\arguments{
\item{dupcheck}{A data frame or tibble containing the information to check for duplicates. This object is produced by dupcheck_within_silly().}

\item{remove_both}{Logical. If set to TRUE, the function will remove duplicates from both "ID1" and "ID2" columns. If set to FALSE (default), the function will remove duplicates based on the "Missing1" and "Missing2" columns.}
}
\value{
A data frame containing the results after removing duplicates.
}
\description{
This function removes duplicates from the input data frame based on specific criteria.
}
\details{
The \code{remove_dups} function takes an input data frame \code{dupcheck} and checks for duplicates based on the columns "silly", "ID1", "ID2", "Missing1", "Missing2", and "proportion". If there are no duplicates or if the input \code{dupcheck} is empty, the function will return an empty data frame with a warning message. The function will first check if \code{dupcheck} is a tibble; if not, it assumes the data is stored in a column named "report" and extracts the tibble from that column.

If the \code{remove_both} argument is set to TRUE, the function removes duplicates from both "ID1" and "ID2" columns using tidyr::pivot_longer. Otherwise, it calculates which duplicates to remove based on the "Missing1" and "Missing2" columns using dplyr::mutate and dplyr::case_when.

The function then processes the data for each unique value of "silly" and removes the corresponding duplicates using the \code{remove_ids} function.
}
\examples{
\dontrun{
password = "************"
create_locuscontrol(markersuite = "Sockeye2011_96SNPs", username = "awbarclay", password = password)
sillyvec = c("SMCDO03", "SNEVA13")
loki2r(sillyvec = sillyvec, username = "awbarclay", password = password)
remove_ind_miss_loci(sillyvec = sillyvec)

dupcheck <- dupcheck_within_silly(sillyvec = sillyvec, loci = LocusControl$locusnames, quantile = NULL, minproportion = 0.95, ncores = 8)
removed_dups <- remove_dups(dupcheck)
}

}
\seealso{
\code{\link{remove_ids}}
}
