% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na_indv.r
\name{remove_na_indv}
\alias{remove_na_indv}
\title{Remove IDs with NA Values for all Loci}
\usage{
remove_na_indv(sillyvec, LocusCtl = LocusControl)
}
\arguments{
\item{sillyvec}{a character vector of silly codes (e.g. \code{sillyvec <- c("KQUART06","KQUART08","KQUART09")})}

\item{LocusCtl}{An optional \code{LocusControl} object (default: \code{LocusControl}) created by \code{\link[=create_locuscontrol]{create_locuscontrol()}} that specifies the columns to be checked for NA values. The function will use the \code{locusnames} attribute of the LocusCtl object to identify the columns.}
}
\value{
A tibble with two columns: \code{SILLY_CODE} contains the element names (\code{SILLY_CODE}) from \code{sillyvec}, and \code{IDs_Removed} contains the identifiers (IDs) of the rows removed for each element in the list.
}
\description{
This function removes any individuals from a SILLY.gcl object that were not genotyped for all Loci in markersuite.
}
\details{
In Loki, NA genotypes mean that no genotyping was performed.
This function checks each silly in \code{sillyvec} for individuals with NAs across all loci and removes them.
This function should be run immediately after pulling genotypes using \code{\link[=loki2r]{loki2r()}}, \code{\link[=loki2r_gaps]{loki2r_gaps()}}, or \code{\link[=loki2r_proj]{loki2r_proj()}}
}
\examples{
\dontrun{

removedInd <- GCLr::remove_na_indv(sillyvec = sillyvec, LocusCtl = LocusControl)

}

}
