% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.r
\name{save_objects}
\alias{save_objects}
\title{Save R Objects to Disk}
\usage{
save_objects(objects, path, rds = FALSE)
}
\arguments{
\item{objects}{character vector of objects you wish to save}

\item{path}{character vector of where you want to save objects}

\item{rds}{logical; if set to \code{TRUE} the objects will be saved in rds format,
default is \code{FALSE} for backwards compatibility.}
}
\value{
All objects are saved as "object.txt" or "object.rds"
}
\description{
This function saves R objects with either \code{\link[base:dput]{base::dput()}} or \code{\link[base:saveRDS]{base::saveRDS()}}.
}
\details{
\code{\link[base:dput]{base::dput()}} saves object code to a text file.
\code{\link[base:saveRDS]{base::saveRDS()}} saves objects in rds format, which is a compressed format - saves and reads back into R much faster than a text file.
}
\examples{
LocusControl <- GCLr::ex_LocusControl
pairwisefst <- GCLr::ex_pairwise_fst

dir.create(path.expand("~/objects"))

GCLr::save_objects(objects = c("LocusControl", "pairwisefst"), path = path.expand("~/objects"), rds = TRUE)

GCLr::save_objects(objects = c("LocusControl", "pairwisefst"), path = path.expand("~/objects"), rds = FALSE)

}
\seealso{
\code{\link[base:dput]{base::dput()}}

\code{\link[base:saveRDS]{base::saveRDS()}}
}
