% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silly_n.GCL.R
\name{silly_n.GCL}
\alias{silly_n.GCL}
\title{Get sample size of .gcl objects}
\usage{
silly_n.GCL(sillyvec)
}
\arguments{
\item{sillyvec}{A character vector of silly codes with associated *.gcl objects that you want sample sizes for.}
}
\value{
This function returns a two-column tibble of sillys (silly) and sample sizes (n)
}
\description{
This function produces a tibble of .gcl object sample sizes
}
\examples{

load("V:/Analysis/2_Central/Chinook/Cook Inlet/2019/2019_UCI_Chinook_baseline_hap_data/2019_UCI_Chinook_baseline_hap_data.RData")

silly_n.GCL(sillyvec = sillyvec157)
}
