% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_combchip_loki_import.R
\name{split_combchip_loki_import}
\alias{split_combchip_loki_import}
\title{Split a Combined Chip Import File into Multiple Files}
\usage{
split_combchip_loki_import(files, nlines = 9216)
}
\arguments{
\item{files}{the file path to the Biomark files including .csv extension}

\item{nlines}{the maximum number of lines you want in each file (default = 9216)}
}
\description{
This function takes a combined Biomark import file and splits it up into
multiple files so genotypes can be imported into Loki in smaller batches (see details)
}
\details{
The function reads in a Biomark file and splits it up into multiple files with a maximum of number of lines = \code{nlines}
and writes them out in the same folder as the original file with letters at the end of the file names. For example,
if you have a file named "S254 03_04 Combined Chip Run.csv" with 18,432 (192 samps X 24 loci X 4 chips) lines (not including the header lines) and you want to break it up into files with 9,216 lines each (96 samps X 96 loci), the function will produce
2 files, and name them "S254 03_04 Combined Chip Run_a.csv" and "S254 03_04 Combined Chip Run_b.csv"
}
\examples{
\dontrun{

files <- list.files(path = "C:/Users/awbarclay/Documents/biomark split tests", pattern = "*Combined Chip Run*", full.names = TRUE)

GCLr::split_combchip_loki_import(files = files, nlines = 9216)

}

}
