% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_estimator_bayes.r
\name{stratified_estimator_bayes}
\alias{stratified_estimator_bayes}
\title{Stratified Estimator using Bayesian output}
\usage{
stratified_estimator_bayes(
  groupvec,
  groupnames,
  maindir,
  mixvec,
  catchvec,
  CVvec = rep(0, length(catchvec)),
  newname,
  priorname = "",
  ext = "RGN",
  nchains = 5,
  burn = 0.5,
  alpha = 0.1,
  threshold = 5e-07,
  xlsx = FALSE,
  PosteriorOutput = FALSE
)
}
\arguments{
\item{groupvec}{If \code{ext = "RGN"} \code{groupvec = seq(groupnames)}
If \verb{ext = "bot} a vector the same length as the number of populations, each element is the group number for the corresponding population.
e.g. groupvec = c(1,1,1,2,2,3) will combine the first 3 pops into group 1, the 4th & 5th pops into group 2 and the 6th pop into group 3.
The elements of groupvec do not need to be in order -- e.g. groupvec = c(1,3,1,2,2,3) is acceptable.}

\item{groupnames}{Character string vector of names for the groups.}

\item{maindir}{The pathname to the sub-directories that contain the BAYES output files. This is also where the Excel output file will be written to.}

\item{mixvec}{Character string vector of mixture names, one for each stratum.}

\item{catchvec}{Vector of harvest numbers. Must have the same length as mixvec. Each element is the harvest for its corresponding stratum (mixture).}

\item{newname}{Chosen name of the output file without extension.}

\item{priorname}{Character string giving the name (type) of the prior used.}

\item{ext}{Used to indicate if "RGN" (group estimates) or "BOT" (population estimates) files are to be read in. Defaults to "RGN".}

\item{nchains}{The number of chains to read in from each mixture. Defaults to 3.}

\item{burn}{Proportion of chain length to use as burn-in. Defaults to 0.5.}

\item{alpha}{A proportion. Used to create a 100*(1-alpha)\% confidence interval. Defaults to 0.1 (90\% CI).}

\item{xlsx}{Logical; if TRUE, the output will be written to an Excel file.}

\item{PosteriorOutput}{Logical; if TRUE, the function will return both the summary statistics and the posterior output. If FALSE, it will return only the summary statistics.}
}
\value{
A summary of the stratified estimator, including mean, standard deviation, median, 90\% confidence interval, proportion of values less than the threshold, and Gelman-Rubin statistic.
}
\description{
This function reads in BAYES output files (all chains) from multiple temporal strata (mixtures) and, by weighting by harvest from each stratum, produces a seasonal total harvest
estimate and C.I. for all defined regions (groups). Appologies for the number of arguments--this is necessary to give the required flexibility. The last four arguments will rarely need to be changed.
}
\details{
WARNING -- A certain directory structure is required for successful execution. Particularly, each mixture's BAYES output files must be within their own sub-directory with the exact name as the mixture.
Also, the BAYES output files need names that begin with the mixture name, followed by the the prior name, followed by Chain -- e.g. "MixnamePriorChain1RGN.RGN" or "MixnamePriorChain1BOT.BOT".
The exact same priorname name must be used in all BAYES output file names.

The output is an Excel file summarizing total harvest.
}
\examples{
\dontrun{
groupvec <- 1:4

groupnames <- c("KenaiTrib","KenaiMain","KasilofMain","CIOther")

maindir <- "V:/Analysis/2_Central/Chinook/Cook Inlet/2014/Mixture/ESSN_Chinook_2020_postseason/bayes/output"

mixvec <- c("Kasilof_early","Kasilof_late","Kenai_Late")

catchvec <- c(356, 301, 195)

newname <-  "TotalHarvest"

GCLr::stratified_estimator_bayes(groupvec = groupvec, groupnames = groupnames, maindir = maindir, mixvec = mixvec, catchvec = catchvec, newname = newname, xlsx = TRUE)

}

}
