% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_estimator_rubias.r
\name{stratified_estimator_rubias}
\alias{stratified_estimator_rubias}
\title{Summary statistics from a stratified estimate of rubias output}
\usage{
stratified_estimator_rubias(
  rubias_output = NULL,
  mixvec = NULL,
  naive = FALSE,
  catchvec = NULL,
  cv = NULL,
  newname = NULL,
  group_names = NULL,
  group_names_new = NULL,
  groupvec = NULL,
  groupvec_new = NULL,
  path = "rubias/output",
  alpha = 0.1,
  burn_in = 5000,
  bias_corr = FALSE,
  threshold = 5e-07,
  ncores = 4
)
}
\arguments{
\item{rubias_output}{Output list object from \code{run_rubias_mix()} or \code{\link[rubias:infer_mixture]{rubias::infer_mixture()}}.}

\item{mixvec}{Character vector of mixture sillies, used to read in output .csv files if \code{rubias_output = NULL}.}

\item{naive}{Default is FALSE for using stock-specific total catch output to summarize harvest and stock proportions. Set it as TRUE if you want the summary done by the old way (stock-specific harvets = harvest * stock proportion), or you are using fishing effort instead of catch number.}

\item{catchvec}{Numeric vector of harvest for each strata, must be in the same order as \code{mixvec}.}

\item{cv}{Numeric vector of harvest estimate coefficients of variation for each stratum, must be the same order as \code{mixvec}. (Only for catch numbers with CVs).}

\item{newname}{Character vector of length 1 specifying the name of the stratified estimate.}

\item{group_names}{Character vector of group_names, used to sort repunit as a factor, can get from .csv.}

\item{group_names_new}{Character vector of new group_names, used to roll up groups from fine-scale to broad-scale for bias correction.}

\item{groupvec}{Numeric vector indicating the group affiliation of each pop in sillyvec, used if resuming to new groups.}

\item{groupvec_new}{A numeric vector indicating the new group affiliation of each group, used if resuming fine-scale groups to broad-scale groups with bias correction.}

\item{path}{Character vector of where to find output from each mixture as a .csv (created by \code{run_rubias_mix()}).}

\item{alpha}{Numeric constant specifying credibility intervals, default is 0.1, which gives 90\% CIs (i.e., 5\% and 95\%).}

\item{burn_in}{Numeric constant specifying how many sweeps were used for burn_in in \code{run_rubias_mix()} or \code{\link[rubias:infer_mixture]{rubias::infer_mixture()}}.}

\item{bias_corr}{Logical switch indicating whether you want bias-corrected values from \code{method = "PB"} or not,
currently can NOT do bias correction if not using the same repunits that were run in the mixture.}

\item{threshold}{Numeric constant specifying how low stock comp is before assuming 0, used for \code{P=0} calculation, default is from BAYES.}

\item{ncores}{The number of CPU cores utilized in this operation.}
}
\value{
Returns a tibble with 13 fields (18 if with CVs) for each repunit (reporting group).
\itemize{
\item stratified_mixture: Character of stratified mixture.
\item repunit: Factor of reporting groups (only a factor for ordering, plotting purposes).
\item mean_harv: Mean stock catch number.
\item sd_harv: Standard deviation for catch.
\item median_harv: Median stock catch number.
\item \verb{5\%_harv} (depends on alpha level): Lower bound of credibility interval for catch number.
\item \verb{95\%_harv}: Upper bound of credibility interval for catch number.
\item mean: Mean stock composition.
\item sd: Standard deviation.
\item median: Median stock composition.
\item \verb{5\%}: Lower bound of credibility interval.
\item \verb{95\%}: Upper bound of credibility interval.
\item P=0: The proportion of the stock comp distribution that was below threshold (i.e., posterior probability that stock comp = 0).
(For catch numbers with CVs):
\item mean_backcalc: mean_harv/sum(mean_harv); mean stock composition.
\item sd_backcalc: sd_harv/sum(mean_harv); standard deviation.
\item median_backcalc: median_harv/sum(mean_harv); median stock composition.
\item \verb{5\%_backcalc}: \verb{5\%_harv}/sum(mean_harv); lower bound of credibility interval.
\item \verb{95\%_backcalc}: \verb{95\%_harv}/sum(mean_harv); upper bound of credibility interval.
}
}
\description{
This function computes summary statistics from a stratified estimate of \pkg{rubias} output,
as a column, instead of a single matrix. It can take either the \code{rubias_output} list object
from \code{run_rubias_mix()} or \code{\link[rubias:infer_mixture]{rubias::infer_mixture()}}, OR it can read in the .csv files created by \code{run_rubias_mix()}.
}
\details{
NOTE: Currently this function only allows bias correction for the reporting groups run in the mixture.
It cannot do bias correction for different baseline groupvecs, because current \pkg{rubias} output only
gives the bias-corrected means for each \code{mixture_collection} and \code{repunit} (i.e., \code{rho}, not \code{pi}).

UPDATE:
\itemize{
\item This function CAN do bias correction if you are rolling up groups from fine-scale to broad-scale. To use this functionality, specify \code{group_names} as the original, fine-scale groups, \code{groupvec_new} as the groupvec to go from fine-scale to broad-scale groups, and \code{group_names_new} as the broad-scale groups (i.e., length(groupvec_new) = length(group_names), and max(groupvec_new) == length(group_names_new)).
\item 3/18/2022 corrected some errors in calculations (only when catch numbers had cv's). New output format now includes both catch numbers and stock props columns with '_backcalc' are the stock props back-calculated using simulated harvest stock props are cut off at 1.
\item Multi-chain compatible
}
}
\examples{
path <- "V:/Analysis/1_SEAK/Sockeye/Mixture/D101to103 Seine/output_PB"

GCLr::stratified_estimator_rubias(path = path,
                            mixvec = c("D101__28293031", "D101__323334", "D101__3536"),
                            group_names = c("Alaska", "Nass", "Skeena", "Other", "McDonald", "Hugh Smith", "Klawock"),
                            catchvec = c(34860, 38169, 21562),
                            newname = "D101_21_stratified_NBDomestic",
                            group_names_new = c("Alaska", "Nass", "Skeena", "Other", "McDonald", "Hugh Smith"),
                            groupvec_new = c(1, 2, 3, 4, 5, 6, 1),
                            alpha = 0.1,
                            burn_in = 5000,
                            bias_corr = TRUE,
                            cv = c(0.2, 0.5, 0.4))

}
