% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_LD.r
\name{summarize_LD}
\alias{summarize_LD}
\title{Summarize LD Results}
\usage{
summarize_LD(LDresults, alpha = 0.05, prop_sign_pops = 0.5)
}
\arguments{
\item{LDresults}{The output from \code{\link[=test_LD]{test_LD()}}; a tibble with 3 variables:
\itemize{
\item \code{Locus_pair}
\item \code{Pop}
\item \code{Pvalue}
}}

\item{alpha}{A numeric value indicating the significance level for filtering the \code{\link[=test_LD]{test_LD()}} output; default is 0.05.}

\item{prop_sign_pops}{A numeric value indicating the proportion of populations with significant test results; default is 0.5.
Only locus pairs that have > prop_sign_pops will be plotted.}
}
\value{
An interactive bar plot with the proportion of populations with significant tests on the y-axis and the locus pairs on the x-axis.
}
\description{
This function takes the output from \code{\link[=test_LD]{test_LD()}}, filters tests by the supplied significance level (alpha),
and makes a bar plot of the proportion of pops with significant test results for each locus pair.
}
\examples{
\dontrun{

sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

loci <- GCLr::ex_LocusControl$locusnames[-c(10, 12, 13, 32, 33, 97, 98)]

dir.create(path = "~/GENEPOP")

GCLr::gcl2genepop(sillyvec = sillyvec, loci = loci, path = path.expand("~/GENEPOP"), VialNums = TRUE, usat = FALSE, 
                  ncores = parallel::detectCores(), npops = 1, LocusCtl = GCLr::ex_LocusControl) 

LDresults <- GCLr::test_LD(genepopFiles = list.files(path.expand("~/GENEPOP")), path = path.expand("~/GENEPOP"), batches = 14, iterations = 1, ncores = parallel::detectCores())

GCLr::summarize_LD(LDresults, alpha = 0.05, prop_sign_pops = 0.5)

}

}
\seealso{
\code{\link[=gcl2genepop]{gcl2genepop()}}

\code{\link[=test_LD]{test_LD()}}
}
