% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rubias_individual_assign.r
\name{summarize_rubias_individual_assign}
\alias{summarize_rubias_individual_assign}
\title{Summarize \code{rubias} individual assignments.}
\usage{
summarize_rubias_individual_assign(
  rubias_output = NULL,
  mixnames = NULL,
  path = "rubias/output"
)
}
\arguments{
\item{rubias_output}{Output list object from \code{\link[=run_rubias_mix]{run_rubias_mix()}} or \code{\link[rubias:infer_mixture]{rubias::infer_mixture()}}.}

\item{mixnames}{A character vector of mixture names to include in the individual assignment summary.}

\item{path}{Character vector of where to find output from each mixture as a .csv (created by \code{\link[=run_rubias_mix]{run_rubias_mix()}}).}
}
\value{
A tibble with the following variables:
\itemize{
\item mixture_collection: The name of the mixture.
\item indiv: The individual (aka SillySource).
\item And a variable for each group in group_names containing the posterior means of group membership for each individual.
}
}
\description{
This function will summarize \code{rubias} individual assignments.
}
\examples{
path <- "V:/Analysis/2_Central/Sockeye/Cook Inlet/2012 Baseline/Mixture/UCI_sockeye_2020_postseason/rubias/output"
mixnames <- c("DriftDW_20", "DriftCorr_20", "UpperSub_20", "Kasilof600ft_20", "WestKalgin_20", "Eastern_20", "General_North_20", "General_South_20")
GCLr::summarize_rubias_individual_assign(rubias_output = NULL, mixnames = mixnames, path = path)

}
