% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_collections.r
\name{temporal_collections}
\alias{temporal_collections}
\title{Get Temporal Collections for Hierarchical ANOVA}
\usage{
temporal_collections(sillyvec, region = NULL, min.samps = 50, sep = ".")
}
\arguments{
\item{sillyvec}{a vector of pooled silly codes (a.k.a. pooled pops) without the ".gcl" extension}

\item{region}{an optional tibble containing two numeric variables:
\itemize{
\item \code{pop_no}: the position number for each population in sillyvec
\item \code{region}: the region number for each population in sillyvec
}}

\item{min.samps}{the minimum number of samples for including a collection}

\item{sep}{the separator used when pooling collections into populations; default is "."}
}
\value{
a tibble with the following variables:
\itemize{
\item \code{silly}: non-pooled silly codes,
\item \code{spop}: numbers indicating the population affiliation for each collection,
\item \code{sregion}: numbers indicating the regional affiliation for each collection
}
}
\description{
This function retrieves temporal collections from a vector of pooled sillys for hierarchical ANOVA.
}
\note{
Single-silly pops and sillys with sample sizes < "min.samps" will not be included in the output.
}
\examples{
load("V:/Analysis/2_Central/Coho/Cook Inlet/2019/2019_Cook_Inlet_coho_baseline/2019_Cook_Inlet_coho_baseline_new.Rdata")
source(paste0(path.expand("~/R/"), "Functions.R"))#GCL functions
sillyvec <- Final_Pops$collection
region <- Final_Pops \%>\% select(pop = order, region = region)
temporal_collections(sillyvec = sillyvec, region = region, min.samps = 50, sep = ".")

}
