% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_comp.r
\name{var_comp}
\alias{var_comp}
\title{Variance Components Calculation}
\usage{
var_comp(sillyvec, loci, groupvec, fstatfile = NULL, dir)
}
\arguments{
\item{sillyvec}{A character vector specifying the population silly codes for analysis.}

\item{loci}{A character vector of locus names to include in the analysis. Can include haploid markers and/or combined loci.}

\item{groupvec}{A numeric vector specifying the reporting group affiliation for each silly code.}

\item{fstatfile}{Optional character string specifying the path to an existing FSTAT file (".dat"). If provided, the function verifies compatibility with the input data. Default is NULL.}

\item{dir}{A character string specifying the directory where an FSTAT file (".dat") will be created. This argument is used only when \code{fstatfile} is NULL.}
}
\value{
A matrix containing the four variance components for each locus:
\itemize{
\item P: For populations (reporting groups in this case).
\item S: For subpopulations within populations (populations within reporting groups).
\item I: For individuals within subpopulations (individuals within populations).
\item G: For alleles within individuals (akin to heterozygosity).
}
Component definitions in () assume that you are running this function on silly codes that are populations. This function could also be run on silly codes that are collections, but the definitions in () would be different.
}
\description{
This function calculates the variance components of a three-level hierarchy. This function is intended to help with marker selection (i.e., going from a 96 to 24 SNP baseline).
}
\details{
The original function is "BackwardEliminationThetaP.GCL" in the temp folder.
}
\examples{
GCLr::var_comp(sillyvec = Kodiak49Pops, loci = loci90, groupvec = Kodiak49GroupVec9, dir = "V:/WORK/Sockeye/Kodiak/Kodiak Afognak Baseline Look 2014/Kyle/ThetaP")

}
\references{
Weir, B.S. 1995. Genetic Data Analysis. Sinauer Associates, Inc. Sunderland, MA. (pg 184)
}
