% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tree_color.r
\name{add_tree_color}
\alias{add_tree_color}
\title{Add Tree Color}
\usage{
add_tree_color(
  tree,
  currentnames,
  treenames,
  groupvec,
  regioncol,
  regionpch = NULL,
  write_nexus = FALSE,
  file = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree object (class = "phylo") produced by the \pkg{ape} package.}

\item{currentnames}{A character vector of the tip labels (pop names) in the tree object.}

\item{treenames}{A character vector of new tip labels (pop names) to replace current names.}

\item{groupvec}{A numeric vector indicating the group affiliation for each population.}

\item{regioncol}{A vector of colors the same length as \code{max(groupvec)} (i.e., the number of groups). The colors can be hexadecimal, R color names, or the number corresponding to an R color name in \code{\link[=colors]{colors()}}.}

\item{regionpch}{A vector with the same structure as \code{regioncol}, but each element is an R pch (plot character) number.}

\item{write_nexus}{Logical value indicating whether to write out a Nexus tree file that can be opened in FigTree. (see details)}

\item{file}{When \code{write_nexus = TRUE}, the full file path for writing out a Nexus tree file including the .nex extension.}
}
\value{
This function creates a new tree object with colors and tip labels and can also write out a Nexus tree file that can be opened and modified in FigTree (see details)
}
\description{
This function takes a tree object produced by \code{\link[ape:nj]{ape::nj()}} and adds group colors to the branch lengths for each reporting groups. Additionally, this function can used to modified the tip labels with new name and optional symbols.
}
\details{
FigTree can be downloaded from GitHub: \href{https://github.com/rambaut/figtree/releases}{FigTree Releases}.
}
\examples{
tree <- ape::nj(GCLr::ex_pairwise_fst) #Create tree from pairwise Fst matrix
currentnames <- tree$tip.label #current tip labels
treenames <- currentnames #tip labels for tree
groupvec <- GCLr::ex_baseline \%>\%
            dplyr::select(repunit, collection) \%>\%
            dplyr::distinct() \%>\%
            dplyr::mutate(repunit = factor(repunit) \%>\% as.numeric()) \%>\%
            dplyr::pull(repunit)

r_colors <- c("green", "blue", "red") # R color names example
hex_colors <- GCLr::col2hex(r_colors) \%>\% dplyr::pull(hex)# Hexadecimal colors example
r_color_numbers <- match(r_colors, colors()) # R color numbers example

colortree <- GCLr::add_tree_color(tree = tree,
                            currentnames = currentnames,
                            treenames = treenames,
                            groupvec = groupvec,
                            regioncol = hex_colors,
                            regionpch = unique(groupvec),
                            write_nexus = FALSE,
                            file = NULL)

colortree$tree$edge.length[colortree$tree$edge.length < 0] <- 0 #Remove negative branch lengths
ape::plot.phylo(x = colortree$tree, edge.color = colortree$color, edge.width = 3, use.edge.length = TRUE, show.tip.label = TRUE, adj = .02, cex = .7, font = 1, label.offset = 0.002)
ape::tiplabels(pch = colortree$pch , offset = .0015, cex = .5)

}
