% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_input_copy_paste.r
\name{bayes_input_copy_paste}
\alias{bayes_input_copy_paste}
\title{Copy and Paste BAYES Input Files}
\usage{
bayes_input_copy_paste(origindir, targetdir, mixvec, i = "A")
}
\arguments{
\item{origindir}{Original directory of files.
A character vector indicating where BAYES files are located.}

\item{targetdir}{Target directory structure.
A character vector indicating where to paste BAYES files (i.e., on the server).}

\item{mixvec}{Mixtures to be moved.
A character vector indicating which mixtures to move.}

\item{i}{Bayes folder identifier.
Indicates which "Bayes" folder to start putting files into.
Default is to put the first mixture into "Bayes A", second into "Bayes B", etc.
This function assumes that the Bayes folders on the V drive are listed alphabetically.}
}
\value{
This function has no explicit return value. It copies and pastes the files onto the server.
}
\description{
This function copies the BAYES input files (.ctl, .mix) from the source directory
and pastes them into the target directory to run BAYES.
}
\examples{
\dontrun{
bayes_input_copy_paste(origindir = "V:/Analysis/4_Westward/Sockeye/KMA Commercial Harvest 2014-2016/Mixtures/BAYES/Late August 89loci",
                       targetdir = "C:/Users/krshedd/BAYES", 
                       mixvec = dget(file = "V:/Analysis/4_Westward/Sockeye/KMA Commercial Harvest 2014-2016/Mixtures/Objects/LateAugustMixtures2014Strata.txt"),
                       i = "A")

}

}
