% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bayes_ctl.r
\name{create_bayes_ctl}
\alias{create_bayes_ctl}
\title{Create Bayes Control Files}
\usage{
create_bayes_ctl(
  sillyvec,
  loci,
  mixvec,
  baseline_name,
  nreps = 40000,
  nchains,
  groupvec,
  priorvec,
  initmat,
  dir,
  seeds = matrix(sample(seq(10000), 3 * nchains), nrow = 3),
  thin = c(1, 1, 1),
  mixfortran,
  basefortran,
  switches = "F T F T F T F"
)
}
\arguments{
\item{sillyvec}{Character vector of population sillys used to create the baseline file.}

\item{loci}{Character vector of the loci used to produce the baseline and mixture files.}

\item{mixvec}{Character vector of mixture ".gcl" objects for which control files are to be created.}

\item{baseline_name}{Character string of the baseline file name without the .bse extension.}

\item{nreps}{Numeric; the number of MCMC iterations for the mixtures.}

\item{nchains}{The number of MCMC chains for the mixtures.}

\item{groupvec}{Numeric vector of group numbers corresponding to each silly in \code{sillyvec}.}

\item{priorvec}{Numeric vector of priors corresponding to each silly in \code{sillyvec}.}

\item{initmat}{Numeric matrix of initial starting values for each chain with, nrow(\code{initmat}) == length(\code{sillyvec}) & ncol(\code{initmat}) == nchains.}

\item{dir}{Character vector of the directory where the control files will be saved.}

\item{seeds}{Matrix of random seeds containing 3 seeds per chain with, nrow(\code{seeds}) == 3 & ncol(\code{seeds}) == \code{nchains}.}

\item{thin}{Thinning intervals for MCMC sampling of 1) stock proportions, 2) baseline allele or type relative frequencies, and 3) stock assignments of each mixture individual.}

\item{mixfortran}{The Fortran format of the mixture files, as created by \code{\link[=create_bayes_mix]{create_bayes_mix()}}.}

\item{basefortran}{The Fortran format of the baseline file,as created by \code{\link[=create_bayes_base]{create_bayes_base()}}.}

\item{switches}{Character string of logical switches:
\itemize{
\item Switch 1: Baseline file printed (default: "F")
\item Switch 2: Mixture file printed (default: "T")
\item Switch 3: MCMC sample of baseline relative frequencies printed (default: "F")
\item Switch 4: Baseline file contains counts, not relative frequencies (default: "T")
\item Switch 5: Stock assignment distribution for each individual in the mixture sample printed (default: "F")
\item Switch 6: Stock-group or regional estimates printed (default: "T")
\item Switch 7: Individual stock estimates are suppressed (default: "F")
}}
}
\value{
This function writes out bayes control (.ctl) files to \code{dir}
}
\description{
This function creates BAYES control (.ctl) files for each mixture in \code{mixvec} and each chain in \code{nchains}. It takes
various arguments such as the population sillys used to create the baseline file, the loci used to produce the baseline
and mixture files, the mixtures for which control files are to be created, the directory where the control files will
be saved, the baseline name (without the .bse extension), the number of MCMC iterations, the number of MCMC chains, the
group numbers and priors for each silly, the initial starting values for each chain, the random seeds, the thinning
intervals for MCMC sampling, the Fortran format of the mixture and baseline files, and various logical switches.
}
\examples{
\dontrun{
sillyvec <- c("KSUSC18FW", "KSUSCN18", "KSUSC19FW", "KSUSCN19")
loci <- c("locus1", "locus2", "locus3")
mixvec <- c("mixture1", "mixture2")
baseline_name <- "baseline"
nreps <- 40000
nchains <- 5
groupvec <- c(1, 2, 3, 4)
priorvec <- c(0.1, 0.2, 0.3, 0.4)
initmat <- matrix(0.1, nrow = 4, ncol = 5)
dir <- "control_files"
seeds <- matrix(12345, nrow = 3, ncol = 5)
thin <- c(1, 1, 1)
mixfortran <- "(1X,I6)"
basefortran <- "(1X,I6)"
switches <- "F T F T F T F"
create_bayes_ctl(sillyvec, loci, mixvec, baseline_name, nreps, nchains, groupvec, priorvec, initmat, dir, seeds, thin, mixfortran, basefortran, switches)
}

}
\seealso{
See bayes manual for addtional details:  \verb{V:\\Software\\BAYES\\MANUAL.DOC}
}
