% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hierfstat_data.r
\name{create_hierfstat_data}
\alias{create_hierfstat_data}
\alias{create_hierfstat_data.GCL}
\title{Create a hierfstat data object}
\usage{
create_hierfstat_data(sillyvec, region = NULL, pop, loci, ncores = 4)
}
\arguments{
\item{sillyvec}{a vector of silly codes without the ".gcl" extension}

\item{region}{optional; a numeric vector indicating the regional affiliation for each silly in \code{sillyvec}. Include this argument when computing hierarchical F statistics.}

\item{pop}{a numeric vector indicating the population affiliation for each silly in \code{sillyvec}. If there is only one silly per population in \code{sillyvec}, then this should be a sequence of numbers of length \code{sillyvec}.}

\item{loci}{a character vector of locus names.}

\item{ncores}{a numeric value indicating the number of cores to use.}
}
\value{
This function returns a \link{hierfstat} data object containing region (if supplied), population, sub-population numbers, and genotypes in single-column format.
}
\description{
This function creates a \pkg{hierfstat} data object using the provided inputs. The \pkg{hierfstat} data object contains information about region (if supplied), population, sub-population numbers, and genotypes in single-column format. The object is used for computing hierarchical F statistics using functions such as \code{\link[=varcomp]{varcomp()}} and \code{\link[=varcomp.glob]{varcomp.glob()}} from the \link{hierfstat} package.
}
\examples{
sillyvec <- c("KQUART06", "KQUART08", "KQUART10")
region <- c(1, 2, 3)
pop <- c(1, 1, 2)
loci <- c("locus1", "locus2", "locus3")
ncores <- 4
create_hierfstat_data(sillyvec, region, pop, loci, ncores)

}
