% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hwler_ctl.r
\name{create_hwler_ctl}
\alias{create_hwler_ctl}
\title{Create HWLER Control Files}
\usage{
create_hwler_ctl(
  sillyvec,
  loci,
  input,
  mixbase = "mix",
  nsamples = c(1000, 2, 5),
  nchains = 5,
  dir,
  initval = "1.0",
  seeds = matrix(sample(seq(10000), 3 * nchains), nrow = 3),
  thin = c(1, 1, 1),
  inputfortran,
  switches = "T T T F T T T T T"
)
}
\arguments{
\item{sillyvec}{A character vector of silly codes for the populations used to create the baseline file.}

\item{loci}{character vector of the loci used to produce the baseline and mixture files.}

\item{input}{The base name for the control files.}

\item{mixbase}{A character vector indicating mixture and/or baseline individuals are present in the data.}

\item{nsamples}{A vector specifying the number of samples for each chain.}

\item{nchains}{The number of MCMC chains to analyze the mixtures.}

\item{dir}{The directory path where the control files will be saved.}

\item{initval}{The initial starting value for each MCMC chain.}

\item{seeds}{A matrix of random seeds containing 3 seeds per chain, where \code{nrow(seeds) == 3} and \code{ncol(seeds) == nchains}.}

\item{thin}{Thinning intervals for MCMC sample of 1) stock proportions, 2) baseline allele or type relative frequencies, and 3) stock assignments of each mixture individual.}

\item{inputfortran}{The input is in The Fortran format.}

\item{switches}{A character string of logical switches, with default value "T T T F T T T T T" (see details).}
}
\value{
Writes out HWLER control (.ctl) files.
}
\description{
This function creates HWLER (Hardy-Weinberg and Linkage Equilibrium Sampler) control files for each chain based on the specified parameters.
}
\details{
The \code{switches} argument has 9 program options turned “on” with “T” for true or “off” with “F” for false:
\enumerate{
\item Baseline printed
\item Mixture printed
\item Output stock assignments of individuals (baseline & mixture) or snapshots of cluster assignments of individuals (baseline or mixture only)
\item Suppress the printing of MCMC samples of stock composition (baseline & mixture) or snapshots of numbers of individuals assigned to clusters (baseline or mixture only)
\item Baseline provided
\item Mixture provided
\item Individual identification provided in data file
\item Output cluster information for the binary tree
\item Sample the Dirichlet mass parameter
}
}
\examples{
GCLr::create_hwler_ctl()

}
