% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hwler_ctl.r
\name{create_hwler_ctl}
\alias{create_hwler_ctl}
\title{Create HWLER Control Files}
\usage{
create_hwler_ctl(
  sillyvec,
  loci,
  input,
  mixbase = "mix",
  nsamples = c(1000, 2, 5),
  nchains = 5,
  dir,
  initval = "1.0",
  seeds = matrix(sample(seq(10000), 3 * nchains), nrow = 3),
  thin = c(1, 1, 1),
  inputfortran,
  switches = "T T T F T T T T T"
)
}
\arguments{
\item{sillyvec}{A character vector of population sillys used to create the baseline file.}

\item{loci}{character vector of the loci used to produce the baseline and mixture files.}

\item{input}{The base name for the control files.}

\item{mixbase}{A character vector indicating mixture and/or baseline individuals are present in the data.}

\item{nsamples}{A vector specifying the number of samples for each chain.}

\item{nchains}{The number of MCMC chains to analyze the mixtures.}

\item{dir}{The directory path where the control files will be saved.}

\item{initval}{The initial starting value for each MCMC chain.}

\item{seeds}{A matrix of random seeds containing 3 seeds per chain, where nrow(seeds) == 3 and ncol(seeds) == nchains.}

\item{thin}{Thinning intervals for MCMC sample of 1) stock proportions, 2) baseline allele or type relative frequencies, and 3) stock assignments of each mixture individual.}

\item{inputfortran}{The input is in The Fortran format.}

\item{switches}{A character string of logical switches, with default value "T T T F T T T T T" (see details).}
}
\value{
Writes out HWLER control (.ctl) files.
}
\description{
This function creates HWLER control files for each chain based on the specified parameters.
}
\details{
The \code{switches} argument has 9 program options turned “on” with “T” for true or “off” with “F” for false:
\enumerate{
\item Print summary of baseline data to summary file
\item Print mixture data to summary file
\item Output MCMC population assignments of individuals or snapshots of cluster assignments of individuals
\item Suppress the output of MCMC samples of population proportions or snapshots of numbers of individuals assigned to each cluster
\item Saseline samples provided
\item Mixture samples provided
\item Alphanumeric identification for each baseline or mixture individual provided at the end of each line of data file
\item Output MCMC samples of partitions for making a binary tree
\item Perform MCMC sampling of the Dirichlet mass parameter
}
}
\examples{
GCLr::create_hwler_ctl()

}
