% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hwler_input.R
\name{create_hwler_input}
\alias{create_hwler_input}
\title{Create a HWLER input file}
\usage{
create_hwler_input(sillyvec, loci, file_name, samplevec, LocCtl = LocusControl)
}
\arguments{
\item{sillyvec}{a vector of silly codes without the ".gcl" extension. These can be mixture sillys, baseline sillys, or mixture and baseline sillys (see details)}

\item{loci}{character vector of locus names}

\item{file_name}{the path to write out the input file with .input extension (e.g. "HWLER/input/myhwler.input").}

\item{samplevec}{a vector of numbers the same length as sillys, with baseline pops getting numbered 1:npops and mixtures getting zeros. This lets HWLER know if an individual is mixture or baseline.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
this function returns an object called \code{inputfortran} containing the FORTRAN format of the input file - this object is needed for \code{\link[=create_hwler_input]{create_hwler_input()}}
}
\description{
This function creates a HWLER input file containing mixture data only, baseline data only, or a combination mixture and baseline.
}
\details{
The HWLER input file can contain only mixture individuals, only baseline individuals, or a combination of mixture and baseline individuals depending on the type of analysis you are doing.
}
\examples{
\dontrun{

 load("V:/Analysis/2_Central/Sockeye/Cook Inlet/Missing Baseline Analysis/Missing Baseline Analysis.RData")
 create_HWLER_mixture.GCL(sillyvec = c(PopNames69, "Western07"), loci = loci34, file_name = "HWLER/input/myHWLER.input", samplevec = c(seq(length(PopNames69)), 0))

}

}
\seealso{
See HWLER manual for additional details: \link{HWLER manual}(system.file("HWLER", "HWLER_manual.doc", package = "GCLr"))
}
