% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mds_plot.R
\name{create_mds_plot}
\alias{create_mds_plot}
\title{Create an Interactive Multidimensional Scaling (MDS) Plot}
\usage{
create_mds_plot(
  file,
  dist.mat,
  pop_names,
  groupvec,
  group_names,
  group_colors,
  title = NULL,
  labels = FALSE,
  showlegend = TRUE
)
}
\arguments{
\item{file}{the file path (including html extension) to save plot as an html. If NULL, no file will be saved (defult = NULL)}

\item{dist.mat}{pairwise distance matrix}

\item{pop_names}{a character vector of population names corresponding to each population in \code{dist.mat} to add hover labels to the points on the plot, if no names are supplied the labels will default to dimnames(dist.mat)[\link{1}]}

\item{groupvec}{A numeric vector indicating the group affiliation of each population in \code{dist.mat}}

\item{group_names}{A character vector of group names with a length equal to the maximum value in \code{groupvec}}

\item{group_colors}{a character vector of colors corresponding to each group in \code{group_names}}

\item{title}{the title to place at the top of the plot; if NULL (default) no title will be added to the plot}

\item{labels}{weather to add labels the points on the MDS plot (default = FALSE)}

\item{showlegend}{whether to include a legend for the group colors (default = TRUE)}
}
\value{
an interactive MDS plot
}
\description{
This function takes a pairwise matrix of genetic distances (e.g. Fst, CSE, Nei's) and creates an interactive MDS plot that can be save as an .html.
}
\examples{
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

loci <- GCLr::ex_baseline[,-c(1:5)] \%>\%
  names() \%>\%
  gsub(pattern = "*\\\\.1", x = ., replacement = "") \%>\%
  unique()

dist.mat <- GCLr::pairwise_fst(sillyvec = sillyvec, loci = loci, inputfile = system.file("genepop", "ex_genepop.txt", package = "GCLr"), ncores = 4)

group_info <- GCLr::ex_baseline \%>\% 
  dplyr::group_by(collection) \%>\% 
  dplyr::summarise(group = unique(repunit) \%>\% factor())

GCLr::create_mds_plot(file = NULL, dist.mat = dist.mat, pop_names = group_info$collection, groupvec = group_info$group \%>\% as.numeric(), group_names = levels(group_info$group), group_colors = c("red", "blue", "green"), title = "Example MDS plot using pairwise Fst matrix", labels = TRUE, showlegend = TRUE)

}
