% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pwfst_tree.r
\name{create_pwfst_tree}
\alias{create_pwfst_tree}
\title{Create Pairwise Fst Tree}
\usage{
create_pwfst_tree(
  sillyvec,
  loci,
  dir,
  nboots = 1000,
  ncores = 4,
  returnbootstrapFst = FALSE
)
}
\arguments{
\item{sillyvec}{A character vector of silly codes without the ".gcl" extension.}

\item{loci}{A character vector of locus names.}

\item{dir}{Directory to save the \code{PairwiseFstTree} list object using \code{\link[base:dput]{base::dput()}}.}

\item{nboots}{A numeric value indicating the number of bootstrap iterations.}

\item{ncores}{A numeric value for the number of cores to use in a \pkg{foreach} \verb{\%dopar\%} loop.
If the number of cores exceeds the number on your device, \code{ncores} defaults to \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{returnbootstrapFst}{A logical value indicating whether to return the Fst matrix for each bootstrap iteration.}
}
\value{
A list with the following 4 or 5 components:
\itemize{
\item \code{tree}: a neighbor joining tree list of 4 created by \code{\link[ape:nj]{ape::nj()}} containing:
\itemize{
\item \code{edge}: numeric matrix
\item \code{edge.length}: numeric vector
\item \code{tip.lable}: character vector of population names, inherited from \code{sillyvec}
\item \code{Nnode}: integer
}
\item \code{bootstrap}: numeric vector of node boodstrap values
\item \code{PairwiseFst}: numeric matrix (\code{length(sillyvec)} x \code{length(sillyvec)}) of pairwise Fst values
\item \code{vc}: list of \code{choose(n = length(sillyvec), k = 2)} with pairwise variance components from \code{\link[hierfstat:varcomp]{hierfstat::varcomp()}}
\item \code{BootstrapFst}: list of \code{nboots} with numeric matrix of pairwise Fst values for each bootstrap iteration
(optional, only returned if \code{returnbootstrapFst = TRUE})
}
This list object is saved in \code{dir} using \code{\link[base:dput]{base::dput()}}, and is named \code{paste(dir,"\\\\", length(sillyvec), "Pops", length(loci), "Loci_", "PairwiseFstTree.txt", sep = "")}
}
\description{
This function generates a matrix of pairwise Fst values, a neighbor joining tree, bootstrap values for tree nodes, and variance components.
It can utilize multicore processing with \pkg{foreach} to speed up the calculation of variance components.
}
\details{
Older versions of this function used to write out an FSTAT .dat file, but this is no longer the case.
The function now calls on \code{\link[=create_hierfstat_data]{create_hierfstat_data()}} to create a \pkg{hierfstat} data object internally.
Depending on the size of your baseline and number of bootstrap iterations, this function can take a while.
}
\examples{

source(paste0(path.expand("~/R/"), "Functions.R")) # GCL functions
load("V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/Susitna_Chinook_baseline_2020.Rdata")
PWFSTtree <- create_pwfst_tree(sillyvec = sillyvec31, loci = loci82, dir = "output", nboots = 1000, ncores = 8, returnbootstrapFst = FALSE)

}
