% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rubias_base.r
\name{create_rubias_base}
\alias{create_rubias_base}
\title{Create rubias Baseline}
\usage{
create_rubias_base(
  sillyvec,
  loci,
  group_names,
  groupvec,
  file = "rubias/baseline",
  baseline_name
)
}
\arguments{
\item{sillyvec}{A character vector of the silly codes for the populations to include in the baseline.}

\item{loci}{A character vector of the loci to include in the baseline.}

\item{group_names}{A character vector of reporting group names, where \code{length(group_names) == max(`groupvec`)}.}

\item{groupvec}{A numeric vector indicating the reporting group affiliation of each population in \code{sillyvec}, where \code{length(groupvec) == length(sillyvec)}.}

\item{file}{A character vector of where to save the baseline as ".csv" file (see details).}

\item{baseline_name}{A character vector of what to name the baseline ".csv" file (see details).}
}
\value{
A dataframe in \pkg{rubias} baseline format and writes out a ".csv" file of the baseline.
}
\description{
This function creates the baseline dataframe needed for \pkg{rubias}.
}
\details{
This function saves the baseline as a ".csv" file for posterity. To read in these ".csv" files, use \code{readr::read_csv(file = file, col_types = cols(.default = "c")} to make sure all columns are character vectors (if homozygous for T, it will become a logical vector).
}
\examples{
create_locuscontrol(markersuite = "Coho_Baseline2016_95SNPs", username = "awbarclay", password = password)
dat <- readr::read_csv("V:/Analysis/2_Central/Coho/Cook Inlet/2019/2019_Cook_Inlet_coho_baseline/output/Final_Pops.csv")
load("V:/Analysis/2_Central/Coho/Cook Inlet/2019/2019_Cook_Inlet_coho_baseline/2019_Cook_Inlet_coho_baseline.Rdata")
#' sillyvec <- dat$collection
old2new_gcl(sillyvec)
group_names <- dat$groups_10 \%>\% unique()
groupvec <- dat$groups_10 \%>\% factor(levels = group_names) \%>\% as.numeric()

UCIcoho104pops_96loci.rubias_base <- GCLr::create_rubias_base(sillyvec = sillyvec, loci = loci, group_names = group_names, groupvec = groupvec, file = "rubias/baseline", baseline_name = "UCIcoho104pops_96loci")

}
