% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_comb_rubias_output.r
\name{custom_comb_rubias_output}
\alias{custom_comb_rubias_output}
\title{Summarize \code{rubias} Output}
\usage{
custom_comb_rubias_output(
  rubias_output = NULL,
  mixvec = NULL,
  group_names = NULL,
  group_names_new = NULL,
  groupvec = NULL,
  groupvec_new = NULL,
  path = "rubias/output",
  alpha = 0.1,
  burn_in = 5000,
  bias_corr = FALSE,
  threshold = 5e-07,
  plot_trace = TRUE,
  ncores = 4
)
}
\arguments{
\item{rubias_output}{}

\item{mixvec}{}

\item{group_names}{}

\item{group_names_new}{}

\item{groupvec}{}

\item{groupvec_new}{}

\item{path}{}

\item{alpha}{}

\item{burn_in}{}

\item{bias_corr}{}

\item{threshold}{}

\item{plot_trace}{}

\item{ncores}{}
}
\value{
A tibble with 8 columns:
\itemize{
\item \code{mixture_collection}: factor of mixtures (only a factor for ordering, plotting purposes)
\item \code{repunit}: factor of reporting groups (only a factor for ordering, plotting purposes)
\item \code{mean}: mean posterior of stock proportion
\item \code{sd}: sd of posterior of stock proportion
\item \code{medoam}: median posterior of stock proportion
\item \code{sd}: sd of posterior of stock proportion
\item \code{loCI}: lower 5\% CI from posterior of stock proportion
\item \code{hiCI}: upper 95\% CI from posterior of stock proportion
\item \code{P=0}: proportion of posterior with stock proportion < \code{threshold} (i.e., 0)
}
}
\description{
A short description...
}
\examples{
\dontrun{
rubias_out <- GCLr::custom_comb_rubias_output()
}

}
