% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failure_rate.r
\name{failure_rate}
\alias{failure_rate}
\alias{FailureRate.GCL.R}
\title{Calculate Failure Rates by Silly, Locus, Plate, and Project}
\usage{
failure_rate(sillyvec)
}
\arguments{
\item{sillyvec}{Character vector of SILLY codes in the project.}
}
\value{
A list of failure rates by SILLY, locus, plate, and project.
}
\description{
This function calculates failure rates by SILLY, locus, plate, and project.
It should be run after the "remove_na_indv" step but before any other steps
in the quality control (QC) process, such as removing duplicate fish,
alternate fish, or missing fish. Note that this function does not connect to LOKI;
it simply calculates failure rates (0's / total fish run) from the silly.gcl objects.
}
\details{
The \code{\link[=failure_rate]{failure_rate()}} function calculates failure rates by SILLY, locus, plate, and project. It performs the following steps:
\itemize{
\item Pools all collections into one master silly using the \code{\link[=pool_collections]{pool_collections()}} function
\item Creates a tibble of Dose 1 scores and attributes from the master silly.
\item Calculates failure rates by SILLY, locus, plate, and project using group_by and summarise functions from the \code{dplyr} package.
\item Generates plots of failure rates by SILLY and locus, as well as by plate and locus, using \code{ggplot2} and \code{plotly} packages.
\item Returns a list of failure rates and plots.
}

Note: This function relies on the \code{\link[=pool_collections]{pool_collections()}} function from the GCL package.
}
\examples{
failure_rate(sillyvec = c("SILLY1", "SILLY2", "SILLY3"))

}
\seealso{
\code{\link{remove_na_indv}}: Function to remove NA individuals
\code{\link{pool_collections}}: Function to pool collections
}
\author{
krshedd
}
\concept{QC Functions}
