% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcl2fstat.r
\name{gcl2fstat}
\alias{gcl2fstat}
\title{Convert ".gcl" Objects to FSTAT}
\usage{
gcl2fstat(sillyvec, loci, path, ncores = 4, LocusCtl = LocusControl)
}
\arguments{
\item{sillyvec}{A character vector of silly codes to include in the FSTAT file.}

\item{loci}{AA character vector of locus names to include in the FSTAT file.}

\item{path}{The full file path to write out the FSTAT file; with "\\" or "/" separator between folders.}

\item{ncores}{The number of cores for multithreading using \code{\link[=doParallel]{doParallel()}} and \code{\link[=foreach]{foreach()}}. Default is 4.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
Writes out an FSTAT file to the specified path.
}
\description{
Write out an FSTAT input file from ".gcl" objects.
}
\details{
This function requires a \code{LocusControl} object. Run \code{\link[=create_locuscontrol]{create_locuscontrol()}} prior to this function.
}
\examples{
load("V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/Susitna_Chinook_baseline_2020.Rdata")
GCLr::gcl2fstat(sillyvec = sillyvec31, loci = loci82, path = "FSTAT/FSTATfile.dat", ncores = 8)

}
