% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcl2fstat.r
\name{gcl2fstat}
\alias{gcl2fstat}
\title{Convert ".gcl" Objects to FSTAT}
\usage{
gcl2fstat(sillyvec, loci, path, ncores = 4, LocusCtl = LocusControl)
}
\arguments{
\item{sillyvec}{A character vector of silly codes to include in the FSTAT file.}

\item{loci}{A character vector of locus names to include in the FSTAT file.}

\item{path}{The full file path to write out the FSTAT file including ".dat" extension; with "\\" or "/" separator between folders.}

\item{ncores}{The number of cores for multithreading using \code{\link[=doParallel]{doParallel()}} and \code{\link[=foreach]{foreach()}}. Default is 4.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
Writes out an FSTAT file to the specified path.
}
\description{
Write out an FSTAT input file from ".gcl" objects.
}
\details{
This function requires a \code{LocusControl} object. Run \code{\link[=create_locuscontrol]{create_locuscontrol()}} prior to this function.
Use this function when you want to run an analysis in \code{FSTAT}.
}
\examples{

sillys <- GCLr::base2gcl(GCLr::ex_baseline)

loci <- GCLr::ex_baseline[,-c(1:5)] \%>\%
  names() \%>\%
  gsub(pattern = "*\\\\.1", x = ., replacement = "") \%>\%
  unique()
  
GCLr::gcl2fstat(sillyvec = sillys, loci = loci, path = path.expand("~/FSTATfile.dat"), ncores = 4)

}
\seealso{
\code{\link[=create_hierfstat_data]{create_hierfstat_data()}}
}
