% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcl2genepop.r
\name{gcl2genepop}
\alias{gcl2genepop}
\title{Convert ".gcl" Objects to GENEPOP File}
\usage{
gcl2genepop(
  sillyvec,
  loci,
  path,
  VialNums = TRUE,
  usat = FALSE,
  ncores = 4,
  npops = NULL,
  LocusCtl = LocusControl
)
}
\arguments{
\item{sillyvec}{A character vector of silly codes to include in the GENEPOP file.}

\item{loci}{A character vector of locus names to include in the GENEPOP file.}

\item{path}{When \code{npops = NULL} full file path to write out the GENEPOP file; with "\\" or "/" separator between folders.
When setting \code{npops}, \code{path} is the folder were the generic genepop files are saved.}

\item{VialNums}{Logical; if TRUE (default), vial numbers will be included for each individual next to their silly code separated by an underscore (e.g., "KCRESC10_1"). If FALSE, only the silly code will be included for each individual (e.g., "KCRESC10").}

\item{usat}{Logical; whether the data are from microsatellites (TRUE) or not (FALSE). This is included because GENEPOP only accepts numeric alleles; SNP alleles have to be converted from character to numeric, and microsatellite alleles are already numeric.}

\item{ncores}{The number of cores for multithreading using \code{\link[=doParallel]{doParallel()}} and \code{\link[=foreach]{foreach()}}. Default is 4.}

\item{npops}{Optional; the number of populations for each file. If supplied, multiple files with generic file names will be written out, with \code{npops} per file (see details).}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
Writes out a GENEPOP file to the specified path.
}
\description{
Write out a GENEPOP input file from ".gcl" objects.
}
\details{
This function requires a \code{LocusControl} object. Run \code{\link[=create_locuscontrol]{create_locuscontrol()}} prior to this function.
The \code{npops} argument is useful when running Linkage Disequilibrium tests in \pkg{genepop} using the \link{gcltest_LD} function, so things finish faster.
For example, if there are 10 pops in \code{sillyvec} and \code{npops} = 2, the following files will be written to the folder supplied in the \code{path} argument: Pops1to2.gen.txt, Pops3to4.gen.txt, Pops5to6.gen.txt, Pops7to8.gen.txt, Pops9to10.gen.txt
}
\examples{
### SNP
GCLr::gcl2genepop(sillyvec = sillyvec, loci = loci, path = "GENEPOP/genepopfile.gen", VialNums = TRUE, usat = FALSE, ncores = 8, npops = NULL)

### uSAT
GCLr::gcl2genepop(sillyvec = sillyvec, loci = loci, path = "Genepop/genepopfile.gen", VialNums = TRUE, usat = TRUE, ncores = 4)

### SNP with npops argument
GCLr::gcl2genepop(sillyvec = sillyvec, loci = loci, path = "Genepop", VialNums = TRUE, usat = FALSE, ncores = 8, npops = 2)

}
