% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcl2genepop.r
\name{gcl2genepop}
\alias{gcl2genepop}
\title{Convert ".gcl" Objects to GENEPOP File}
\usage{
gcl2genepop(
  sillyvec,
  loci,
  path,
  VialNums = TRUE,
  usat = FALSE,
  ncores = 4,
  npops = NULL,
  LocusCtl = LocusControl
)
}
\arguments{
\item{sillyvec}{A character vector of silly codes to include in the GENEPOP file.}

\item{loci}{A character vector of locus names to include in the GENEPOP file.}

\item{path}{The full file path to write out the GENEPOP file; with "\\" or "/" separator between folders.}

\item{VialNums}{Logical; if TRUE (default), vial numbers will be included for each individual next to their silly code separated by an underscore (e.g., "KCRESC10_1"). If FALSE, only the silly code will be included for each individual (e.g., "KCRESC10").}

\item{usat}{Logical; whether the data are from microsatellites (TRUE) or not (FALSE). This is included because GENEPOP only accepts numeric alleles; SNP alleles have to be converted from character to numeric, and microsatellite alleles are already numeric.}

\item{ncores}{The number of cores for multithreading using \code{\link[=doParallel]{doParallel()}} and \code{\link[=foreach]{foreach()}}. Default is 4.}

\item{npops}{Optional; the number of populations for each file. If supplied, multiple files with generic file names will be written out, with \code{npops} per file (see details).}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
Writes out a GENEPOP file to the specified path.
}
\description{
Write out a GENEPOP input file from ".gcl" objects.
}
\details{
This function requires a \code{LocusControl} object. Run \code{\link[=create_locuscontrol]{create_locuscontrol()}} prior to this function.
The \code{npops} argument is useful when running Linkage Disequilibrium tests in \pkg{genepop} using the \link{gcltest_LD} function so things finish faster. For example, if there are 10 pops in \code{sillyvec} and \code{npops} = 2, the following files will be written to the folder supplied in the \code{path} argument: Pops1to2.gen.txt, Pops3to4.gen.txt, Pops5to6.gen.txt, Pops7to8.gen.txt, Pops9to10.gen.txt
}
\examples{
source(paste0(path.expand("~/R/"), "Functions.R"))#GCL functions

### SNP
attach("V:/Analysis/2_Central/Coho/Cook Inlet/2019/2019_Cook_Inlet_coho_baseline/2019_Cook_Inlet_coho_baseline.RData")
old2new_locuscontrol()
sapply(sillyvec104, function(silly){assign(paste0(silly, ".gcl"), get(paste0(silly, ".gcl")), pos = -1, envir = .GlobalEnv)})
sillyvec <- sillyvec104
loci <- loci81
detach()
old2new_gcl(sillyvec)
GCLr::gcl2genepop(sillyvec = sillyvec, loci = loci, path = "GENEPOP/genepopfile.gen", VialNums = TRUE, usat = FALSE, ncores = 8, npops = NULL)

### uSAT
attach("V:/Analysis/2_Central/Chinook/Susitna-Watana/2017/Upper and Middle River Analysis/SuWuUpperMiddleRiverAnalysis_tidy.RData")
sapply(Sillys, function(silly){assign(paste0(silly, ".gcl"), get(paste0(silly, ".gcl")), pos = -1, envir = .GlobalEnv)})
sillyvec <- Sillys
loci <- loci13
LocusControl <- LocusControl
detach()
GCLr::gcl2genepop(sillyvec = sillyvec, loci = loci, path = "V:/Analysis/2_Central/Chinook/Susitna-Watana/2017/Upper and Middle River Analysis/Genepop/genepopfile.gen", VialNums = TRUE, usat = TRUE, ncores = 4)

### SNP with npops argument
attach("C:/2019 coho baseline/2019_Cook_Inlet_coho_baseline_new.RData")
LocusControl <- LocusControl
sapply(sillyvec104, function(silly){assign(paste0(silly, ".gcl"), get(paste0(silly, ".gcl")), pos = -1, envir = .GlobalEnv)})
sillyvec <- sillyvec104
loci <- loci81
detach()
GCLr::gcl2genepop(sillyvec = sillyvec[1:36], loci = loci, path = "GENEPOP", VialNums = TRUE, usat = FALSE, ncores = 8, npops = 2)

}
