% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcl2nexus.r
\name{gcl2nexus}
\alias{gcl2nexus}
\title{Convert ".gcl" Objects to NEXUS File}
\usage{
gcl2nexus(sillyvec, loci, path, VialNums = TRUE, PopNames = NULL, ncores = 4)
}
\arguments{
\item{sillyvec}{A character vector of silly codes to include in the NEXUS file.}

\item{loci}{A character vector of locus names to include in the NEXUS file.}

\item{path}{The full file path to write out the NEXUS file; with "\\" or "/" separator between folders.}

\item{VialNums}{Logical; if TRUE (default), vial numbers will be included for each individual next to their silly code separated by an underscore (e.g., "KCRESC10_1"). If FALSE, only the SILLY code will be included for each individual (e.g., "KCRESC10").}

\item{PopNames}{Optional character vector the same length as \code{sillyvec} to give populations new names. If NULL, \code{PopNames} defaults to \code{sillyvec}.}

\item{ncores}{The number of cores for multithreading using \code{\link[=doParallel]{doParallel()}} and \code{\link[=foreach]{foreach()}}. Default is 4.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
Writes out a NEXUS file to the specified path.
}
\description{
Write out a NEXUS file in GDA input format from ".gcl" objects.
}
\details{
This function requires a \code{LocusControl} object. Run \code{\link[=create_locuscontrol]{create_locuscontrol()}} prior to this function.
}
\examples{
load("V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/Susitna_Chinook_baseline_2020.Rdata")
combine_loci(sillyvec31, markerset = c("Ots_U211", "Ots_U212-297"), update = TRUE)
combine_loci(sillyvec31, markerset = c("Ots_UNKN4-150", "Ots_UNKN6-187"), update = TRUE)
GCLr::gcl2nexus(sillyvec = sillyvec31, loci = c(loci80, "Ots_U211.Ots_U212-297", "Ots_UNKN4-150.Ots_UNKN6-187"), path = "nexusfile.nex", VialNums = TRUE, PopNames = Final_Pops$location, ncores = 8)

}
