% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierfstat_global_stats.r
\name{hierfstat_global_stats}
\alias{hierfstat_global_stats}
\title{Compute global hierarchical F-statistics and variance components.}
\usage{
hierfstat_global_stats(levels, genotypes)
}
\arguments{
\item{levels}{A data frame containing the different levels (factors) from the outermost (e.g., region) to the innermost before the individual.}

\item{genotypes}{A data frame containing the genotypes in single-column numeric format (e.g., 101, 102, 202).}
}
\value{
A list containing 3 tibbles:
\itemize{
\item \code{loci} - variance components by locus,
\item \code{overall} - variance components summed over all loci,
\item \code{F} - hierarchical F-statistics (see \code{\link[hierfstat:varcomp.glob]{hierfstat::varcomp.glob()}} documentation).
}
}
\description{
This is a wrapper function for \code{\link[hierfstat:varcomp.glob]{hierfstat::varcomp.glob()}}. This wrapper will compute variance components and hierarchical F-statistics for haploid and diploid loci and combine the results into a single output.
}
\examples{
load("V:/Analysis/2_Central/Coho/Cook Inlet/2019/2019_Cook_Inlet_coho_baseline/2019_Cook_Inlet_coho_baseline_new.Rdata")

source(paste0(path.expand("~/R/"), "Functions.R")) # GCL functions

sillyvec104 <- Final_Pops$collection

region <- Final_Pops \%>\% select(pop = order, region = region)

temporal_collections <- temporal_collections(sillyvec = sillyvec104, region = region, min.samps = 50, sep = ".")

combine_loci(sillyvec = sillyvec187, markerset = c("Oki_102213-604", "Oki_101119-1006")) # Combining two loci to test haploid calcs

fstat.dat <- create_hierfstat_data(sillyvec = temporal_collections$silly, region = temporal_collections$region, pop = temporal_collections$pop, loci = c("Oki_102213-604.Oki_101119-1006", loci81), ncores = 8)

hierfstat_global_stats(levels = fstat.dat[,1:3], genotypes = fstat.dat[,-c(1:3)])

}
