% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.r
\name{load_objects}
\alias{load_objects}
\title{Load R Objects}
\usage{
load_objects(path, pattern = NULL, rds = FALSE)
}
\arguments{
\item{path}{A character vector specifying the path where the objects to load reside.}

\item{pattern}{Optional character vector argument to manually specify a pattern (i.e., specific object or regular expression). Accepts a list of patterns if you want to load multiple objects (e.g., \code{c("pattern1", "pattern2")}).}

\item{rds}{Logical; if set to TRUE, the function will load files in RDS format. If set to FALSE, the function will load text files produced by \code{\link[=dput]{dput()}}. The default is FALSE for backward compatibility.}
}
\value{
A character vector of the objects loaded into the workspace.
}
\description{
This function loads R objects saved with \code{\link[=dput]{dput()}} or \code{\link[=saveRDS]{saveRDS()}}, serving as a wrapper for \code{\link[=dget]{dget()}} and \code{\link[=readRDS]{readRDS()}}.
}
\examples{
setwd("V:/Analysis/1_SEAK/Sockeye/Mixture/Lynn Canal Inseason/2018/")
GCLr::load_objects(path = "Objects", pattern = "^loci") # Loads loci objects from "Objects" directory
GCLr::load_objects(path = "Objects", pattern = c("^loci", "sample_size")) # Loads loci and sample size objects from "Objects" directory

}
