% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_stats.r
\name{locus_stats}
\alias{locus_stats}
\title{Calculate Basic Locus Statistics}
\usage{
locus_stats(sillyvec, loci, ncores = 4, LocusCtl = LocusControl)
}
\arguments{
\item{sillyvec}{A character vector of silly codes without the ".gcl" extension.}

\item{loci}{A character vector of locus names.}

\item{ncores}{A numeric value for the number of cores to use in a \pkg{foreach} \verb{\%dopar\%} loop (default = 4).
The number of cores cannot exceeds the number on your device (\code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}} (default = LocusControl).}
}
\value{
A tibble with 1 row per locus in \code{loci} + "Overall" and the following 4 columns:
\itemize{
\item \code{locus}: locus name
\item \code{Ho}: observed heterozygosity calculated by \code{\link[hierfstat:basic.stats]{hierfstat::basic.stats()}}
\item \code{Fis}: Fis calculated by \code{\link[hierfstat:varcomp]{hierfstat::varcomp()}}
\item \code{Fst}: Fst calculated by \code{\link[hierfstat:varcomp]{hierfstat::varcomp()}}
}
}
\description{
This function calculates basic locus statistics (Ho, Fis, Fst) for each locus across collections (silly code).
}
\details{
This function uses \code{\link[hierfstat:hierfstat]{hierfstat::hierfstat()}} to calculate locus statistics for Ho, Fis, and Fst.
}
\examples{
  
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)
  
GCLr::locus_stats(sillyvec = sillyvec, loci = GCLr::ex_LocusControl$locusnames[-c(10, 12, 13, 32, 33)], ncores = parallel::detectCores(), LocusCtl = GCLr::ex_LocusControl )

}
\seealso{
\code{\link[hierfstat:hierfstat]{hierfstat::hierfstat()}}
\code{\link[hierfstat:varcomp]{hierfstat::varcomp()}}
\code{\link[hierfstat:basic.stats]{hierfstat::basic.stats()}}
\code{\link[=create_hierfstat_data]{create_hierfstat_data()}}
}
