% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loki2r_gaps.r
\name{loki2r_gaps}
\alias{loki2r_gaps}
\alias{LOKI2R_GAPS.GCL}
\title{Create GCL Objects (GAPS)}
\usage{
loki2r_gaps(sillyvec, username, password)

LOKI2R_GAPS.GCL(sillyvec, username, password)
}
\arguments{
\item{sillyvec}{Vector of silly codes to pull from LOKI.}

\item{username}{Username for LOKI connection.}

\item{password}{Password for LOKI connection.}
}
\value{
The runtime of the function in seconds.
}
\description{
This function connects to LOKI and creates a ".gcl" object for each silly code in \code{sillyvec}. A "*.gcl" object is a Gene Conservation Laboratory genotypes object with associated sample attributes.
}
\details{
This function requires the R package "RJDBC" for connecting to LOKI.
It connects to LOKI using the provided \code{username} and \code{password}, and retrieves genotypes and sample attributes for each silly code in \code{sillyvec}.
The genotypes and attributes are stored in separate "*.gcl" objects named after each silly code. This function requires an OJDBC driver object, which is an object in the GCLr package called \link{drv}.
}
\examples{
sillyvec <- c("KQUART06", "KQUART08", "KQUART09")
username <- "jjasper"
loki2r_gaps(sillyvec, username, password)

}
