% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loki2r_gaps.r
\name{loki2r_gaps}
\alias{loki2r_gaps}
\title{Create GCL Objects (GAPS)}
\usage{
loki2r_gaps(sillyvec, username, password, LocusControl = LocusControl)
}
\arguments{
\item{sillyvec}{Vector of silly codes to pull from LOKI.}

\item{username}{Username for LOKI connection.}

\item{password}{Password for LOKI connection.}

\item{LocusControl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}}
}
\value{
The runtime of the function in seconds.
}
\description{
This function connects to LOKI and creates a ".gcl" object for each silly code in \code{sillyvec}. A "*.gcl" object is a Gene Conservation Laboratory genotypes object with associated sample attributes.
}
\details{
This function requires the R package "RJDBC" for connecting to LOKI.
It connects to LOKI using the provided \code{username} and \code{password}, and retrieves genotypes and sample attributes for each silly code in \code{sillyvec}.
The genotypes and attributes are stored in separate "*.gcl" objects named after each silly code.
}
\examples{
\dontrun{
sillyvec <- c("KQUART06", "KQUART08", "KQUART09")
loki2r_gaps(sillyvec = sillyvec, username = "awbarclay", password, LocusControl = GCLr::LocusControl)
}

}
