% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baseline_eval.r
\name{plot_baseline_eval}
\alias{plot_baseline_eval}
\alias{plot_baseline_eval_summary.GCL}
\title{Plot Baseline Evaluation Results}
\usage{
plot_baseline_eval(
  summary,
  file,
  method = c("MCMC", "PB", "both"),
  test_groups = NULL,
  group_colors = NULL
)
}
\arguments{
\item{summary}{A baseline evaluation summary list produced by \code{\link[=summarize_rubias_baseline_eval]{summarize_rubias_baseline_eval()}}.}

\item{file}{A tibble produced by \code{\link[=base_eval_sample_sizes]{base_eval_sample_sizes()}} containing the following variables:
\itemize{
\item \code{test_group}
\item \code{scenario}
\item \code{repunit}
\item  \code{samps}
}}

\item{method}{A character string indicating the rubias output to summarize. Select one of three choices:
\itemize{
\item {"MCMC"(plot MCMC output)}
\item{"PB" (plot bias corrected output)}
\item{"both" (plot both outputs)}
}}

\item{test_groups}{A character vector of group names to include in the plots. This also sets the order in which they are plotted. If \code{test_groups} is not supplied, all test groups in \code{summary} will be plotted (see details).}

\item{group_colors}{A character vector of R \code{colors()} the same length as \code{test_groups}. If \code{group_colors} is not supplied, colors will be automatically selected using \code{rainbow()}.}
}
\value{
If \code{method == "MCMC"} or \code{"PB"}, the function produces a single page pdf file containing the plots and returns the faceted plots. If \code{method == "both"}, the function produces a two-page pdf file containing the plots for both methods and no plots are returned.
}
\description{
This function takes the baseline evaluation summary object produced by \code{\link[=summarize_rubias_baseline_eval]{summarize_rubias_baseline_eval()}} and produces plots of correct allocation for each test_group with summary statistics placed in the upper left corner of each plot.
}
\details{
The solid line on the plots indicates where the true proportion equals the estimated proportion. The two dotted lines indicate where the estimates fall within +/- 10\% of the true proportion.
If a large number of \code{test_groups} are supplied, the faceted plots will become too small to see on one page. If so, you may need to supply a subset of \code{test_groups} to plot.
This function is not intended for producing publication-ready plots; however, the code from this function can be copied and modified to produce plots formatted for publication.
}
\examples{
load("V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/Susitna_Chinook_baseline_2020.Rdata")
require(tidyverse)
tests <- sample_sizes \%>\% group_by(test_group, scenario) \%>\% summarize(test_group = test_group \%>\% unique(), scenario = scenario \%>\% unique(), .groups = "drop_last")#Total of 510 tests  #
mixvec <- tests \%>\% unite(col = "mixvec", test_group, scenario, sep ="_" ) \%>\% pull()
path <-  "V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/rubias/3groups/output"
summary <- summarize_rubias_base_eval (mixvec = mixvec, sample_sizes = sample_sizes, method = "both", group_names = NULL, group_names_new = NULL, groupvec = NULL, groupvec_new = NULL, path = path, alpha = 0.1, burn_in = 5000, threshold = 5e-7, ncores = 8)

plot_baseline_eval(summary = summary, file = "Baseline_eval_plots.pdf", method = "PB", test_groups = groups3, group_colors = c("green", "magenta", "red"))

}
