% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq_fis_4snps.r
\name{plot_freq_fis_4snps}
\alias{plot_freq_fis_4snps}
\title{Plot Allele Frequency & Fis for SNPs}
\usage{
plot_freq_fis_4snps(
  sillyvec,
  loci,
  groupvec = NULL,
  alpha = 0.05,
  groupcol = NULL,
  file = NULL,
  group.pch = 21,
  point.size = 3,
  pval.cex = 3,
  pval.digits = 2,
  line.width = 0.7,
  ncores = 4,
  LocusCtl = LocusControl
)
}
\arguments{
\item{sillyvec}{A character vector of silly codes without the ".gcl" extension, objects must exist in your environment.}

\item{loci}{A character vector of locus names. Note, only loci with ploidy = 2 and nalleles = 2 will be plotted.}

\item{groupvec}{An optional numeric vector indicating the group affiliation of each silly in \code{sillyvec},
must be the same length as \code{sillyvec} (default = \code{NULL}). If \code{NULL}, \code{groupvec} will be \code{1:length(sillyvec)}}

\item{alpha}{A numeric vector of length 1 specifying the critical HWE p-value for plotting.}

\item{file}{A character vector of length 1 with the full file path including ".pdf" extension where plots will be saved.}

\item{group.pch}{A vector of point shape \code{pch} values corresponding to each group with \code{length = max(groupvec)}.
If only 1 \code{pch} is supplied it will be recycled for all groups (default \code{pch} = 19).}

\item{point.size}{A numeric vector of length 1 specifying the size of points on the plot, used by \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
(default = 1).}

\item{pval.cex}{A numeric vector of length 1 specifying the character expansion (i.e. size of text) for HWE p-values (default = 3).}

\item{pval.digits}{A numeric vector of length 1 specifying the number of digits to show for HWE p-values (default = 2).}

\item{line.width}{A numeric vector of length 1 specifying the width of the line connecting the points, used by \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}
default = 0.5).}

\item{ncores}{A numeric value for the number of cores to use in a \pkg{foreach} \verb{\%dopar\%} loop (default = 4).
If the number of cores exceeds the number on your device (\code{\link[parallel:detectCores]{parallel::detectCores()}}), then all cores will be used.}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}} (default = LocusControl).}

\item{group_col}{An optional character vector of colors corresponding to each group in \code{groupvec} (default = \code{NULL}).}
}
\value{
A tibble of HWE p-values, Fis values, and allele frequencies and a ("\emph{.pdf") of locus-specific plots:
\itemize{
\item \code{output}: a data.frame with 7 columns containing the full HWE output:
\itemize{
\item \code{population}: silly code
\item \code{locus}: locus name
\item \code{pval}: HWE p-value as calculated by \code{\link[HardyWeinberg:HWExactMat]{HardyWeinberg::HWExactMat()}}
\item \code{fis}: Fis estimate as calculated by \code{\link[HardyWeinberg:HWf]{HardyWeinberg::HWf()}}
\item \code{allele freq}: allele frequency (proportion) of the 1st allele in \code{LocusCtl} as calculated by \code{\link[=calc_freq_pop]{calc_freq_pop()}}
}
\item \code{file}: a ("}.pdf") of stacked allele frequency and Fis plots for all sillys in \code{sillyvec}
with one page per locus in \code{loci}
}
}
\description{
This function plots ("*.pdf") and returns allele frequencies, Fis values, and HWE p-values for each locus.
}
\details{
HWE p-values are printed if less than alpha. Loci with a ploidy of 1 (e.g., mitochondrial SNPs or microhaplotypes) are excluded.
The allele frequency is that of the 1st allele in \code{LocusCtl} (i.e., alphabetical alleles).
}
\examples{
\dontrun{
freqs_fis <- GCLr::plot_freq_fis_4snps()
}

}
\seealso{
\code{\link[HardyWeinberg:HardyWeinberg-package]{HardyWeinberg::HardyWeinberg-package()}}
\code{\link[HardyWeinberg:HWExactMat]{HardyWeinberg::HWExactMat()}}
\code{\link[HardyWeinberg:HWf]{HardyWeinberg::HWf()}}
\code{\link[=calc_freq_pop]{calc_freq_pop()}}
\code{\link[=plot_allele_freq]{plot_allele_freq()}}
}
