% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quick_tree.r
\name{plot_quick_tree}
\alias{plot_quick_tree}
\title{Plot Quick Phylogenetic Tree}
\usage{
plot_quick_tree(
  sillyvec,
  loci,
  inputfile,
  popnames = NULL,
  ncores = 4,
  groupvec = NULL,
  regioncol = NULL,
  cex = 1
)
}
\arguments{
\item{sillyvec}{a vector of silly codes without the ".gcl" extension.}

\item{loci}{a vector of locus names.}

\item{inputfile}{the file path of the genepop input file including the .txt extension. If the file does not exist, then one will be created.}

\item{popnames}{an optional vector of population names corresponding to sillyvec in order to add them as the dimnames of the output matrix. If not supplied, sillyvec will be used as the dimnames.}

\item{ncores}{a numeric value indicating the number of cores to use}

\item{groupvec}{a numeric vector indicating the group affiliation for each population in sillyvec}

\item{regioncol}{a vector of colors of the same length as max(groupvec) (i.e., the number of groups). The colors can be hexadecimal, R color names, or the number corresponding to an R color name in \code{colors()}.}

\item{cex}{a numeric value giving the factor scaling of the tip labels (Character Expansion).}
}
\value{
a neighbor-joining phylogram based on pairwise Fst values.
}
\description{
This function quickly creates a neighbor-joining Fst tree using the \pkg{genepop} and \pkg{ape} packages.
}
\examples{
source(paste0(path.expand("~/R/"), "Functions.R"))  # GCL functions
load("V:/Analysis/2_Central/Chinook/Susitna River/Susitna_Chinook_baseline_2020/Susitna_Chinook_baseline_2020.Rdata")
plot_quick_tree(sillyvec = sillyvec31, loci = loci82, inputfile = "genepop/Susitna31pops82loci.txt", popnames = treenames31, ncores = 8, groupvec = groupvec4, regioncol = grcol4, cex = 0.75)

}
