% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sampn_by_locus.r
\name{plot_sampn_by_locus}
\alias{plot_sampn_by_locus}
\title{Plot Sample Size by Locus}
\usage{
plot_sampn_by_locus(SampSizeByLocus)
}
\arguments{
\item{SampSizeByLocus}{A data frame with two columns: "silly" and "count". The "silly"
column contains categorical data, and the "count" column contains the count of samples
for each category. This data frame is produced by \code{\link[=sampn_by_locus]{sampn_by_locus()}}.}
}
\value{
A \code{plotly} interactive plot showing a heatmap of the proportion of counts for each
locus relative to the overall sample size.
}
\description{
This function takes a data frame containing information about sample sizes by locus
and creates a heatmap plot to visualize the proportion of counts for each locus
relative to the overall sample size. It uses the \code{plotly} and \code{ggplot2} packages
for plotting.
}
\details{
The function first calculates the proportion of counts for each locus by dividing the
count for each locus by the overall sample size. It then creates a long-format data frame
using the \code{pivot_longer} function from the \code{tidyr} package. The long-format data frame
includes three columns: "silly" (categorical data), "locus" (locus names), and "proportion"
(proportion of counts for each locus). Next, the function performs a full join between the
original data and the \code{silly_n} data, where \code{silly_n} is a function that computes the
overall sample size. The function finally uses the \code{ggplot2} package to create a heatmap
plot with "silly" on the x-axis, "locus" on the y-axis, and the "proportion" mapped to
the fill color.
}
\examples{
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

sampn_df <- GCLr::sampn_by_locus(sillyvec, loci = GCLr::ex_LocusControl$locusnames[-c(10, 12, 13, 32, 33)], LocusCtl = GCLr::ex_LocusControl)

GCLr::plot_sampn_by_locus(sampn_df)
}
\seealso{
\code{\link[=sampn_by_locus]{sampn_by_locus()}}
}
