% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prior_diff_allele_freq.r
\name{post_prior_diff_allele_freq}
\alias{post_prior_diff_allele_freq}
\title{Get Difference in Allele Frequencies Between Baseline and BAYES Posteriors}
\usage{
post_prior_diff_allele_freq(sillyvec, loci, mydir, nchains, mixname)
}
\arguments{
\item{sillyvec}{sillyvec Character vector of population sillys used to create the baseline file.}

\item{loci}{Character vector of the loci used to produce the baseline and mixture files.}

\item{mydir}{the director where the BAYES mixture output files are located.  e.g., "BAYES/output" This directory should contain a folder named \code{mixname}, where the .FRQ files are saved.}

\item{nchains}{the number of MCMC chains analyzed for the mixture}

\item{mixname}{the name of the mixture. There should be a folder in \code{mydir} named mix}
}
\value{
a list containing 4 elements:
\itemize{
\item \code{Outliers}: a logical matrix with \code{nrows = length(sillyvec)} and \code{ncols = nchains} indicating where a population was an outlier for a given chain.
\item \code{MeanAbsDiffByPop}: a list with one element per chain containing a vector of mean (among loci) absolute allele frequency differences for each population in \code{sillyvec}
\item \code{PriorAlleleFreqs}: a 3 dimensional array of baseline allele frequencies for each locus in \code{loci}, silly in \code{sillyvec}, and allele.
\item \code{PosteriorAlleleFreqs}: a list with one element per chain containing a vector of BAYES MCMC allele frequencies for each population in \code{sillyvec} and each locus in \code{loci}
}

list(Outliers = Outliers, MeanAbsDiffByPop = myabsdiffsByPop, PriorAlleleFreqs = rawQ, PosteriorAlleleFreqs = myQ)
}
\description{
This function reads in BAYES MCMC samples of baseline frequencies for each population (one file per chain) and compares them to the original baseline allele frequencies (see details).
}
\details{
During a mixed stock analysis in BAYES, the original allele frequencies get updated at each MCMC iteration.
Looking at the difference between the average MCMC allele frequency and the original baseline frequencies can be useful when looking into reasons for non-convergence among chains (i.e., Gelman-Ruban shrink factor > 1.2)
Populations/reporting groups with large changes in allele frequencies for certain population/reporting groups can indicate that the mixture contains fish from a distinct population that is not represented in the baseline.
}
\examples{
\dontrun{

post_prior_diff_allele_freq(sillyvec = sillyvec, loci = loci, mydir = "BAYES/output", nchains = 5, mixname = "my.mixname")

}

}
\seealso{
See \link{BAYES manual}(system.file("BAYES", "MANUAL.DOC", package = "GCLr")) for additional details.
}
