% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genepop_dis.r
\name{read_genepop_dis}
\alias{read_genepop_dis}
\title{Read Genepop DIS Output}
\usage{
read_genepop_dis(file, loci = NULL, sillyvec = NULL)
}
\arguments{
\item{file}{The full file path to the \code{genepop} LD output, including the ".DIS" extension.}

\item{loci}{An optional character vector of locus names (default = \code{NULL}).
If \code{NULL} the locus names will come directly from the (.DIS) file and get truncated to 8 characters.
If supplying \code{loci}, make sure it is the same \code{loci} used in \code{\link[=gcl2genepop]{gcl2genepop()}}.}

\item{sillyvec}{An optional character vector of silly codes without the ".gcl" extension (default = \code{NULL}).
If \code{NULL}, \code{Pop} names will come directly from the (".DIS") file and likely include "_fishID" extensions.
If supplying \code{sillyvec}, make sure it is the same \code{sillyvec} used in \code{\link[=gcl2genepop]{gcl2genepop()}}.}
}
\value{
A non-tidy tibble with pairwise loci LD p-values per population and overall populations:
\itemize{
\item \code{Locus1}: locus name for 1st locus in pair
\item \code{Locus2}: locus name for 2nd locus in pair
\item \code{sillyvec_1}: pairwise loci LD p-values for 1st silly in \code{sillyvec}
\item \code{sillyvec_2}: pairwise loci LD p-values in \code{sillyvec}
\item \code{sillyvec_n}: pairwise loci LD p-values in \code{sillyvec}
\item \code{Overall}: pairwise loci LD p-values overall sillys in \code{sillyvec}
}
\description{
This function reads in output from a \code{genepop} Linkage Disequilibrium test ("*.DIS") file.
}
\details{
Designed and tested on Genepop v4.8.3. If the LD (".DIS) file contains tests for only 1 populations, the tibble will not contain an "Overall" column.
P-values with "No contingency table" are replaced with "1" and p-values = "0" are replaced with 1 / (n_Batches x n_Iterations per batch).
}
\examples{
\dontrun{
genepop_ld <- GCLr::read_genepop_dis(file = "~/R/test.txt.DIS", loci = loci, sillyvec = sillyvec)
}
 
}
\seealso{
\code{\link[genepop:genepop-package]{genepop::genepop-package()}}
\code{\link[genepop:Linkage]{genepop::test_LD()}}
\code{\link[=test_LD]{test_LD()}}
}
