% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genepop_pwfst.r
\name{read_genepop_pwfst}
\alias{read_genepop_pwfst}
\title{Read Pairwise Fst GENEPOP Output}
\usage{
read_genepop_pwfst(file, popnames, outfile = NULL)
}
\arguments{
\item{file}{the path to the GENEPOP .MIG file}

\item{popnames}{a character vector of population names corresponding to the populations in the GENEPOP file used to create the "*.MIG" file}

\item{outfile}{the path including file name to save a tab delimited text file of the pairwise matrix}
}
\value{
This function returns a named pairwise Fst matrix. If \code{!is.null(outfile)} the function produces a tab delimited text file of the pairwise matrix, otherwise no file is produced.
}
\description{
This function reads in the pairwise FST half matrix from a GENEPOP "*.MIG" file and produces a full matrix with dimnames. The matrix from this function can be used to create a pairwise FST tree using \code{\link[ape:nj]{ape::nj()}}
}
\examples{
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

loci <- GCLr::ex_LocusControl$locusnames[-c(10, 12, 13, 32, 33, 97, 98)]

dir.create(path = "~/GENEPOP")

GCLr::gcl2genepop(sillyvec = sillyvec, loci = loci, path = path.expand("~/GENEPOP/my_genepop.txt"), VialNums = TRUE, usat = FALSE,
                  ncores = parallel::detectCores(), npops = NULL, LocusCtl = GCLr::ex_LocusControl)

genepop::Fst(inputFile = path.expand("~/GENEPOP/my_genepop.txt"), pairs = TRUE, outputFile = path.expand("~/GENEPOP/my_genepop.txt"))

GCLr::read_genepop_pwfst(file = path.expand("~/GENEPOP/my_genepop.txt.MIG"), popnames = paste0("Pop", seq(14)))

}
\seealso{
\code{\link[genepop:Fst]{genepop::Fst()}}

\code{\link[=gcl2genepop]{gcl2genepop()}}:
}
