% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qc_geno.r
\name{read_qc_geno}
\alias{read_qc_geno}
\title{Read QC Data}
\usage{
read_qc_geno(
  qccsvFilepaths,
  skip = 15,
  LocusCtl = LocusControl,
  type = c("Biomark", "uSat", "GT-seq")[1]
)
}
\arguments{
\item{qccsvFilepaths}{character vector specifying the file paths of the QC CSV files.}

\item{skip}{number of lines to skip while reading the CSV files when type = "Biomark". (default = 15).}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}

\item{type}{the type of project ("Biomark", "uSat", or "GT-seq"), (default = "Biomark")}
}
\value{
Returns a few silly objects to the global environment
\itemize{
\item \code{qc.gcl} tibble objects
\item \code{qcSillys}; a character vector of qc sillys
}
}
\description{
This function reads QC data from Biomark CSV files. It's called on by \code{\link[=qc]{qc()}}.
}
