% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ind_miss_loci.r
\name{remove_ind_miss_loci}
\alias{remove_ind_miss_loci}
\title{Remove Individuals with Missing Loci Data}
\usage{
remove_ind_miss_loci(
  sillyvec,
  loci = LocusControl$locusnames,
  proportion = 0.8,
  LocusCtl = LocusControl
)
}
\arguments{
\item{sillyvec}{A character vector of silly codes without the ".gcl" extention (e.g. sillyvec <- c("KQUART06","KQUART08","KQUART10")).}

\item{loci}{A character vector specifying the names of loci (columns) in the datasets to be considered for missing data removal. Default is 'LocusControl$locusnames'.}

\item{proportion}{A numeric value (0 to 1) specifying the maximum proportion of missing data allowed for each individual in the specified loci. Default is 0.8.}

\item{LocusCtl}{An object of class 'LocusControl', containing locus names created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}. Default is 'LocusControl'.}
}
\value{
A data frame containing the individuals removed from each dataset in 'sillyvec', along with their corresponding dataset names.
}
\description{
This function removes individuals from a data set based on the proportion of missing data in specified loci.
}
\details{
This function takes a vector of dataset names ('sillyvec') and removes individuals from each dataset based on the proportion of missing data in the specified loci. The 'loci' argument allows the user to specify the specific loci (columns) to be considered for missing data removal. By default, all loci in 'LocusControl' will be used.

The function calculates the proportion of missing data for each individual in the specified loci. If the proportion of missing data for an individual exceeds the 'proportion' argument, the individual is removed from the dataset. Only the individuals with a proportion of missing data less than or equal to 'proportion' will remain in the dataset.

Note that the function uses the 'get' function to access each dataset and relies on the 'dplyr' and 'purrr' packages for data manipulation and filtering.

If no individuals are removed from any of the datasets, the function will print a message indicating that no individuals were removed. Otherwise, it will print the total number of individuals removed from each dataset in 'sillyvec'.
}
\examples{
\dontrun{
create_locuscontrol(markersuite = "Sockeye2011_96SNPs", username = "awbarclay", password = password)
sillyvec <- c("SMCDO03", "SNEVA13")
loki2r(sillyvec = sillyvec, username = "awbarclay", password = password)

naloci <- dimnames(SMCDO03.gcl)[[2]][100:199]
empty.gcl <- SMCDO03.gcl
empty.gcl[,naloci] <- NA

remove_ind_miss_loci(sillyvec = c(sillyvec, "empty"), loci = LocusControl$locusnames, proportion = 0.8)
}

}
