% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampn_by_locus.r
\name{sampn_by_locus}
\alias{sampn_by_locus}
\title{Sample Size by Locus}
\usage{
sampn_by_locus(
  sillyvec,
  loci = LocusControl$locusnames,
  LocusCtl = LocusControl
)
}
\arguments{
\item{sillyvec}{A vector of sample names to extract data from.}

\item{loci}{A character vector containing the names of loci for which data will be extracted. Defaults to the locus names in the \code{LocusControl} object.}

\item{LocusCtl}{The \code{LocusControl} object containing the data for all loci. Default is the global \code{LocusControl} object.}
}
\value{
A data frame containing the counts of non-missing values for each locus in each sample. The output will have one row per sample and one column per locus, along with an additional 'silly' column indicating the sample name.
}
\description{
This function takes a vector of sample names (\code{sillyvec}) and extracts data for specified loci from a \code{LocusControl} object. It returns a data frame with the counts of non-missing values for each locus in each sample.
}
\details{
The function first checks if all specified loci in the \code{loci} argument exist in the \code{LocusControl} object. If any locus is not found, an error message is thrown.

The function then iterates through each sample in \code{sillyvec}, retrieves the data for the specified loci from the \code{LocusControl} object, and calculates the count of non-missing values for each locus. The results are combined into a single data frame.

Note the output tibble from this function can be fed to \code{\link[=plot_sampn_by_locus]{plot_sampn_by_locus()}} to produce an interactive heatmap of the proportion of fish with scores for each locus and silly
}
\examples{
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

GCLr::sampn_by_locus(sillyvec = sillyvec, loci = GCLr::ex_LocusControl$locusnames[-c(10, 12, 13, 32, 33)], LocusCtl = GCLr::ex_LocusControl)

}
\seealso{
\code{\link[=plot_sampn_by_locus]{plot_sampn_by_locus()}}
}
