% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.r
\name{save_objects}
\alias{save_objects}
\title{Save R objects to disk}
\usage{
save_objects(objects, path, rds = FALSE)
}
\arguments{
\item{objects}{character vector of objects you wish to save}

\item{path}{character vector of where you want to save objects}

\item{rds}{logical; if set to TRUE the objects will be saved in rds format,
default is FALSE for backwards compatibility.}
}
\value{
All objects are saved as "objects.txt" or "objects.rds"
}
\description{
This function saves R objects with either \code{dput} or \code{saveRDS}.
}
\details{
\code{dput} saves object code to a text file.
\code{saveRDS} saves objects in rds format, which is a compressed format - saves and reads back into R much faster than a text file.
}
\examples{
iris <- iris  # bringing into pos = 1
save_objects(objects = "iris", path = "Objects", rds = TRUE)

}
