% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_sillys.r
\name{save_sillys}
\alias{save_sillys}
\title{Save sillys as R objects}
\usage{
save_sillys(sillyvec, path, rds = FALSE)
}
\arguments{
\item{sillyvec}{character vector of sillys you wish to save}

\item{path}{character vector of where you want to save sillys}

\item{rds}{logical; if set to TRUE the sillys will be saved in rds format,
default is FALSE for backwards compatibility.}
}
\value{
All sillyvec are saved as "sillyvec.txt" or "sillyvec.rds", the .gcl is removed
}
\description{
This function saves sillys as R objects with \code{dput}, it is a wrapper for \code{dput}.
}
\examples{
iris <- iris  # bringing into pos = 1
save_sillys(sillyvec = "SGILL15D15", path = "Raw genotypes")

}
