% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silly_n.r
\name{silly_n}
\alias{silly_n}
\title{Get Sample Sizes from *.gcl Objects}
\usage{
silly_n(sillyvec)
}
\arguments{
\item{sillyvec}{A character vector of silly codes with associated *.gcl objects for which you want to obtain sample sizes.}
}
\value{
A two-column tibble containing the silly codes (silly) and their corresponding sample sizes (n).
}
\description{
This function retrieves the sample sizes of *.gcl objects associated with given silly codes.
}
\examples{
 
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

GCLr::silly_n(sillyvec = sillyvec)

}
