% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_LD.r
\name{summarize_LD}
\alias{summarize_LD}
\alias{Summarize_LD.GCL}
\title{Summarize LD Results}
\usage{
summarize_LD(LDresults, alpha = 0.05, prop_sign_pops = 0.5)

Summarize_LD.GCL(LDresults, alpha = 0.05, prop_sign_pops = 0.5)
}
\arguments{
\item{LDresults}{The output from \code{\link[=test_LD]{test_LD()}}; a tibble with 3 variables:
\itemize{
\item \code{Locus_pair}
\item \code{Pop}
\item \code{Pvalue}
}}

\item{alpha}{A numeric value indicating the significance level for filtering the \code{\link[=test_LD]{test_LD()}} output; default is 0.05.}

\item{prop_sign_pops}{A numeric value indicating the proportion of populations with significant test results; default is 0.5.
Only locus pairs that have > prop_sign_pops will be plotted.}
}
\value{
An interactive bar plot with the proportion of populations with significant tests on the y-axis and the locus pairs on the x-axis.
}
\description{
This function takes the output from \code{\link[=test_LD]{test_LD()}}, filters tests by the supplied significance level (alpha),
and makes a bar plot of the proportion of pops with significant test results for each locus pair.
}
\examples{
attach("V:/Analysis/5_Coastwide/Chum/Baseline for South Peninsula MSA/South_AK_Peninsula_chum_baseline analysis.Rdata")
LDresults <- LD
detach()
summarize_LD(LDresults, alpha = 0.05, prop_sign_pops = 0.5)

}
