% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_LD.r
\name{test_LD}
\alias{test_LD}
\title{Perform LD Testing in Parallel}
\usage{
test_LD(
  path,
  genepopFiles,
  dememorizations = 10000,
  batches = 100,
  iterations = 5000,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{path}{The folder where the \code{genepop} files are located and where the results will be written}

\item{genepopFiles}{The names of the \code{genepop} files located in the specified path}

\item{dememorizations}{Integer value indicating the length of the dememorization step of the Markov chain algorithm}

\item{batches}{Integer value indicating the number of batches}

\item{iterations}{Integer value indicating the iterations per batch}

\item{ncores}{The number of cores for multicoring using \code{doParallel} and \code{foreach}. Default is the number of available cores minus 1.}
}
\value{
A summarized tibble with the following columns:
\itemize{
\item \code{Locus_pair} - Combined locus pairs
\item \code{Pop} - Population name
\item \code{Pvalue} - P-value for LD test
}
}
\description{
This function performs LD testing using \code{\link[genepop:Linkage]{genepop::test_LD()}} in parallel. It takes a list of \code{genepop} files with n populations per file and tests for LD. The results are combined into a single summary object containing p-values for each locus pair and population.
}
\examples{
\dontrun{
GCLr::test_LD(genepopFiles = my.files, path = "GENEPOP", batches = 1, iterations = 1, ncores = 18)
}

}
