% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_comp.r
\name{var_comp}
\alias{var_comp}
\title{Variance Components Calculation}
\usage{
var_comp(
  sillyvec,
  loci,
  groupvec,
  fstatfile = NULL,
  LocusCtl = LocusControl,
  ncores = parallel::detectCores()
)
}
\arguments{
\item{sillyvec}{A character vector specifying the population silly codes for analysis.}

\item{loci}{A character vector of locus names to include in the analysis. Can include haploid markers and/or combined loci.}

\item{groupvec}{A numeric vector specifying the reporting group affiliation for each silly code.}

\item{fstatfile}{Optional character string specifying the path to an existing FSTAT file (".dat"). If provided, the function verifies compatibility with the input data. (default = NULL)}

\item{LocusCtl}{an object created by \code{\link[=create_locuscontrol]{create_locuscontrol()}}, (default = LocusControl)}
}
\value{
A tibble containing the four variance components for each locus:
\itemize{
\item P: For populations (reporting groups in this case).
\item S: For subpopulations within populations (populations within reporting groups).
\item I: For individuals within subpopulations (individuals within populations).
\item G: For alleles within individuals (akin to heterozygosity).
}
}
\description{
This function calculates the variance components of a three-level hierarchy. This function is intended to help with marker selection (i.e., going from a 96 to 24 SNP baseline).
}
\details{
Component definitions in the output assume that you are running this function on silly codes that are populations (see Value section).
This function could also be run on silly codes that are collections, but the definitions in the output would be different.
}
\examples{
sillyvec <- GCLr::base2gcl(GCLr::ex_baseline)

groupvec <- GCLr::ex_baseline \%>\% 
  dplyr::group_by(collection) \%>\% 
  dplyr::summarize(group = unique(repunit)) \%>\% 
  dplyr::mutate(group = factor(group) \%>\% as.numeric()) \%>\% 
  dplyr::pull(group)

loci <- GCLr::ex_baseline[,-c(1:5)] \%>\%
  names() \%>\%
  gsub(pattern = "*\\\\.1", x = ., replacement = "") \%>\%
  unique()

GCLr::var_comp(sillyvec = sillyvec, loci = loci, groupvec = groupvec, LocusCtl = GCLr::ex_LocusControl)

}
\references{
Weir, B.S. 1995. Genetic Data Analysis. Sinauer Associates, Inc. Sunderland, MA. (pg 184)
}
